Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    6.656487 18.842074
#> 2    5.295532 12.809827
#> 3    8.739060 12.127714
#> 4    7.578688 16.644230
#> 5    6.688274 17.594804
#> 6    7.853479 11.063922
#> 7    8.041579 16.740757
#> 8   10.081280 17.526556
#> 9   10.171586 16.590265
#> 10   8.191422 11.396123
#> 11  11.941483 15.816271
#> 12   7.709778 18.122810
#> 13  13.629291 14.138172
#> 14  13.273229 11.800854
#> 15   9.707912 12.797155
#> 16  10.795023 14.401853
#> 17  11.244621 14.787856
#> 18   8.968437  9.795419
#> 19   6.337682 13.411676
#> 20  10.577892 12.723606
#> 21   8.826818  7.522809
#> 22  13.792230 19.531146
#> 23  11.518197 11.720893
#> 24   9.885944 13.834200
#> 25   7.807499 11.098447
#> 26  13.118445 18.309590
#> 27  10.981589 13.541205
#> 28  11.636939 15.048786
#> 29   5.110082 19.760691
#> 30   9.291797 19.297068
#> 31  10.452733 12.705443
#> 32   8.168314 16.639062
#> 33  10.515498 15.119797
#> 34   8.278400 15.407388
#> 35   8.247734  6.120185
#> 36  11.257725  9.978341
#> 37   9.814965 17.050192
#> 38  13.061034 12.653152
#> 39  10.989178 11.233271
#> 40   7.997267 18.336961
#> 41   6.253706 11.049978
#> 42  10.946327 12.894988
#> 43   5.654115 16.655762
#> 44  11.819798 14.492713
#> 45   6.326762 15.817566
#> 46   9.540094 15.432668
#> 47  11.203654 14.525177
#> 48   7.515420 11.569262
#> 49  11.181835 15.347398
#> 50   7.635086 14.762731
#> 51  14.919930 15.483081
#> 52   9.241905 13.023057
#> 53   8.936573 19.646576
#> 54  11.386384 12.274473
#> 55  11.765991  8.225806
#> 56  11.864044 14.645639
#> 57  13.453082 11.758749
#> 58  12.109667 19.701228
#> 59   9.654986 16.087007
#> 60  11.285499 17.344458
#> 61   6.751875 12.914062
#> 62   7.952976 13.709413
#> 63   7.797479 16.149464
#> 64  10.344445 18.752244
#> 65   6.329312 17.170481
#> 66  13.966690 11.379929
#> 67  11.818444 16.508659
#> 68  11.591760 13.655863
#> 69  11.916345 19.785227
#> 70   8.763261 18.828413
#> 71  12.086002 12.829185
#> 72   8.904530 20.288810
#> 73  11.452690 13.580326
#> 74  10.548050 13.543482
#> 75  13.323994 16.715632
#> 76   8.534218  6.095884
#> 77  10.713696 16.026853
#> 78  10.667779 17.168739
#> 79   9.935832 13.901428
#> 80  11.555466 10.881143
#> 81  10.723980 17.275569
#> 82  11.096786 17.618720
#> 83   6.833553 10.623607
#> 84   8.790925 18.957591
#> 85  10.465551 15.075098
#> 86  10.855970 17.559831
#> 87   8.193881 14.596332
#> 88  10.752266 13.428337
#> 89  12.749745 15.555881
#> 90   8.360548 18.855555
#> 91   8.235750 15.234929
#> 92   8.871209 17.332710
#> 93   8.777340 12.011937
#> 94   8.380560 14.290104
#> 95  11.114206 11.654124
#> 96  12.969385 13.973824
#> 97  12.739706 17.406059
#> 98   8.817966 13.428907
#> 99  10.253520 14.201333
#> 100 11.117617  5.849145
#> 
#> 
#> $misty.uniqueid
#> [1] "f1c8daf67d51f91da3c77dace81373eb"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   20.3038482 17.400335
#> 2   11.0557667 18.757272
#> 3   11.2937597 13.607678
#> 4    5.7658646 16.823636
#> 5   14.6121446 10.316543
#> 6   11.8703514 20.542947
#> 7   15.3912349 15.819401
#> 8    8.3717066 10.073532
#> 9    8.3824641 10.213243
#> 10  17.1228996 15.211919
#> 11  14.7088932 17.470986
#> 12  12.1608659 11.142669
#> 13  10.6644117 18.019670
#> 14  16.4913886 18.990248
#> 15   9.4906750 11.232403
#> 16  11.2671632 11.142925
#> 17  11.5793774 19.743879
#> 18  10.1621701  6.786021
#> 19  10.6417220 18.697596
#> 20  10.2966667 14.198851
#> 21   9.9101831 13.303549
#> 22   6.0355173  9.862257
#> 23   8.3983251 16.366564
#> 24   8.3768295 22.577196
#> 25  17.1373174 14.923767
#> 26   5.4771712 17.691212
#> 27  22.4439739 20.695397
#> 28   9.4031972 13.304878
#> 29  12.0106314 22.766467
#> 30   9.4285231  9.632249
#> 31  12.6641136  9.948471
#> 32   3.0495263 23.217567
#> 33   7.2166538 22.070903
#> 34   6.3626902 10.935312
#> 35   9.4753942 11.329258
#> 36  12.8682505  9.212486
#> 37   5.2300807  7.160484
#> 38  16.5438709 17.373494
#> 39  -1.3136901 19.691622
#> 40  11.7314485 15.861174
#> 41  14.9353376 12.554014
#> 42  13.8425180 19.504627
#> 43  -3.4054230 18.776280
#> 44   4.6761868 16.613217
#> 45   1.0571907 20.036823
#> 46   8.2736480  9.227748
#> 47  14.7132218 12.615997
#> 48   9.3538990  6.885467
#> 49  11.9962994 28.737600
#> 50   2.7337337 11.710645
#> 51   6.8209271 10.533756
#> 52  14.8252598 19.229130
#> 53  16.9224224 15.924024
#> 54  13.8279423 20.342391
#> 55   6.5612180 17.085579
#> 56  10.9022791 20.277896
#> 57   0.5486068 13.749521
#> 58  12.3509200  7.474955
#> 59   8.7860348 10.809673
#> 60  10.7357745  4.819315
#> 61   4.9917159 12.359154
#> 62  14.5506888 11.782330
#> 63  14.3927948 13.156204
#> 64  13.7007677 15.601976
#> 65   9.9631466 12.719095
#> 66   5.8744477 17.029973
#> 67  14.2363441 23.805197
#> 68  16.1570706 17.592306
#> 69   7.6440009 14.210536
#> 70  15.3986262  6.875991
#> 71  13.8600348 16.588217
#> 72   7.6037685 16.255138
#> 73  12.3631243 12.643720
#> 74  -0.3325145  8.916863
#> 75   8.6220429 21.574738
#> 76  16.1260136  8.711393
#> 77  13.6313472 10.793923
#> 78   8.4145598 29.015532
#> 79  13.8024956  8.023227
#> 80   2.8830581 17.822368
#> 81  13.3063715 16.791667
#> 82  19.0667349 15.367703
#> 83  13.5702059 16.594828
#> 84   5.3279586 12.778793
#> 85  15.1712356  6.941634
#> 86   4.5048300 13.393867
#> 87  13.0727559 17.274661
#> 88   9.2991540 16.054438
#> 89  15.1140079 16.637717
#> 90  10.4786855 14.114797
#> 91  12.2102852 16.452716
#> 92   7.7542832 13.599448
#> 93  16.2079840 20.457791
#> 94  10.9191109  9.425564
#> 95   4.1073157 15.331025
#> 96  16.8983019 14.273337
#> 97   6.5960377 24.019765
#> 98   9.1022678 12.675206
#> 99  15.0478473 16.035806
#> 100 14.9209633 12.051978
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    6.656487 18.842074
#> 2    5.295532 12.809827
#> 3    8.739060 12.127714
#> 4    7.578688 16.644230
#> 5    6.688274 17.594804
#> 6    7.853479 11.063922
#> 7    8.041579 16.740757
#> 8   10.081280 17.526556
#> 9   10.171586 16.590265
#> 10   8.191422 11.396123
#> 11  11.941483 15.816271
#> 12   7.709778 18.122810
#> 13  13.629291 14.138172
#> 14  13.273229 11.800854
#> 15   9.707912 12.797155
#> 16  10.795023 14.401853
#> 17  11.244621 14.787856
#> 18   8.968437  9.795419
#> 19   6.337682 13.411676
#> 20  10.577892 12.723606
#> 21   8.826818  7.522809
#> 22  13.792230 19.531146
#> 23  11.518197 11.720893
#> 24   9.885944 13.834200
#> 25   7.807499 11.098447
#> 26  13.118445 18.309590
#> 27  10.981589 13.541205
#> 28  11.636939 15.048786
#> 29   5.110082 19.760691
#> 30   9.291797 19.297068
#> 31  10.452733 12.705443
#> 32   8.168314 16.639062
#> 33  10.515498 15.119797
#> 34   8.278400 15.407388
#> 35   8.247734  6.120185
#> 36  11.257725  9.978341
#> 37   9.814965 17.050192
#> 38  13.061034 12.653152
#> 39  10.989178 11.233271
#> 40   7.997267 18.336961
#> 41   6.253706 11.049978
#> 42  10.946327 12.894988
#> 43   5.654115 16.655762
#> 44  11.819798 14.492713
#> 45   6.326762 15.817566
#> 46   9.540094 15.432668
#> 47  11.203654 14.525177
#> 48   7.515420 11.569262
#> 49  11.181835 15.347398
#> 50   7.635086 14.762731
#> 51  14.919930 15.483081
#> 52   9.241905 13.023057
#> 53   8.936573 19.646576
#> 54  11.386384 12.274473
#> 55  11.765991  8.225806
#> 56  11.864044 14.645639
#> 57  13.453082 11.758749
#> 58  12.109667 19.701228
#> 59   9.654986 16.087007
#> 60  11.285499 17.344458
#> 61   6.751875 12.914062
#> 62   7.952976 13.709413
#> 63   7.797479 16.149464
#> 64  10.344445 18.752244
#> 65   6.329312 17.170481
#> 66  13.966690 11.379929
#> 67  11.818444 16.508659
#> 68  11.591760 13.655863
#> 69  11.916345 19.785227
#> 70   8.763261 18.828413
#> 71  12.086002 12.829185
#> 72   8.904530 20.288810
#> 73  11.452690 13.580326
#> 74  10.548050 13.543482
#> 75  13.323994 16.715632
#> 76   8.534218  6.095884
#> 77  10.713696 16.026853
#> 78  10.667779 17.168739
#> 79   9.935832 13.901428
#> 80  11.555466 10.881143
#> 81  10.723980 17.275569
#> 82  11.096786 17.618720
#> 83   6.833553 10.623607
#> 84   8.790925 18.957591
#> 85  10.465551 15.075098
#> 86  10.855970 17.559831
#> 87   8.193881 14.596332
#> 88  10.752266 13.428337
#> 89  12.749745 15.555881
#> 90   8.360548 18.855555
#> 91   8.235750 15.234929
#> 92   8.871209 17.332710
#> 93   8.777340 12.011937
#> 94   8.380560 14.290104
#> 95  11.114206 11.654124
#> 96  12.969385 13.973824
#> 97  12.739706 17.406059
#> 98   8.817966 13.428907
#> 99  10.253520 14.201333
#> 100 11.117617  5.849145
#> 
#> 
#> $misty.uniqueid
#> [1] "f1c8daf67d51f91da3c77dace81373eb"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   20.3038482 17.400335
#> 2   11.0557667 18.757272
#> 3   11.2937597 13.607678
#> 4    5.7658646 16.823636
#> 5   14.6121446 10.316543
#> 6   11.8703514 20.542947
#> 7   15.3912349 15.819401
#> 8    8.3717066 10.073532
#> 9    8.3824641 10.213243
#> 10  17.1228996 15.211919
#> 11  14.7088932 17.470986
#> 12  12.1608659 11.142669
#> 13  10.6644117 18.019670
#> 14  16.4913886 18.990248
#> 15   9.4906750 11.232403
#> 16  11.2671632 11.142925
#> 17  11.5793774 19.743879
#> 18  10.1621701  6.786021
#> 19  10.6417220 18.697596
#> 20  10.2966667 14.198851
#> 21   9.9101831 13.303549
#> 22   6.0355173  9.862257
#> 23   8.3983251 16.366564
#> 24   8.3768295 22.577196
#> 25  17.1373174 14.923767
#> 26   5.4771712 17.691212
#> 27  22.4439739 20.695397
#> 28   9.4031972 13.304878
#> 29  12.0106314 22.766467
#> 30   9.4285231  9.632249
#> 31  12.6641136  9.948471
#> 32   3.0495263 23.217567
#> 33   7.2166538 22.070903
#> 34   6.3626902 10.935312
#> 35   9.4753942 11.329258
#> 36  12.8682505  9.212486
#> 37   5.2300807  7.160484
#> 38  16.5438709 17.373494
#> 39  -1.3136901 19.691622
#> 40  11.7314485 15.861174
#> 41  14.9353376 12.554014
#> 42  13.8425180 19.504627
#> 43  -3.4054230 18.776280
#> 44   4.6761868 16.613217
#> 45   1.0571907 20.036823
#> 46   8.2736480  9.227748
#> 47  14.7132218 12.615997
#> 48   9.3538990  6.885467
#> 49  11.9962994 28.737600
#> 50   2.7337337 11.710645
#> 51   6.8209271 10.533756
#> 52  14.8252598 19.229130
#> 53  16.9224224 15.924024
#> 54  13.8279423 20.342391
#> 55   6.5612180 17.085579
#> 56  10.9022791 20.277896
#> 57   0.5486068 13.749521
#> 58  12.3509200  7.474955
#> 59   8.7860348 10.809673
#> 60  10.7357745  4.819315
#> 61   4.9917159 12.359154
#> 62  14.5506888 11.782330
#> 63  14.3927948 13.156204
#> 64  13.7007677 15.601976
#> 65   9.9631466 12.719095
#> 66   5.8744477 17.029973
#> 67  14.2363441 23.805197
#> 68  16.1570706 17.592306
#> 69   7.6440009 14.210536
#> 70  15.3986262  6.875991
#> 71  13.8600348 16.588217
#> 72   7.6037685 16.255138
#> 73  12.3631243 12.643720
#> 74  -0.3325145  8.916863
#> 75   8.6220429 21.574738
#> 76  16.1260136  8.711393
#> 77  13.6313472 10.793923
#> 78   8.4145598 29.015532
#> 79  13.8024956  8.023227
#> 80   2.8830581 17.822368
#> 81  13.3063715 16.791667
#> 82  19.0667349 15.367703
#> 83  13.5702059 16.594828
#> 84   5.3279586 12.778793
#> 85  15.1712356  6.941634
#> 86   4.5048300 13.393867
#> 87  13.0727559 17.274661
#> 88   9.2991540 16.054438
#> 89  15.1140079 16.637717
#> 90  10.4786855 14.114797
#> 91  12.2102852 16.452716
#> 92   7.7542832 13.599448
#> 93  16.2079840 20.457791
#> 94  10.9191109  9.425564
#> 95   4.1073157 15.331025
#> 96  16.8983019 14.273337
#> 97   6.5960377 24.019765
#> 98   9.1022678 12.675206
#> 99  15.0478473 16.035806
#> 100 14.9209633 12.051978
#> 
#>