Train MISTy models

Trains multi-view models for all target markers, estimates the performance, the contributions of the view specific models and the importance of predictor markers for each target marker.

run_misty(
  views,
  results.folder = "results",
  seed = 42,
  target.subset = NULL,
  bypass.intra = FALSE,
  cv.folds = 10,
  cached = FALSE,
  append = FALSE,
  ...
)

Arguments

views

view composition.

results.folder

path to the top level folder to store raw results.

seed

seed used for random sampling to ensure reproducibility.

target.subset

subset of targets to train models for. If NULL, models will be trained for markers in the intraview.

bypass.intra

a logical indicating whether to train a baseline model using the intraview data (see Details).

cv.folds

number of cross-validation folds to consider for estimating the performance of the multi-view models.

cached

a logical indicating whether to cache the trained models and to reuse previously cached ones if they already exist for this sample.

append

a logical indicating whether to append the performance and coefficient files in the results.folder. Consider setting to TRUE when rerunning a workflow with different target.subset parameters.

...

all additional parameters are passed to ranger() for training the view-specific models (see Details for defaults).

Value

Path to the results folder that can be passed to collect_results().

Details

If bypass.intra is set to TRUE all variable in the intraview the intraview data will be treated as targets only. The baseline intraview model in this case is a trivial model that predicts the average of each target. If the intraview has only one variable this switch is automatically set to TRUE.

Default values passed to ranger() for training the view-specific models: num.trees = 100, importance = "impurity", num.threads = 1, seed = seed.

See also

Examples

# Create a view composition of an intraview and a paraview with radius 10 then
# run MISTy for a single sample.

library(dplyr)

# get the expression data
data("synthetic")
expr <- synthetic[[1]] %>% select(-c(row, col, type))
# get the coordinates for each cell
pos <- synthetic[[1]] %>% select(row, col)

# compose
misty.views <- create_initial_view(expr) %>% add_paraview(pos, l = 10)
#> 
#> Generating paraview

# run with default parameters
run_misty(misty.views)
#> 
#> Training models
#> [1] "/tmp/RtmpMWfNBo/file62e32ca0b3c4/reference/results"