Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    9.142892 16.393385
#> 2    9.465783 13.943560
#> 3    9.810335 11.262750
#> 4   10.360627 15.993334
#> 5   15.279331 13.631338
#> 6   11.600551  9.367000
#> 7   10.272986 17.540637
#> 8    8.690546 11.091164
#> 9   11.238097 20.410711
#> 10   9.008343 12.139154
#> 11   6.644129 11.848259
#> 12   7.580800 12.496396
#> 13  16.189966 14.895417
#> 14  10.107069 11.210695
#> 15  13.902496 14.636737
#> 16  10.935195 18.754577
#> 17  10.260030 19.554459
#> 18  10.513464 13.669984
#> 19  11.662317 15.091780
#> 20  11.329546 17.212485
#> 21   9.351006 13.443272
#> 22   9.766898  8.506501
#> 23   7.888453 14.165038
#> 24   5.063959 12.360342
#> 25  10.139119 11.177554
#> 26   9.037650 12.443616
#> 27  12.556824 13.392453
#> 28  12.349295 20.914221
#> 29   6.696575 17.367068
#> 30   9.336167 14.774742
#> 31  10.624776 11.479659
#> 32  13.546892 14.246438
#> 33  12.571835 17.635840
#> 34  12.779159 10.812975
#> 35   9.186021 17.697883
#> 36   9.615885 13.002516
#> 37   9.481450 19.321155
#> 38   8.496463  5.082661
#> 39  11.919867 17.572389
#> 40   8.721970 17.944395
#> 41   9.792220 14.215198
#> 42  10.724107 16.162698
#> 43   9.723955 11.853637
#> 44  11.867172 14.868685
#> 45   9.462713 12.005517
#> 46   8.417439 16.170925
#> 47  11.694947 11.913272
#> 48   9.783619 15.976626
#> 49   7.810727 13.984781
#> 50   5.589841 10.348084
#> 51   9.338426 15.018089
#> 52   9.171589 14.116097
#> 53  13.216286 17.549513
#> 54   8.325149 12.812848
#> 55   7.593651 18.000674
#> 56   8.777604 13.975025
#> 57   8.890623 16.736588
#> 58  11.068520 10.322416
#> 59  11.028493 17.230290
#> 60   7.771067 16.935474
#> 61  12.241019 14.111416
#> 62  12.232828 12.673303
#> 63  11.859715 10.878273
#> 64  10.863607 12.646655
#> 65   9.293865  6.252061
#> 66  12.309485 17.562880
#> 67   9.425402 14.872808
#> 68   7.294368 17.676825
#> 69   9.027313 13.704298
#> 70   9.244010 12.140135
#> 71  10.315597 17.588271
#> 72   7.482679 13.579633
#> 73   7.887133 12.362669
#> 74  10.867698 17.131006
#> 75  10.793098 16.946039
#> 76   8.344616 12.731451
#> 77  12.017646 20.739101
#> 78  11.445803 18.578711
#> 79   9.801106 11.336545
#> 80  15.081621 16.430858
#> 81  12.998034 15.075573
#> 82   7.854737 11.847093
#> 83  10.824745 14.994232
#> 84  13.901937 13.234428
#> 85  10.905157 17.379135
#> 86   7.905843 14.756253
#> 87   9.178311 18.157743
#> 88   8.720429 12.471440
#> 89   8.388292 15.985588
#> 90  10.156951 13.841807
#> 91   6.165200 13.457487
#> 92   8.858229 13.287433
#> 93  11.074452 14.227778
#> 94   7.105353 18.791416
#> 95  11.510528 15.356714
#> 96  12.590685 17.701177
#> 97  10.438819 20.701206
#> 98  10.164032  9.922782
#> 99   7.620928 23.430470
#> 100  9.934664 17.145230
#> 
#> 
#> $misty.uniqueid
#> [1] "7ea5df9d1da9f18b77fa9818ea2605dd"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   14.1011873 19.742848
#> 2   12.2156412 18.819131
#> 3   11.3501363 15.790395
#> 4    5.1320621 16.260031
#> 5   19.7665612 11.656634
#> 6   11.0273285 21.331675
#> 7    9.4343160 18.258239
#> 8   11.9887273 14.467727
#> 9   15.2295013 12.611135
#> 10   9.0849375 20.102323
#> 11  13.2966287  9.119595
#> 12  10.5631967 13.884667
#> 13  12.6352429  9.164188
#> 14  16.5655873 18.238783
#> 15   8.4316781 20.393283
#> 16  11.5146501 17.518505
#> 17  -0.6335323 12.818916
#> 18  16.3791388 19.421143
#> 19   3.2253939 17.153230
#> 20  10.7758665 14.131669
#> 21  18.2456118 13.203442
#> 22  16.2246284 22.298338
#> 23  11.1702150 11.479576
#> 24   9.8578328 11.526532
#> 25   5.1709759 11.996685
#> 26  13.5603120  5.654598
#> 27  13.8623090 17.205546
#> 28  10.1823029 16.869211
#> 29   2.2216365  9.875181
#> 30   7.9946435  5.398114
#> 31  12.3859983 18.237705
#> 32   2.7121858 15.201699
#> 33  13.6821971 16.577671
#> 34  16.9540054  4.719504
#> 35   7.6074471 13.123602
#> 36   7.2302758 20.186299
#> 37   9.6570519 16.509709
#> 38   7.7847473 11.521746
#> 39   9.9265767 12.477288
#> 40   7.0766853 15.815182
#> 41  23.4009983 12.539164
#> 42   3.8531004 19.259427
#> 43  16.7496080 12.160188
#> 44   8.8677373 16.773691
#> 45   0.2157770 14.432193
#> 46  12.3072504 12.858021
#> 47  10.4122790 16.403665
#> 48  25.1553908 16.885414
#> 49   8.6051366 11.333473
#> 50   4.0915530 14.361771
#> 51   7.3150268 17.678935
#> 52  13.9941715 22.642420
#> 53   4.5416706 23.309709
#> 54  15.5134975 18.509465
#> 55  17.6645689 10.580293
#> 56   4.1520805 13.578944
#> 57  19.3104158 16.352046
#> 58  14.4342451 24.362638
#> 59  10.5187017 12.863263
#> 60  15.2318051 13.904168
#> 61   9.1130044  8.487530
#> 62  17.5372851  6.293770
#> 63   2.0480684 12.051733
#> 64  14.4911917 14.694430
#> 65  10.8248934 15.944544
#> 66  10.5487007 17.147310
#> 67  13.2788592 13.446042
#> 68  18.1013192 16.599150
#> 69   9.1167352 17.081623
#> 70  13.7287578 15.824537
#> 71   9.5376935  9.970923
#> 72   2.2555123 11.019431
#> 73   9.7649665 22.906063
#> 74  -1.3984734 21.682775
#> 75   8.4681389 14.114004
#> 76  16.5904423 13.844990
#> 77  10.9137513  4.791087
#> 78   7.8918079  9.561384
#> 79  13.6614243 10.343523
#> 80  17.4646592 10.191988
#> 81   5.6842623 19.017859
#> 82  10.0136834 24.063663
#> 83  16.3162814 20.653854
#> 84   8.5715147 14.407709
#> 85  12.5966896 15.885167
#> 86  11.8432804  9.393668
#> 87  17.5048808 13.413894
#> 88   9.8047243 12.528775
#> 89   2.8248108 14.917911
#> 90   8.8408710 19.679833
#> 91   1.9856638 25.684271
#> 92  13.9173288 12.809161
#> 93   1.5134808 23.695666
#> 94  11.5957740 20.030302
#> 95   8.7828765 12.402506
#> 96  11.4253554 13.028264
#> 97  13.8427296 15.663133
#> 98   4.4987190 10.819931
#> 99  12.4419760 18.862501
#> 100 15.0748376 10.528965
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    9.142892 16.393385
#> 2    9.465783 13.943560
#> 3    9.810335 11.262750
#> 4   10.360627 15.993334
#> 5   15.279331 13.631338
#> 6   11.600551  9.367000
#> 7   10.272986 17.540637
#> 8    8.690546 11.091164
#> 9   11.238097 20.410711
#> 10   9.008343 12.139154
#> 11   6.644129 11.848259
#> 12   7.580800 12.496396
#> 13  16.189966 14.895417
#> 14  10.107069 11.210695
#> 15  13.902496 14.636737
#> 16  10.935195 18.754577
#> 17  10.260030 19.554459
#> 18  10.513464 13.669984
#> 19  11.662317 15.091780
#> 20  11.329546 17.212485
#> 21   9.351006 13.443272
#> 22   9.766898  8.506501
#> 23   7.888453 14.165038
#> 24   5.063959 12.360342
#> 25  10.139119 11.177554
#> 26   9.037650 12.443616
#> 27  12.556824 13.392453
#> 28  12.349295 20.914221
#> 29   6.696575 17.367068
#> 30   9.336167 14.774742
#> 31  10.624776 11.479659
#> 32  13.546892 14.246438
#> 33  12.571835 17.635840
#> 34  12.779159 10.812975
#> 35   9.186021 17.697883
#> 36   9.615885 13.002516
#> 37   9.481450 19.321155
#> 38   8.496463  5.082661
#> 39  11.919867 17.572389
#> 40   8.721970 17.944395
#> 41   9.792220 14.215198
#> 42  10.724107 16.162698
#> 43   9.723955 11.853637
#> 44  11.867172 14.868685
#> 45   9.462713 12.005517
#> 46   8.417439 16.170925
#> 47  11.694947 11.913272
#> 48   9.783619 15.976626
#> 49   7.810727 13.984781
#> 50   5.589841 10.348084
#> 51   9.338426 15.018089
#> 52   9.171589 14.116097
#> 53  13.216286 17.549513
#> 54   8.325149 12.812848
#> 55   7.593651 18.000674
#> 56   8.777604 13.975025
#> 57   8.890623 16.736588
#> 58  11.068520 10.322416
#> 59  11.028493 17.230290
#> 60   7.771067 16.935474
#> 61  12.241019 14.111416
#> 62  12.232828 12.673303
#> 63  11.859715 10.878273
#> 64  10.863607 12.646655
#> 65   9.293865  6.252061
#> 66  12.309485 17.562880
#> 67   9.425402 14.872808
#> 68   7.294368 17.676825
#> 69   9.027313 13.704298
#> 70   9.244010 12.140135
#> 71  10.315597 17.588271
#> 72   7.482679 13.579633
#> 73   7.887133 12.362669
#> 74  10.867698 17.131006
#> 75  10.793098 16.946039
#> 76   8.344616 12.731451
#> 77  12.017646 20.739101
#> 78  11.445803 18.578711
#> 79   9.801106 11.336545
#> 80  15.081621 16.430858
#> 81  12.998034 15.075573
#> 82   7.854737 11.847093
#> 83  10.824745 14.994232
#> 84  13.901937 13.234428
#> 85  10.905157 17.379135
#> 86   7.905843 14.756253
#> 87   9.178311 18.157743
#> 88   8.720429 12.471440
#> 89   8.388292 15.985588
#> 90  10.156951 13.841807
#> 91   6.165200 13.457487
#> 92   8.858229 13.287433
#> 93  11.074452 14.227778
#> 94   7.105353 18.791416
#> 95  11.510528 15.356714
#> 96  12.590685 17.701177
#> 97  10.438819 20.701206
#> 98  10.164032  9.922782
#> 99   7.620928 23.430470
#> 100  9.934664 17.145230
#> 
#> 
#> $misty.uniqueid
#> [1] "7ea5df9d1da9f18b77fa9818ea2605dd"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   14.1011873 19.742848
#> 2   12.2156412 18.819131
#> 3   11.3501363 15.790395
#> 4    5.1320621 16.260031
#> 5   19.7665612 11.656634
#> 6   11.0273285 21.331675
#> 7    9.4343160 18.258239
#> 8   11.9887273 14.467727
#> 9   15.2295013 12.611135
#> 10   9.0849375 20.102323
#> 11  13.2966287  9.119595
#> 12  10.5631967 13.884667
#> 13  12.6352429  9.164188
#> 14  16.5655873 18.238783
#> 15   8.4316781 20.393283
#> 16  11.5146501 17.518505
#> 17  -0.6335323 12.818916
#> 18  16.3791388 19.421143
#> 19   3.2253939 17.153230
#> 20  10.7758665 14.131669
#> 21  18.2456118 13.203442
#> 22  16.2246284 22.298338
#> 23  11.1702150 11.479576
#> 24   9.8578328 11.526532
#> 25   5.1709759 11.996685
#> 26  13.5603120  5.654598
#> 27  13.8623090 17.205546
#> 28  10.1823029 16.869211
#> 29   2.2216365  9.875181
#> 30   7.9946435  5.398114
#> 31  12.3859983 18.237705
#> 32   2.7121858 15.201699
#> 33  13.6821971 16.577671
#> 34  16.9540054  4.719504
#> 35   7.6074471 13.123602
#> 36   7.2302758 20.186299
#> 37   9.6570519 16.509709
#> 38   7.7847473 11.521746
#> 39   9.9265767 12.477288
#> 40   7.0766853 15.815182
#> 41  23.4009983 12.539164
#> 42   3.8531004 19.259427
#> 43  16.7496080 12.160188
#> 44   8.8677373 16.773691
#> 45   0.2157770 14.432193
#> 46  12.3072504 12.858021
#> 47  10.4122790 16.403665
#> 48  25.1553908 16.885414
#> 49   8.6051366 11.333473
#> 50   4.0915530 14.361771
#> 51   7.3150268 17.678935
#> 52  13.9941715 22.642420
#> 53   4.5416706 23.309709
#> 54  15.5134975 18.509465
#> 55  17.6645689 10.580293
#> 56   4.1520805 13.578944
#> 57  19.3104158 16.352046
#> 58  14.4342451 24.362638
#> 59  10.5187017 12.863263
#> 60  15.2318051 13.904168
#> 61   9.1130044  8.487530
#> 62  17.5372851  6.293770
#> 63   2.0480684 12.051733
#> 64  14.4911917 14.694430
#> 65  10.8248934 15.944544
#> 66  10.5487007 17.147310
#> 67  13.2788592 13.446042
#> 68  18.1013192 16.599150
#> 69   9.1167352 17.081623
#> 70  13.7287578 15.824537
#> 71   9.5376935  9.970923
#> 72   2.2555123 11.019431
#> 73   9.7649665 22.906063
#> 74  -1.3984734 21.682775
#> 75   8.4681389 14.114004
#> 76  16.5904423 13.844990
#> 77  10.9137513  4.791087
#> 78   7.8918079  9.561384
#> 79  13.6614243 10.343523
#> 80  17.4646592 10.191988
#> 81   5.6842623 19.017859
#> 82  10.0136834 24.063663
#> 83  16.3162814 20.653854
#> 84   8.5715147 14.407709
#> 85  12.5966896 15.885167
#> 86  11.8432804  9.393668
#> 87  17.5048808 13.413894
#> 88   9.8047243 12.528775
#> 89   2.8248108 14.917911
#> 90   8.8408710 19.679833
#> 91   1.9856638 25.684271
#> 92  13.9173288 12.809161
#> 93   1.5134808 23.695666
#> 94  11.5957740 20.030302
#> 95   8.7828765 12.402506
#> 96  11.4253554 13.028264
#> 97  13.8427296 15.663133
#> 98   4.4987190 10.819931
#> 99  12.4419760 18.862501
#> 100 15.0748376 10.528965
#> 
#>