Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    8.596023 19.657781
#> 2    9.466313 14.659598
#> 3    9.881802 16.186725
#> 4   10.020854 14.734156
#> 5   10.966107 17.194114
#> 6   10.163220 12.926808
#> 7   13.979367  8.513202
#> 8    8.178288 16.358545
#> 9    9.892375 12.763638
#> 10   8.111392 11.190433
#> 11   8.426143 13.326661
#> 12  11.396180 13.097094
#> 13   9.945130 16.994282
#> 14  10.663248 15.672013
#> 15   7.929667  9.894868
#> 16   9.430216 15.167340
#> 17  12.482863 13.873196
#> 18  10.320205 12.491106
#> 19   7.712228 18.238474
#> 20   8.805123 16.596119
#> 21   8.281478 13.949694
#> 22   9.481291 15.226558
#> 23   5.709575 14.511161
#> 24   7.311326 12.799953
#> 25   9.404840 19.408792
#> 26   9.060685 14.336934
#> 27  11.828128 19.674543
#> 28   4.456874 11.254267
#> 29   8.338042 13.785603
#> 30  11.623143 16.687887
#> 31  11.534613 14.585136
#> 32  12.622254 19.119659
#> 33   7.248200 16.544226
#> 34  11.842744 14.551122
#> 35   9.537017 13.162780
#> 36   8.046398 23.170980
#> 37   7.487196 17.864758
#> 38  12.825033 13.879317
#> 39   7.181147 18.233772
#> 40  10.327658 19.354277
#> 41  10.176092 16.389047
#> 42  12.697738  9.896295
#> 43  12.029319 13.871100
#> 44  11.209920 13.613069
#> 45   7.278276 13.224864
#> 46  11.117034 14.013129
#> 47  10.862955 14.367715
#> 48  16.873290 14.219413
#> 49   8.920871 13.662170
#> 50  10.950914 15.905245
#> 51  13.115642 17.598868
#> 52  14.188761 17.832404
#> 53  10.396316 13.034425
#> 54   8.936958 16.029810
#> 55  11.595497 13.697626
#> 56   7.885133 14.792662
#> 57   8.088950 13.850151
#> 58  13.687237 18.711902
#> 59   7.967801 15.150923
#> 60  12.276044 14.324021
#> 61   9.887599 18.603268
#> 62  13.535965 14.076312
#> 63  11.213081 18.718789
#> 64   4.461386 13.133223
#> 65   9.057756 11.171698
#> 66  10.126197 13.817162
#> 67  10.493494 12.376719
#> 68  10.053853 15.244898
#> 69  11.998775 11.883986
#> 70   8.731713 19.572808
#> 71  10.008972 13.252067
#> 72   7.465988 17.552411
#> 73  12.217631 15.807396
#> 74  10.346211 13.312015
#> 75  10.318242 20.445625
#> 76  10.550908 11.423251
#> 77   8.818203  8.403569
#> 78  10.272786 14.463441
#> 79   8.560380 13.594284
#> 80   9.295969 15.820120
#> 81   9.968761 16.403428
#> 82   9.761316 15.331751
#> 83   9.897964 14.358367
#> 84   7.458657 15.146246
#> 85  11.440233 12.892068
#> 86  11.890540 13.321090
#> 87   8.664927 12.504430
#> 88   8.378082 11.469550
#> 89  10.339366  9.722367
#> 90   9.712107 14.652241
#> 91  13.622534 17.291482
#> 92  14.091377 13.130288
#> 93   9.224685 18.083180
#> 94   7.555096 14.796887
#> 95  11.528002 18.383414
#> 96  12.041716  8.699767
#> 97  10.147825 13.621041
#> 98   6.533522 19.235877
#> 99  10.217667 13.996772
#> 100 11.260443 11.316881
#> 
#> 
#> $misty.uniqueid
#> [1] "aa6b4bc054a810b60488840f95af0466"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>           marker1    marker2
#> 1    12.407050302 17.4618395
#> 2    13.117675090 13.4633146
#> 3    12.061583682 11.6821722
#> 4    14.769949727  5.7612506
#> 5    20.011399398 26.2315143
#> 6     3.047806419 11.5271793
#> 7     7.712229386 13.7107177
#> 8    11.550369154 13.3070396
#> 9    13.203764603 16.2444648
#> 10    8.230319521 13.8581296
#> 11   12.111574823 23.8069341
#> 12   11.287168223 16.8976417
#> 13   17.560717710 19.0611530
#> 14    2.869409418 13.0190612
#> 15   12.252488400 17.5483812
#> 16    7.991326590 13.0490768
#> 17    8.332067487 12.5602868
#> 18   18.293090467 13.7788578
#> 19   11.929593155 16.0806216
#> 20    5.450882262 12.6877139
#> 21    8.193929127 18.4065469
#> 22   12.185582271 16.9420380
#> 23   12.547001707  9.2371299
#> 24    9.975581655 10.5474421
#> 25   10.067844533  5.0716312
#> 26   18.752478447 18.4392623
#> 27   10.111846742 17.2579245
#> 28   10.301721849 20.5783021
#> 29   -0.002588611 17.2092803
#> 30    4.700828847 11.8571819
#> 31   11.037445474 17.2537238
#> 32    4.804932533 18.8318240
#> 33   12.076442898  9.5884201
#> 34    4.548290477 13.3580783
#> 35    6.792017853  0.8487779
#> 36   11.773836741 15.3019866
#> 37    3.890854952 12.6198863
#> 38    1.029390146 15.8016656
#> 39   12.611194454 17.2889677
#> 40   14.829181419 18.0498593
#> 41   11.545935973 11.2873273
#> 42   19.304732194  9.6373435
#> 43    9.083615476 16.7452491
#> 44    9.068899626 26.1212064
#> 45    6.804742745 11.2337746
#> 46    2.144255431 16.9581637
#> 47   13.807881111 26.7037786
#> 48   11.735651339 17.3905270
#> 49    5.155463890 16.3586202
#> 50    1.669881895 15.3777056
#> 51   26.270450011 14.6272137
#> 52    6.856734404 12.1239438
#> 53   11.550392369 17.7562331
#> 54   18.780200591 23.7047249
#> 55   10.025723316 11.6997124
#> 56    8.773215570 18.8029704
#> 57   13.672626839  8.0994957
#> 58    8.010517497 18.6155341
#> 59   17.022172746 13.7230562
#> 60    6.736570275 22.1427864
#> 61   10.187790855 19.4274954
#> 62    7.135744381 17.9047359
#> 63   15.389565502 12.1659687
#> 64   11.402062024 23.7161777
#> 65    8.233860132 21.1194790
#> 66    2.786591574 23.4376491
#> 67  -11.067507565 11.3842731
#> 68   19.743662612 24.9306926
#> 69    7.212460937 15.5706546
#> 70    5.385550268 13.5095917
#> 71   10.755904303  9.8135460
#> 72   18.028548105 10.6066426
#> 73   12.229268179  3.3716751
#> 74    5.698802618 10.5466285
#> 75   11.315649741  6.5340359
#> 76    8.750336911 14.3357142
#> 77   15.024863031  7.9122516
#> 78    5.499653555 10.2607795
#> 79   10.265185894 21.0453388
#> 80   16.718946655 23.7946564
#> 81   15.219623887 15.7894341
#> 82    5.609104399  7.7860052
#> 83    1.379921544 18.0048420
#> 84    4.519856021 12.2350312
#> 85   10.145079936 17.4326868
#> 86    9.034774468 20.3491319
#> 87    8.297531567 10.2630072
#> 88   15.558883034 15.6505592
#> 89    3.210779970 11.0856304
#> 90    9.795424596 10.7954759
#> 91   -2.516975649 21.8882198
#> 92    9.691950980 10.4173852
#> 93   16.409444997 19.5539471
#> 94   11.053514876 13.1257214
#> 95   15.095502679 16.7555720
#> 96    8.054856180 20.3731842
#> 97    5.363771279  6.8815121
#> 98    6.570511227  7.0474321
#> 99    5.087443620 17.7398837
#> 100   8.670908817 28.3263827
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    8.596023 19.657781
#> 2    9.466313 14.659598
#> 3    9.881802 16.186725
#> 4   10.020854 14.734156
#> 5   10.966107 17.194114
#> 6   10.163220 12.926808
#> 7   13.979367  8.513202
#> 8    8.178288 16.358545
#> 9    9.892375 12.763638
#> 10   8.111392 11.190433
#> 11   8.426143 13.326661
#> 12  11.396180 13.097094
#> 13   9.945130 16.994282
#> 14  10.663248 15.672013
#> 15   7.929667  9.894868
#> 16   9.430216 15.167340
#> 17  12.482863 13.873196
#> 18  10.320205 12.491106
#> 19   7.712228 18.238474
#> 20   8.805123 16.596119
#> 21   8.281478 13.949694
#> 22   9.481291 15.226558
#> 23   5.709575 14.511161
#> 24   7.311326 12.799953
#> 25   9.404840 19.408792
#> 26   9.060685 14.336934
#> 27  11.828128 19.674543
#> 28   4.456874 11.254267
#> 29   8.338042 13.785603
#> 30  11.623143 16.687887
#> 31  11.534613 14.585136
#> 32  12.622254 19.119659
#> 33   7.248200 16.544226
#> 34  11.842744 14.551122
#> 35   9.537017 13.162780
#> 36   8.046398 23.170980
#> 37   7.487196 17.864758
#> 38  12.825033 13.879317
#> 39   7.181147 18.233772
#> 40  10.327658 19.354277
#> 41  10.176092 16.389047
#> 42  12.697738  9.896295
#> 43  12.029319 13.871100
#> 44  11.209920 13.613069
#> 45   7.278276 13.224864
#> 46  11.117034 14.013129
#> 47  10.862955 14.367715
#> 48  16.873290 14.219413
#> 49   8.920871 13.662170
#> 50  10.950914 15.905245
#> 51  13.115642 17.598868
#> 52  14.188761 17.832404
#> 53  10.396316 13.034425
#> 54   8.936958 16.029810
#> 55  11.595497 13.697626
#> 56   7.885133 14.792662
#> 57   8.088950 13.850151
#> 58  13.687237 18.711902
#> 59   7.967801 15.150923
#> 60  12.276044 14.324021
#> 61   9.887599 18.603268
#> 62  13.535965 14.076312
#> 63  11.213081 18.718789
#> 64   4.461386 13.133223
#> 65   9.057756 11.171698
#> 66  10.126197 13.817162
#> 67  10.493494 12.376719
#> 68  10.053853 15.244898
#> 69  11.998775 11.883986
#> 70   8.731713 19.572808
#> 71  10.008972 13.252067
#> 72   7.465988 17.552411
#> 73  12.217631 15.807396
#> 74  10.346211 13.312015
#> 75  10.318242 20.445625
#> 76  10.550908 11.423251
#> 77   8.818203  8.403569
#> 78  10.272786 14.463441
#> 79   8.560380 13.594284
#> 80   9.295969 15.820120
#> 81   9.968761 16.403428
#> 82   9.761316 15.331751
#> 83   9.897964 14.358367
#> 84   7.458657 15.146246
#> 85  11.440233 12.892068
#> 86  11.890540 13.321090
#> 87   8.664927 12.504430
#> 88   8.378082 11.469550
#> 89  10.339366  9.722367
#> 90   9.712107 14.652241
#> 91  13.622534 17.291482
#> 92  14.091377 13.130288
#> 93   9.224685 18.083180
#> 94   7.555096 14.796887
#> 95  11.528002 18.383414
#> 96  12.041716  8.699767
#> 97  10.147825 13.621041
#> 98   6.533522 19.235877
#> 99  10.217667 13.996772
#> 100 11.260443 11.316881
#> 
#> 
#> $misty.uniqueid
#> [1] "aa6b4bc054a810b60488840f95af0466"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>           marker1    marker2
#> 1    12.407050302 17.4618395
#> 2    13.117675090 13.4633146
#> 3    12.061583682 11.6821722
#> 4    14.769949727  5.7612506
#> 5    20.011399398 26.2315143
#> 6     3.047806419 11.5271793
#> 7     7.712229386 13.7107177
#> 8    11.550369154 13.3070396
#> 9    13.203764603 16.2444648
#> 10    8.230319521 13.8581296
#> 11   12.111574823 23.8069341
#> 12   11.287168223 16.8976417
#> 13   17.560717710 19.0611530
#> 14    2.869409418 13.0190612
#> 15   12.252488400 17.5483812
#> 16    7.991326590 13.0490768
#> 17    8.332067487 12.5602868
#> 18   18.293090467 13.7788578
#> 19   11.929593155 16.0806216
#> 20    5.450882262 12.6877139
#> 21    8.193929127 18.4065469
#> 22   12.185582271 16.9420380
#> 23   12.547001707  9.2371299
#> 24    9.975581655 10.5474421
#> 25   10.067844533  5.0716312
#> 26   18.752478447 18.4392623
#> 27   10.111846742 17.2579245
#> 28   10.301721849 20.5783021
#> 29   -0.002588611 17.2092803
#> 30    4.700828847 11.8571819
#> 31   11.037445474 17.2537238
#> 32    4.804932533 18.8318240
#> 33   12.076442898  9.5884201
#> 34    4.548290477 13.3580783
#> 35    6.792017853  0.8487779
#> 36   11.773836741 15.3019866
#> 37    3.890854952 12.6198863
#> 38    1.029390146 15.8016656
#> 39   12.611194454 17.2889677
#> 40   14.829181419 18.0498593
#> 41   11.545935973 11.2873273
#> 42   19.304732194  9.6373435
#> 43    9.083615476 16.7452491
#> 44    9.068899626 26.1212064
#> 45    6.804742745 11.2337746
#> 46    2.144255431 16.9581637
#> 47   13.807881111 26.7037786
#> 48   11.735651339 17.3905270
#> 49    5.155463890 16.3586202
#> 50    1.669881895 15.3777056
#> 51   26.270450011 14.6272137
#> 52    6.856734404 12.1239438
#> 53   11.550392369 17.7562331
#> 54   18.780200591 23.7047249
#> 55   10.025723316 11.6997124
#> 56    8.773215570 18.8029704
#> 57   13.672626839  8.0994957
#> 58    8.010517497 18.6155341
#> 59   17.022172746 13.7230562
#> 60    6.736570275 22.1427864
#> 61   10.187790855 19.4274954
#> 62    7.135744381 17.9047359
#> 63   15.389565502 12.1659687
#> 64   11.402062024 23.7161777
#> 65    8.233860132 21.1194790
#> 66    2.786591574 23.4376491
#> 67  -11.067507565 11.3842731
#> 68   19.743662612 24.9306926
#> 69    7.212460937 15.5706546
#> 70    5.385550268 13.5095917
#> 71   10.755904303  9.8135460
#> 72   18.028548105 10.6066426
#> 73   12.229268179  3.3716751
#> 74    5.698802618 10.5466285
#> 75   11.315649741  6.5340359
#> 76    8.750336911 14.3357142
#> 77   15.024863031  7.9122516
#> 78    5.499653555 10.2607795
#> 79   10.265185894 21.0453388
#> 80   16.718946655 23.7946564
#> 81   15.219623887 15.7894341
#> 82    5.609104399  7.7860052
#> 83    1.379921544 18.0048420
#> 84    4.519856021 12.2350312
#> 85   10.145079936 17.4326868
#> 86    9.034774468 20.3491319
#> 87    8.297531567 10.2630072
#> 88   15.558883034 15.6505592
#> 89    3.210779970 11.0856304
#> 90    9.795424596 10.7954759
#> 91   -2.516975649 21.8882198
#> 92    9.691950980 10.4173852
#> 93   16.409444997 19.5539471
#> 94   11.053514876 13.1257214
#> 95   15.095502679 16.7555720
#> 96    8.054856180 20.3731842
#> 97    5.363771279  6.8815121
#> 98    6.570511227  7.0474321
#> 99    5.087443620 17.7398837
#> 100   8.670908817 28.3263827
#> 
#>