Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.906932 19.571650
#> 2   11.182134 15.630704
#> 3   10.324549 17.695439
#> 4   13.892730 14.812309
#> 5   13.538684 14.508501
#> 6   12.786369 15.140426
#> 7   10.858132 11.462156
#> 8    7.666336 13.986407
#> 9    9.384872 16.961758
#> 10   6.986754 14.724923
#> 11   8.027162 15.439420
#> 12  11.938750 13.982282
#> 13  12.304716 14.848763
#> 14   9.392653 13.604254
#> 15  10.413548 16.321333
#> 16   7.336524 14.350925
#> 17   9.866426 11.188312
#> 18  12.202401 14.023282
#> 19   8.404335 15.988154
#> 20  12.023306 16.989569
#> 21  12.729700 16.546465
#> 22  11.483378 15.850786
#> 23  11.749579 14.476399
#> 24  10.256061 10.250145
#> 25   9.040087 13.721235
#> 26   8.503243 14.824441
#> 27   5.742269 19.523974
#> 28   8.859278 14.300025
#> 29  10.902306 13.764064
#> 30   9.583295 15.468381
#> 31  10.360570 13.939756
#> 32  10.630274 22.669428
#> 33  10.206739 21.089988
#> 34   5.997040 13.360453
#> 35   9.589628 11.452581
#> 36  11.364351 14.490449
#> 37  10.877040 15.593094
#> 38  10.249391 16.090701
#> 39   7.404172 15.497029
#> 40   7.146917  8.155525
#> 41   6.652361 16.720351
#> 42  14.741705 15.252583
#> 43   8.354054 15.230957
#> 44   8.241494 18.694400
#> 45   6.098081 15.356324
#> 46   9.462565 15.799903
#> 47  12.474838 11.893896
#> 48  15.245362 13.298906
#> 49   8.908757 15.373315
#> 50   6.993682 10.893628
#> 51   9.995674 15.474782
#> 52   9.699655 15.995029
#> 53   9.879215 16.620933
#> 54  10.703552 15.882448
#> 55  10.416912 15.146961
#> 56  13.634964 17.445457
#> 57  10.666941 17.988573
#> 58  10.326750 13.622185
#> 59   7.778148 16.872704
#> 60   9.174523 13.439443
#> 61  11.725702 15.614752
#> 62  13.509493 17.076636
#> 63   9.147984 11.700780
#> 64  14.615016 17.895762
#> 65  10.489381 20.938900
#> 66  11.391555 15.385872
#> 67   9.937781 15.961602
#> 68  10.961455 17.562784
#> 69   7.574845 20.067703
#> 70   9.618207 14.010115
#> 71  11.703754 14.055106
#> 72  13.138668 18.087473
#> 73  10.822585 19.503463
#> 74   8.348947 12.615397
#> 75  10.394753 18.859605
#> 76  10.874695 14.459568
#> 77  10.010478 11.255884
#> 78  11.691411 19.118779
#> 79   9.944648 12.768662
#> 80  12.022190 17.775252
#> 81  11.808342 12.628196
#> 82   7.522992 16.554401
#> 83  11.457745 16.138213
#> 84  10.005667 21.513217
#> 85   7.316608 13.326687
#> 86  12.148382 16.248440
#> 87  10.919711 14.617441
#> 88   8.255032 16.688279
#> 89  10.718987 11.302949
#> 90   8.840171 18.362213
#> 91  10.442186 15.095267
#> 92  12.450757 20.068387
#> 93  10.734499 16.820464
#> 94   9.147626 14.168584
#> 95   6.539146 13.611178
#> 96   9.461107 20.373374
#> 97   8.500355 18.503853
#> 98  10.947094 14.487601
#> 99  13.214583 14.380480
#> 100  8.879677 12.999821
#> 
#> 
#> $misty.uniqueid
#> [1] "6c36b05b076ec487c7834c060cda91bd"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    4.9857832  8.266761
#> 2    9.1024124 15.194326
#> 3   11.1611958 14.879411
#> 4   10.5399848  9.366488
#> 5   10.4032781 18.505310
#> 6    7.9286017 12.739305
#> 7   14.7546420 15.541335
#> 8   -0.8344226  7.928416
#> 9    7.6025395 19.666701
#> 10   8.5512089 16.752552
#> 11  13.8072922 10.594795
#> 12   8.8125124 11.309923
#> 13   5.7460526 16.857770
#> 14   4.6580039 15.632963
#> 15   7.8483174 13.306495
#> 16  17.6831649 10.327271
#> 17   6.8813816 21.669331
#> 18   5.5811981 17.048850
#> 19   5.6776978 14.120399
#> 20  15.3849857  9.233974
#> 21  16.7299042 17.095213
#> 22   8.8674387 16.306219
#> 23   7.2689054  3.467780
#> 24  17.8162897 16.122038
#> 25   3.9801413 29.509929
#> 26  21.0927022 10.674093
#> 27  13.0949395 14.267646
#> 28  10.1607752 17.736961
#> 29  -0.4527783  7.504429
#> 30  16.7411118 22.075583
#> 31   1.1121882 16.390081
#> 32   8.6576391 16.674211
#> 33  10.5781600 14.001717
#> 34  13.5420884 18.182269
#> 35   8.7908981 20.342485
#> 36  11.3119752 11.742858
#> 37  14.2314228 12.511316
#> 38  12.2969158 14.361997
#> 39  10.1173148 22.413785
#> 40  11.1110077 14.680054
#> 41  11.0174332 17.429599
#> 42  14.8302544 11.811529
#> 43  10.3499883  7.765548
#> 44  10.4303013 19.864739
#> 45  18.2715604 19.368236
#> 46  12.5026848 18.996842
#> 47   9.0004221 28.134218
#> 48   6.1420347 22.537401
#> 49  10.3873417 18.192311
#> 50  -3.3265928 12.367587
#> 51   9.6474414 14.309422
#> 52   0.5878761 18.768527
#> 53  -1.5960471 16.417658
#> 54   6.4698428 15.777875
#> 55  20.5374830 14.752683
#> 56  10.0322728 20.699537
#> 57   6.6981110 16.949358
#> 58   8.0336767 13.760694
#> 59  10.8010237 14.523483
#> 60   3.3096265 18.278783
#> 61  13.7329058 15.283707
#> 62  12.9189346  9.445141
#> 63  11.7298879 18.261793
#> 64   4.5974013 10.370896
#> 65  13.8322976 25.123311
#> 66  17.2482177 21.564891
#> 67   9.4306062 12.519830
#> 68   6.3291985 17.072694
#> 69   6.3502635 11.740213
#> 70   4.2190081 12.416360
#> 71   2.7036297 18.327385
#> 72   6.4596874 17.643834
#> 73  10.1244812 19.439730
#> 74   4.3049871 14.828416
#> 75  14.9506624 11.559381
#> 76  12.2615372  4.549640
#> 77  19.2365631 15.205167
#> 78  16.4728018 10.943410
#> 79   9.9906005 12.213122
#> 80  11.5539114 21.025484
#> 81   5.4454741 11.310352
#> 82  13.4051229 19.478483
#> 83  18.4394493 19.611164
#> 84   6.4903875 23.244311
#> 85   8.4202304 16.813545
#> 86   7.0889250 15.268535
#> 87   5.3856594 13.834851
#> 88   8.1140746 16.449152
#> 89  20.8046504 18.135500
#> 90  15.8299939 11.006019
#> 91  14.3535469 17.878682
#> 92   9.6825608 25.688866
#> 93  -0.1357886 13.817663
#> 94   8.5167786 16.149210
#> 95  11.6131575  9.503652
#> 96  12.9272898 21.472363
#> 97   8.1025985 16.494157
#> 98   9.6255757 19.456922
#> 99  10.8523193 19.178240
#> 100 10.9796551 17.948193
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.906932 19.571650
#> 2   11.182134 15.630704
#> 3   10.324549 17.695439
#> 4   13.892730 14.812309
#> 5   13.538684 14.508501
#> 6   12.786369 15.140426
#> 7   10.858132 11.462156
#> 8    7.666336 13.986407
#> 9    9.384872 16.961758
#> 10   6.986754 14.724923
#> 11   8.027162 15.439420
#> 12  11.938750 13.982282
#> 13  12.304716 14.848763
#> 14   9.392653 13.604254
#> 15  10.413548 16.321333
#> 16   7.336524 14.350925
#> 17   9.866426 11.188312
#> 18  12.202401 14.023282
#> 19   8.404335 15.988154
#> 20  12.023306 16.989569
#> 21  12.729700 16.546465
#> 22  11.483378 15.850786
#> 23  11.749579 14.476399
#> 24  10.256061 10.250145
#> 25   9.040087 13.721235
#> 26   8.503243 14.824441
#> 27   5.742269 19.523974
#> 28   8.859278 14.300025
#> 29  10.902306 13.764064
#> 30   9.583295 15.468381
#> 31  10.360570 13.939756
#> 32  10.630274 22.669428
#> 33  10.206739 21.089988
#> 34   5.997040 13.360453
#> 35   9.589628 11.452581
#> 36  11.364351 14.490449
#> 37  10.877040 15.593094
#> 38  10.249391 16.090701
#> 39   7.404172 15.497029
#> 40   7.146917  8.155525
#> 41   6.652361 16.720351
#> 42  14.741705 15.252583
#> 43   8.354054 15.230957
#> 44   8.241494 18.694400
#> 45   6.098081 15.356324
#> 46   9.462565 15.799903
#> 47  12.474838 11.893896
#> 48  15.245362 13.298906
#> 49   8.908757 15.373315
#> 50   6.993682 10.893628
#> 51   9.995674 15.474782
#> 52   9.699655 15.995029
#> 53   9.879215 16.620933
#> 54  10.703552 15.882448
#> 55  10.416912 15.146961
#> 56  13.634964 17.445457
#> 57  10.666941 17.988573
#> 58  10.326750 13.622185
#> 59   7.778148 16.872704
#> 60   9.174523 13.439443
#> 61  11.725702 15.614752
#> 62  13.509493 17.076636
#> 63   9.147984 11.700780
#> 64  14.615016 17.895762
#> 65  10.489381 20.938900
#> 66  11.391555 15.385872
#> 67   9.937781 15.961602
#> 68  10.961455 17.562784
#> 69   7.574845 20.067703
#> 70   9.618207 14.010115
#> 71  11.703754 14.055106
#> 72  13.138668 18.087473
#> 73  10.822585 19.503463
#> 74   8.348947 12.615397
#> 75  10.394753 18.859605
#> 76  10.874695 14.459568
#> 77  10.010478 11.255884
#> 78  11.691411 19.118779
#> 79   9.944648 12.768662
#> 80  12.022190 17.775252
#> 81  11.808342 12.628196
#> 82   7.522992 16.554401
#> 83  11.457745 16.138213
#> 84  10.005667 21.513217
#> 85   7.316608 13.326687
#> 86  12.148382 16.248440
#> 87  10.919711 14.617441
#> 88   8.255032 16.688279
#> 89  10.718987 11.302949
#> 90   8.840171 18.362213
#> 91  10.442186 15.095267
#> 92  12.450757 20.068387
#> 93  10.734499 16.820464
#> 94   9.147626 14.168584
#> 95   6.539146 13.611178
#> 96   9.461107 20.373374
#> 97   8.500355 18.503853
#> 98  10.947094 14.487601
#> 99  13.214583 14.380480
#> 100  8.879677 12.999821
#> 
#> 
#> $misty.uniqueid
#> [1] "6c36b05b076ec487c7834c060cda91bd"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    4.9857832  8.266761
#> 2    9.1024124 15.194326
#> 3   11.1611958 14.879411
#> 4   10.5399848  9.366488
#> 5   10.4032781 18.505310
#> 6    7.9286017 12.739305
#> 7   14.7546420 15.541335
#> 8   -0.8344226  7.928416
#> 9    7.6025395 19.666701
#> 10   8.5512089 16.752552
#> 11  13.8072922 10.594795
#> 12   8.8125124 11.309923
#> 13   5.7460526 16.857770
#> 14   4.6580039 15.632963
#> 15   7.8483174 13.306495
#> 16  17.6831649 10.327271
#> 17   6.8813816 21.669331
#> 18   5.5811981 17.048850
#> 19   5.6776978 14.120399
#> 20  15.3849857  9.233974
#> 21  16.7299042 17.095213
#> 22   8.8674387 16.306219
#> 23   7.2689054  3.467780
#> 24  17.8162897 16.122038
#> 25   3.9801413 29.509929
#> 26  21.0927022 10.674093
#> 27  13.0949395 14.267646
#> 28  10.1607752 17.736961
#> 29  -0.4527783  7.504429
#> 30  16.7411118 22.075583
#> 31   1.1121882 16.390081
#> 32   8.6576391 16.674211
#> 33  10.5781600 14.001717
#> 34  13.5420884 18.182269
#> 35   8.7908981 20.342485
#> 36  11.3119752 11.742858
#> 37  14.2314228 12.511316
#> 38  12.2969158 14.361997
#> 39  10.1173148 22.413785
#> 40  11.1110077 14.680054
#> 41  11.0174332 17.429599
#> 42  14.8302544 11.811529
#> 43  10.3499883  7.765548
#> 44  10.4303013 19.864739
#> 45  18.2715604 19.368236
#> 46  12.5026848 18.996842
#> 47   9.0004221 28.134218
#> 48   6.1420347 22.537401
#> 49  10.3873417 18.192311
#> 50  -3.3265928 12.367587
#> 51   9.6474414 14.309422
#> 52   0.5878761 18.768527
#> 53  -1.5960471 16.417658
#> 54   6.4698428 15.777875
#> 55  20.5374830 14.752683
#> 56  10.0322728 20.699537
#> 57   6.6981110 16.949358
#> 58   8.0336767 13.760694
#> 59  10.8010237 14.523483
#> 60   3.3096265 18.278783
#> 61  13.7329058 15.283707
#> 62  12.9189346  9.445141
#> 63  11.7298879 18.261793
#> 64   4.5974013 10.370896
#> 65  13.8322976 25.123311
#> 66  17.2482177 21.564891
#> 67   9.4306062 12.519830
#> 68   6.3291985 17.072694
#> 69   6.3502635 11.740213
#> 70   4.2190081 12.416360
#> 71   2.7036297 18.327385
#> 72   6.4596874 17.643834
#> 73  10.1244812 19.439730
#> 74   4.3049871 14.828416
#> 75  14.9506624 11.559381
#> 76  12.2615372  4.549640
#> 77  19.2365631 15.205167
#> 78  16.4728018 10.943410
#> 79   9.9906005 12.213122
#> 80  11.5539114 21.025484
#> 81   5.4454741 11.310352
#> 82  13.4051229 19.478483
#> 83  18.4394493 19.611164
#> 84   6.4903875 23.244311
#> 85   8.4202304 16.813545
#> 86   7.0889250 15.268535
#> 87   5.3856594 13.834851
#> 88   8.1140746 16.449152
#> 89  20.8046504 18.135500
#> 90  15.8299939 11.006019
#> 91  14.3535469 17.878682
#> 92   9.6825608 25.688866
#> 93  -0.1357886 13.817663
#> 94   8.5167786 16.149210
#> 95  11.6131575  9.503652
#> 96  12.9272898 21.472363
#> 97   8.1025985 16.494157
#> 98   9.6255757 19.456922
#> 99  10.8523193 19.178240
#> 100 10.9796551 17.948193
#> 
#>