Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.893553 19.336713
#> 2    8.716630 13.037363
#> 3   11.318842 18.044354
#> 4   12.089139 11.356848
#> 5    8.050416  9.815124
#> 6   11.528261 16.979143
#> 7   12.932232 21.161485
#> 8   10.483600 11.759536
#> 9   10.115427 21.014715
#> 10   8.933661 12.817106
#> 11   9.383509 17.088257
#> 12   9.551742 16.567882
#> 13   9.538110 16.316283
#> 14   9.868694 12.488904
#> 15   9.543073 20.235168
#> 16   8.657915 15.096421
#> 17  10.000277 13.396191
#> 18  10.121528 12.009003
#> 19  12.392267 11.016746
#> 20   8.655345 20.002415
#> 21   9.152732 15.220524
#> 22  12.914748 12.330051
#> 23  10.345900 19.948571
#> 24   5.441971 19.699903
#> 25   6.787275  5.801943
#> 26   7.517273 14.503135
#> 27   6.385582 15.314120
#> 28   6.930604 15.127042
#> 29   9.466654 13.973768
#> 30  12.092599 16.142662
#> 31  10.868484 16.420559
#> 32  10.886872 18.701405
#> 33  11.458173 14.594859
#> 34  10.197605 16.979867
#> 35  12.316712 17.023622
#> 36  12.787968 12.655921
#> 37  10.048792  9.934174
#> 38   5.996131 17.347107
#> 39  11.588354 12.324761
#> 40   7.994129 12.300068
#> 41   9.350278 15.449147
#> 42  10.307255 14.842578
#> 43   8.542882 13.827592
#> 44   7.841134 15.254639
#> 45   7.911655 14.159296
#> 46   8.772467 14.685144
#> 47  11.607896 13.260652
#> 48   8.668603 13.603861
#> 49   4.592925 17.741996
#> 50   9.871547 13.053868
#> 51   8.332126 15.127475
#> 52  12.916660 15.873705
#> 53   8.765162 21.329606
#> 54   9.810293 16.866612
#> 55   6.887290 17.296013
#> 56   8.620576 16.492752
#> 57  12.352788 13.629430
#> 58  11.110800 15.214609
#> 59   9.890170 16.170345
#> 60  11.379065 14.383225
#> 61  10.595063 12.330635
#> 62  13.366918 11.356003
#> 63  11.613177 13.335899
#> 64  10.996199 15.396415
#> 65   7.300979 16.199191
#> 66  11.875745 12.796886
#> 67  11.372694 18.044533
#> 68  11.366189 11.903920
#> 69  13.006432 14.012061
#> 70  11.802448 15.966445
#> 71   8.990203 16.535228
#> 72  10.944799 15.240025
#> 73   8.545093  9.006820
#> 74  10.085778 16.180921
#> 75  11.408121 15.454490
#> 76   8.389366 13.658302
#> 77  14.133595 17.771521
#> 78  12.044818  9.288756
#> 79  11.108889 17.008302
#> 80   6.254854 18.775168
#> 81   9.007435 12.773444
#> 82   8.638840 12.306843
#> 83   9.584633 19.612040
#> 84   9.589058 20.425620
#> 85  10.562234 13.913311
#> 86   7.438105 16.566215
#> 87  11.402290 16.111998
#> 88   8.167156 12.159125
#> 89  11.629100 18.569552
#> 90   8.512066 12.809143
#> 91  10.810755 16.698804
#> 92   9.646116 11.213771
#> 93  12.759592 11.329508
#> 94  10.935274 15.051273
#> 95  10.306820 15.951924
#> 96  10.917398 12.797590
#> 97   9.233312 14.655319
#> 98  12.870451 14.839605
#> 99  12.267915 14.733885
#> 100  9.691086 15.428780
#> 
#> 
#> $misty.uniqueid
#> [1] "7f5e6f7969ad2b054ff3b7d5531e4c52"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    3.3086486 14.893949
#> 2   16.0361181 14.020024
#> 3    9.7172071 13.673548
#> 4    6.8306837 21.106980
#> 5   13.0639657 13.759436
#> 6    9.0011500  8.270088
#> 7   14.9228834 14.663006
#> 8    7.9514210 14.143656
#> 9   15.1681090 11.587082
#> 10   7.7732663  9.704692
#> 11  12.0420200 18.844784
#> 12   7.0494006 21.529089
#> 13  17.4981963 14.231938
#> 14   8.7591065 17.802864
#> 15  11.7420800 22.719638
#> 16  14.1794431 13.713949
#> 17  10.8933061  8.502784
#> 18   9.2346525  7.766908
#> 19   9.9145657  9.214681
#> 20   4.8142260 18.834355
#> 21   8.1659811 16.671785
#> 22  12.8583886  6.531424
#> 23   8.9576468 13.043432
#> 24   5.8965337 17.331890
#> 25  12.2775456 18.682129
#> 26  12.4584198  8.378604
#> 27   3.5878462 -1.488901
#> 28   8.3322811 16.390013
#> 29  -0.2286532 11.668675
#> 30  10.2104546 23.411720
#> 31  -0.9247928 12.923810
#> 32  14.6055343 14.219899
#> 33   8.0903588 18.872490
#> 34   7.8659796 19.787248
#> 35  11.3540152 14.039155
#> 36  14.0296091  9.977807
#> 37   9.7634632 18.411759
#> 38  10.5839389 18.762160
#> 39  11.8556398 20.534441
#> 40   6.8581685  8.134803
#> 41   7.4861817 19.580352
#> 42   6.1421205 13.329981
#> 43  18.4517222  7.955995
#> 44   6.0188309 17.317868
#> 45  18.2861212 17.195661
#> 46  12.9984510  3.412588
#> 47   8.4081198 12.003692
#> 48  19.2036255 20.824294
#> 49  10.7896437 14.210982
#> 50  10.5056723 13.923299
#> 51   8.5714421 10.489731
#> 52   6.9781209 11.432314
#> 53  12.3625277 14.729526
#> 54  11.5832401 27.804328
#> 55  11.5498026 17.722831
#> 56  15.5316761 22.224422
#> 57   3.4373386 14.539016
#> 58   7.9867981 18.172316
#> 59  13.4599305 13.055509
#> 60  10.9405330 12.114604
#> 61  -0.1949835 14.027883
#> 62   1.6695752  7.906043
#> 63   5.2938134 21.159029
#> 64   0.7664494 12.868675
#> 65   9.8382329 12.556003
#> 66  17.7917874 18.798992
#> 67  20.6824835  6.004403
#> 68   4.5930607  7.931406
#> 69  14.6333914 15.552621
#> 70  13.5744943 18.472358
#> 71  14.7955247 15.603963
#> 72  10.9113115  9.893060
#> 73  12.9718418 15.388286
#> 74   3.2622034 10.697569
#> 75  12.4166464 15.698106
#> 76   7.9435470 19.306237
#> 77   3.8064415 12.280630
#> 78  -3.4174992 17.261276
#> 79  19.6420254 13.935052
#> 80  12.1416772 19.856171
#> 81   5.1243474 15.129243
#> 82  14.4592107 16.155270
#> 83  14.4382337 21.636290
#> 84   5.4509838 12.731711
#> 85  16.0861979 19.637993
#> 86  17.1891628  7.306491
#> 87  15.0005903 16.337340
#> 88  13.2382652 14.685484
#> 89   9.1588828 23.682504
#> 90   2.2848105 14.338907
#> 91  11.3185039 14.454009
#> 92   9.7510659 10.853013
#> 93  18.2600208  9.080724
#> 94  12.7874515 13.826092
#> 95  18.1239639 12.534028
#> 96  17.9742560  8.334584
#> 97   6.4722372 20.270199
#> 98  -2.2690390 22.797751
#> 99   7.9284494 17.496032
#> 100 12.2192415 11.750491
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.893553 19.336713
#> 2    8.716630 13.037363
#> 3   11.318842 18.044354
#> 4   12.089139 11.356848
#> 5    8.050416  9.815124
#> 6   11.528261 16.979143
#> 7   12.932232 21.161485
#> 8   10.483600 11.759536
#> 9   10.115427 21.014715
#> 10   8.933661 12.817106
#> 11   9.383509 17.088257
#> 12   9.551742 16.567882
#> 13   9.538110 16.316283
#> 14   9.868694 12.488904
#> 15   9.543073 20.235168
#> 16   8.657915 15.096421
#> 17  10.000277 13.396191
#> 18  10.121528 12.009003
#> 19  12.392267 11.016746
#> 20   8.655345 20.002415
#> 21   9.152732 15.220524
#> 22  12.914748 12.330051
#> 23  10.345900 19.948571
#> 24   5.441971 19.699903
#> 25   6.787275  5.801943
#> 26   7.517273 14.503135
#> 27   6.385582 15.314120
#> 28   6.930604 15.127042
#> 29   9.466654 13.973768
#> 30  12.092599 16.142662
#> 31  10.868484 16.420559
#> 32  10.886872 18.701405
#> 33  11.458173 14.594859
#> 34  10.197605 16.979867
#> 35  12.316712 17.023622
#> 36  12.787968 12.655921
#> 37  10.048792  9.934174
#> 38   5.996131 17.347107
#> 39  11.588354 12.324761
#> 40   7.994129 12.300068
#> 41   9.350278 15.449147
#> 42  10.307255 14.842578
#> 43   8.542882 13.827592
#> 44   7.841134 15.254639
#> 45   7.911655 14.159296
#> 46   8.772467 14.685144
#> 47  11.607896 13.260652
#> 48   8.668603 13.603861
#> 49   4.592925 17.741996
#> 50   9.871547 13.053868
#> 51   8.332126 15.127475
#> 52  12.916660 15.873705
#> 53   8.765162 21.329606
#> 54   9.810293 16.866612
#> 55   6.887290 17.296013
#> 56   8.620576 16.492752
#> 57  12.352788 13.629430
#> 58  11.110800 15.214609
#> 59   9.890170 16.170345
#> 60  11.379065 14.383225
#> 61  10.595063 12.330635
#> 62  13.366918 11.356003
#> 63  11.613177 13.335899
#> 64  10.996199 15.396415
#> 65   7.300979 16.199191
#> 66  11.875745 12.796886
#> 67  11.372694 18.044533
#> 68  11.366189 11.903920
#> 69  13.006432 14.012061
#> 70  11.802448 15.966445
#> 71   8.990203 16.535228
#> 72  10.944799 15.240025
#> 73   8.545093  9.006820
#> 74  10.085778 16.180921
#> 75  11.408121 15.454490
#> 76   8.389366 13.658302
#> 77  14.133595 17.771521
#> 78  12.044818  9.288756
#> 79  11.108889 17.008302
#> 80   6.254854 18.775168
#> 81   9.007435 12.773444
#> 82   8.638840 12.306843
#> 83   9.584633 19.612040
#> 84   9.589058 20.425620
#> 85  10.562234 13.913311
#> 86   7.438105 16.566215
#> 87  11.402290 16.111998
#> 88   8.167156 12.159125
#> 89  11.629100 18.569552
#> 90   8.512066 12.809143
#> 91  10.810755 16.698804
#> 92   9.646116 11.213771
#> 93  12.759592 11.329508
#> 94  10.935274 15.051273
#> 95  10.306820 15.951924
#> 96  10.917398 12.797590
#> 97   9.233312 14.655319
#> 98  12.870451 14.839605
#> 99  12.267915 14.733885
#> 100  9.691086 15.428780
#> 
#> 
#> $misty.uniqueid
#> [1] "7f5e6f7969ad2b054ff3b7d5531e4c52"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    3.3086486 14.893949
#> 2   16.0361181 14.020024
#> 3    9.7172071 13.673548
#> 4    6.8306837 21.106980
#> 5   13.0639657 13.759436
#> 6    9.0011500  8.270088
#> 7   14.9228834 14.663006
#> 8    7.9514210 14.143656
#> 9   15.1681090 11.587082
#> 10   7.7732663  9.704692
#> 11  12.0420200 18.844784
#> 12   7.0494006 21.529089
#> 13  17.4981963 14.231938
#> 14   8.7591065 17.802864
#> 15  11.7420800 22.719638
#> 16  14.1794431 13.713949
#> 17  10.8933061  8.502784
#> 18   9.2346525  7.766908
#> 19   9.9145657  9.214681
#> 20   4.8142260 18.834355
#> 21   8.1659811 16.671785
#> 22  12.8583886  6.531424
#> 23   8.9576468 13.043432
#> 24   5.8965337 17.331890
#> 25  12.2775456 18.682129
#> 26  12.4584198  8.378604
#> 27   3.5878462 -1.488901
#> 28   8.3322811 16.390013
#> 29  -0.2286532 11.668675
#> 30  10.2104546 23.411720
#> 31  -0.9247928 12.923810
#> 32  14.6055343 14.219899
#> 33   8.0903588 18.872490
#> 34   7.8659796 19.787248
#> 35  11.3540152 14.039155
#> 36  14.0296091  9.977807
#> 37   9.7634632 18.411759
#> 38  10.5839389 18.762160
#> 39  11.8556398 20.534441
#> 40   6.8581685  8.134803
#> 41   7.4861817 19.580352
#> 42   6.1421205 13.329981
#> 43  18.4517222  7.955995
#> 44   6.0188309 17.317868
#> 45  18.2861212 17.195661
#> 46  12.9984510  3.412588
#> 47   8.4081198 12.003692
#> 48  19.2036255 20.824294
#> 49  10.7896437 14.210982
#> 50  10.5056723 13.923299
#> 51   8.5714421 10.489731
#> 52   6.9781209 11.432314
#> 53  12.3625277 14.729526
#> 54  11.5832401 27.804328
#> 55  11.5498026 17.722831
#> 56  15.5316761 22.224422
#> 57   3.4373386 14.539016
#> 58   7.9867981 18.172316
#> 59  13.4599305 13.055509
#> 60  10.9405330 12.114604
#> 61  -0.1949835 14.027883
#> 62   1.6695752  7.906043
#> 63   5.2938134 21.159029
#> 64   0.7664494 12.868675
#> 65   9.8382329 12.556003
#> 66  17.7917874 18.798992
#> 67  20.6824835  6.004403
#> 68   4.5930607  7.931406
#> 69  14.6333914 15.552621
#> 70  13.5744943 18.472358
#> 71  14.7955247 15.603963
#> 72  10.9113115  9.893060
#> 73  12.9718418 15.388286
#> 74   3.2622034 10.697569
#> 75  12.4166464 15.698106
#> 76   7.9435470 19.306237
#> 77   3.8064415 12.280630
#> 78  -3.4174992 17.261276
#> 79  19.6420254 13.935052
#> 80  12.1416772 19.856171
#> 81   5.1243474 15.129243
#> 82  14.4592107 16.155270
#> 83  14.4382337 21.636290
#> 84   5.4509838 12.731711
#> 85  16.0861979 19.637993
#> 86  17.1891628  7.306491
#> 87  15.0005903 16.337340
#> 88  13.2382652 14.685484
#> 89   9.1588828 23.682504
#> 90   2.2848105 14.338907
#> 91  11.3185039 14.454009
#> 92   9.7510659 10.853013
#> 93  18.2600208  9.080724
#> 94  12.7874515 13.826092
#> 95  18.1239639 12.534028
#> 96  17.9742560  8.334584
#> 97   6.4722372 20.270199
#> 98  -2.2690390 22.797751
#> 99   7.9284494 17.496032
#> 100 12.2192415 11.750491
#> 
#>