Add custom views to the current view composition

Add one or more custom views to the current view composition.

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.708136 17.338892
#> 2    8.063452 16.935237
#> 3   11.526050 12.463397
#> 4    8.118746 15.468370
#> 5    8.697268 16.510748
#> 6    9.903531 12.507144
#> 7    7.508341 10.499780
#> 8   12.014483 13.394182
#> 9    6.881952 13.760937
#> 10   7.676754 11.459994
#> 11   4.107654 17.620411
#> 12  10.599197 17.197705
#> 13  10.390003 16.832037
#> 14  11.636882 18.923598
#> 15   9.890546 15.447949
#> 16   9.412611 12.102145
#> 17   6.708660 16.643182
#> 18   7.414161 14.191177
#> 19   7.761435 11.486742
#> 20  13.005823 10.213666
#> 21   4.714042 17.008542
#> 22  12.971489 16.725485
#> 23  12.567386 14.892043
#> 24   9.287064 14.121517
#> 25  12.187767 19.808875
#> 26  11.336973 17.227807
#> 27  11.682824 12.336190
#> 28  10.200598 16.060850
#> 29   9.648570 17.871150
#> 30  10.133797 13.047927
#> 31   9.391570 13.488612
#> 32  11.821952 13.938674
#> 33   8.029999 15.081510
#> 34   6.836897 17.533481
#> 35   6.485547 17.230936
#> 36  14.114226 11.031424
#> 37   9.357184 11.925674
#> 38  10.896379 18.616170
#> 39   9.724946 15.665223
#> 40  10.004341 15.425168
#> 41   7.633070 16.904912
#> 42  11.624551 11.142866
#> 43  12.897686 12.339255
#> 44  11.312416 19.677489
#> 45   6.920959 15.062921
#> 46  10.954999 14.262024
#> 47   9.784877 17.336778
#> 48   6.720485 10.451302
#> 49  12.574081 14.671935
#> 50  16.767328 15.722675
#> 51  13.580231 19.353845
#> 52  10.822746 17.189646
#> 53   9.649511 16.366196
#> 54  10.179572 14.304580
#> 55   8.502072 15.050008
#> 56   9.791009 17.782440
#> 57  10.071835 13.694397
#> 58  10.422272 15.891948
#> 59  12.093779 15.293535
#> 60  13.485515 14.666463
#> 61  13.346433 14.874927
#> 62  12.890830 12.689273
#> 63   8.589204 17.596739
#> 64  11.291155 14.470902
#> 65  12.876345 11.984263
#> 66  11.302933 17.922970
#> 67  10.691038 11.136571
#> 68  11.438800  4.897267
#> 69   8.391994 15.463690
#> 70  10.529460 11.093294
#> 71  12.068344 12.106271
#> 72  10.349628 22.950400
#> 73   7.796353 16.254614
#> 74   9.372471 18.944458
#> 75   6.688858  8.865601
#> 76   7.297304 18.297327
#> 77   9.986277 17.739690
#> 78   6.468258  9.853976
#> 79   8.191803 18.991306
#> 80  12.092579  9.949009
#> 81   9.153742 12.479139
#> 82  10.610007 10.864665
#> 83   9.294966 12.983331
#> 84  12.817284  9.028809
#> 85   9.831501 12.490954
#> 86   7.730421 20.342751
#> 87   9.319529 16.289399
#> 88   9.120920 16.103369
#> 89  10.109056 12.649024
#> 90   8.920085 12.764154
#> 91  12.571526 13.646958
#> 92  11.128986 18.545857
#> 93  10.734639 17.216337
#> 94  11.121090 14.147088
#> 95  12.834402 18.598181
#> 96   7.952489 13.102465
#> 97  12.420485 11.495382
#> 98  10.461877 18.510073
#> 99  10.018214 16.286349
#> 100  7.829329 15.732821
#> 
#> 
#> $misty.uniqueid
#> [1] "d03b6bb28df6d812604eb8187b1a411e"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>       marker1    marker2
#> 1    4.743614 18.2127663
#> 2   10.323613 10.3417881
#> 3   11.684054 15.6584911
#> 4   12.153762 25.8435140
#> 5   17.093875  3.6107069
#> 6    4.227177 20.5062241
#> 7    3.570817 15.4657347
#> 8   13.782238 19.6503953
#> 9   15.161480 20.7682897
#> 10   5.549935 12.2177003
#> 11  13.955666 22.2695360
#> 12  14.707730  0.5732081
#> 13  12.928969 15.1521873
#> 14  18.227977 18.5174771
#> 15  21.757741 17.1235993
#> 16  10.264716 11.6276311
#> 17  14.836210 19.5256559
#> 18   7.215275 17.5333340
#> 19   4.765966 18.1902287
#> 20   8.648636 17.2800730
#> 21   9.964157 11.7127957
#> 22   7.983732 19.8166351
#> 23  11.334645  8.4797249
#> 24  10.794732  6.5887309
#> 25  -1.019853 14.3348266
#> 26  14.418201 11.6660481
#> 27  17.270236 16.7550212
#> 28  12.197017 20.1253656
#> 29  13.554810 17.2417847
#> 30   5.662013 14.4883810
#> 31   8.263552  2.6204936
#> 32  12.524518  5.6932906
#> 33  10.389303 18.9667519
#> 34  10.333762 19.3686747
#> 35  13.229151 21.9449065
#> 36   5.354759  5.6812029
#> 37   9.492726 13.7446424
#> 38   6.355689 22.0801972
#> 39  12.454934 11.2629800
#> 40   7.337349 15.3706519
#> 41  10.965888 12.3485470
#> 42  13.718588 14.6933584
#> 43   9.472093 13.3222762
#> 44   6.318437 17.8841948
#> 45  19.722982  5.5201605
#> 46   6.826898 18.1448819
#> 47   6.113687  8.3744892
#> 48  10.499592 12.3679345
#> 49   8.253145 25.3362612
#> 50   5.801361 10.6228350
#> 51  15.441615 11.1490347
#> 52   9.104823 10.6783740
#> 53   3.709376 23.4215005
#> 54   5.479055 13.3396191
#> 55  13.475332 10.9514353
#> 56  14.840395 16.6001184
#> 57  22.794761  8.8597355
#> 58  10.551089 11.9596280
#> 59  12.341175 19.5827725
#> 60   8.719957 17.2181985
#> 61  11.253563  6.2793630
#> 62   9.758069 24.2172115
#> 63   5.661062 17.7023628
#> 64   8.174392 11.3061589
#> 65  14.538064 17.0733642
#> 66  13.125715 19.1344325
#> 67   7.097998 14.6482140
#> 68  12.653318 19.5734569
#> 69   6.646129 18.2747904
#> 70   9.921261 12.8187236
#> 71  12.465388 12.8460265
#> 72  15.014997 21.3020611
#> 73  11.377016 18.7907211
#> 74   8.980861 18.6189239
#> 75   4.352479 16.8005403
#> 76  10.306558 17.4196308
#> 77  13.340929 17.0929812
#> 78   5.885496 12.4329252
#> 79  13.241918 26.6146898
#> 80  22.446838 19.6407021
#> 81  12.401111  9.8247955
#> 82  11.204073 15.6944605
#> 83  12.286967 14.2445881
#> 84  10.306512 15.5125049
#> 85   5.684352  8.4350241
#> 86  13.117308 12.0464586
#> 87  10.469743 21.2135226
#> 88   4.322114 11.7530621
#> 89  13.356335 11.6940553
#> 90  12.892764 11.1671893
#> 91  10.604154 17.5452732
#> 92  15.117405  7.7097965
#> 93  11.899475  9.6318635
#> 94  13.695708 17.9048515
#> 95   6.792726 22.0766588
#> 96  11.358081  9.9960813
#> 97   8.891527 12.9578713
#> 98   8.391330 16.1283254
#> 99   4.218566 -1.2211042
#> 100 15.258792 20.2318924
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.708136 17.338892
#> 2    8.063452 16.935237
#> 3   11.526050 12.463397
#> 4    8.118746 15.468370
#> 5    8.697268 16.510748
#> 6    9.903531 12.507144
#> 7    7.508341 10.499780
#> 8   12.014483 13.394182
#> 9    6.881952 13.760937
#> 10   7.676754 11.459994
#> 11   4.107654 17.620411
#> 12  10.599197 17.197705
#> 13  10.390003 16.832037
#> 14  11.636882 18.923598
#> 15   9.890546 15.447949
#> 16   9.412611 12.102145
#> 17   6.708660 16.643182
#> 18   7.414161 14.191177
#> 19   7.761435 11.486742
#> 20  13.005823 10.213666
#> 21   4.714042 17.008542
#> 22  12.971489 16.725485
#> 23  12.567386 14.892043
#> 24   9.287064 14.121517
#> 25  12.187767 19.808875
#> 26  11.336973 17.227807
#> 27  11.682824 12.336190
#> 28  10.200598 16.060850
#> 29   9.648570 17.871150
#> 30  10.133797 13.047927
#> 31   9.391570 13.488612
#> 32  11.821952 13.938674
#> 33   8.029999 15.081510
#> 34   6.836897 17.533481
#> 35   6.485547 17.230936
#> 36  14.114226 11.031424
#> 37   9.357184 11.925674
#> 38  10.896379 18.616170
#> 39   9.724946 15.665223
#> 40  10.004341 15.425168
#> 41   7.633070 16.904912
#> 42  11.624551 11.142866
#> 43  12.897686 12.339255
#> 44  11.312416 19.677489
#> 45   6.920959 15.062921
#> 46  10.954999 14.262024
#> 47   9.784877 17.336778
#> 48   6.720485 10.451302
#> 49  12.574081 14.671935
#> 50  16.767328 15.722675
#> 51  13.580231 19.353845
#> 52  10.822746 17.189646
#> 53   9.649511 16.366196
#> 54  10.179572 14.304580
#> 55   8.502072 15.050008
#> 56   9.791009 17.782440
#> 57  10.071835 13.694397
#> 58  10.422272 15.891948
#> 59  12.093779 15.293535
#> 60  13.485515 14.666463
#> 61  13.346433 14.874927
#> 62  12.890830 12.689273
#> 63   8.589204 17.596739
#> 64  11.291155 14.470902
#> 65  12.876345 11.984263
#> 66  11.302933 17.922970
#> 67  10.691038 11.136571
#> 68  11.438800  4.897267
#> 69   8.391994 15.463690
#> 70  10.529460 11.093294
#> 71  12.068344 12.106271
#> 72  10.349628 22.950400
#> 73   7.796353 16.254614
#> 74   9.372471 18.944458
#> 75   6.688858  8.865601
#> 76   7.297304 18.297327
#> 77   9.986277 17.739690
#> 78   6.468258  9.853976
#> 79   8.191803 18.991306
#> 80  12.092579  9.949009
#> 81   9.153742 12.479139
#> 82  10.610007 10.864665
#> 83   9.294966 12.983331
#> 84  12.817284  9.028809
#> 85   9.831501 12.490954
#> 86   7.730421 20.342751
#> 87   9.319529 16.289399
#> 88   9.120920 16.103369
#> 89  10.109056 12.649024
#> 90   8.920085 12.764154
#> 91  12.571526 13.646958
#> 92  11.128986 18.545857
#> 93  10.734639 17.216337
#> 94  11.121090 14.147088
#> 95  12.834402 18.598181
#> 96   7.952489 13.102465
#> 97  12.420485 11.495382
#> 98  10.461877 18.510073
#> 99  10.018214 16.286349
#> 100  7.829329 15.732821
#> 
#> 
#> $misty.uniqueid
#> [1] "d03b6bb28df6d812604eb8187b1a411e"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>       marker1    marker2
#> 1    4.743614 18.2127663
#> 2   10.323613 10.3417881
#> 3   11.684054 15.6584911
#> 4   12.153762 25.8435140
#> 5   17.093875  3.6107069
#> 6    4.227177 20.5062241
#> 7    3.570817 15.4657347
#> 8   13.782238 19.6503953
#> 9   15.161480 20.7682897
#> 10   5.549935 12.2177003
#> 11  13.955666 22.2695360
#> 12  14.707730  0.5732081
#> 13  12.928969 15.1521873
#> 14  18.227977 18.5174771
#> 15  21.757741 17.1235993
#> 16  10.264716 11.6276311
#> 17  14.836210 19.5256559
#> 18   7.215275 17.5333340
#> 19   4.765966 18.1902287
#> 20   8.648636 17.2800730
#> 21   9.964157 11.7127957
#> 22   7.983732 19.8166351
#> 23  11.334645  8.4797249
#> 24  10.794732  6.5887309
#> 25  -1.019853 14.3348266
#> 26  14.418201 11.6660481
#> 27  17.270236 16.7550212
#> 28  12.197017 20.1253656
#> 29  13.554810 17.2417847
#> 30   5.662013 14.4883810
#> 31   8.263552  2.6204936
#> 32  12.524518  5.6932906
#> 33  10.389303 18.9667519
#> 34  10.333762 19.3686747
#> 35  13.229151 21.9449065
#> 36   5.354759  5.6812029
#> 37   9.492726 13.7446424
#> 38   6.355689 22.0801972
#> 39  12.454934 11.2629800
#> 40   7.337349 15.3706519
#> 41  10.965888 12.3485470
#> 42  13.718588 14.6933584
#> 43   9.472093 13.3222762
#> 44   6.318437 17.8841948
#> 45  19.722982  5.5201605
#> 46   6.826898 18.1448819
#> 47   6.113687  8.3744892
#> 48  10.499592 12.3679345
#> 49   8.253145 25.3362612
#> 50   5.801361 10.6228350
#> 51  15.441615 11.1490347
#> 52   9.104823 10.6783740
#> 53   3.709376 23.4215005
#> 54   5.479055 13.3396191
#> 55  13.475332 10.9514353
#> 56  14.840395 16.6001184
#> 57  22.794761  8.8597355
#> 58  10.551089 11.9596280
#> 59  12.341175 19.5827725
#> 60   8.719957 17.2181985
#> 61  11.253563  6.2793630
#> 62   9.758069 24.2172115
#> 63   5.661062 17.7023628
#> 64   8.174392 11.3061589
#> 65  14.538064 17.0733642
#> 66  13.125715 19.1344325
#> 67   7.097998 14.6482140
#> 68  12.653318 19.5734569
#> 69   6.646129 18.2747904
#> 70   9.921261 12.8187236
#> 71  12.465388 12.8460265
#> 72  15.014997 21.3020611
#> 73  11.377016 18.7907211
#> 74   8.980861 18.6189239
#> 75   4.352479 16.8005403
#> 76  10.306558 17.4196308
#> 77  13.340929 17.0929812
#> 78   5.885496 12.4329252
#> 79  13.241918 26.6146898
#> 80  22.446838 19.6407021
#> 81  12.401111  9.8247955
#> 82  11.204073 15.6944605
#> 83  12.286967 14.2445881
#> 84  10.306512 15.5125049
#> 85   5.684352  8.4350241
#> 86  13.117308 12.0464586
#> 87  10.469743 21.2135226
#> 88   4.322114 11.7530621
#> 89  13.356335 11.6940553
#> 90  12.892764 11.1671893
#> 91  10.604154 17.5452732
#> 92  15.117405  7.7097965
#> 93  11.899475  9.6318635
#> 94  13.695708 17.9048515
#> 95   6.792726 22.0766588
#> 96  11.358081  9.9960813
#> 97   8.891527 12.9578713
#> 98   8.391330 16.1283254
#> 99   4.218566 -1.2211042
#> 100 15.258792 20.2318924
#> 
#>