Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    6.367578 13.660528
#> 2   11.412268 18.795591
#> 3   10.435780 19.133470
#> 4   11.005681 18.005244
#> 5   12.210940 16.537712
#> 6   10.472112 12.640753
#> 7   10.162135 17.476992
#> 8   10.168064 15.088304
#> 9   10.373959 12.054767
#> 10   8.112558 17.418659
#> 11  10.624522 16.243355
#> 12   6.833295 18.747411
#> 13   9.780231 17.684040
#> 14   7.298886 14.388469
#> 15  13.049443 12.599050
#> 16   8.780371 13.582626
#> 17  12.122010 14.869543
#> 18  11.920200 16.522173
#> 19   6.750819 16.912246
#> 20   9.667504 12.540667
#> 21  13.870126 18.718635
#> 22   8.175921 15.291707
#> 23  13.103743 18.656798
#> 24  12.545750 15.188442
#> 25   5.609032 12.780907
#> 26   9.068654 17.894356
#> 27   8.658243 13.556836
#> 28  12.272979 16.250483
#> 29  11.720532 15.888718
#> 30  10.073815 18.450650
#> 31   8.100606 13.584757
#> 32  10.688784 13.038888
#> 33   8.953193 13.220668
#> 34  10.677577 11.257240
#> 35   9.315523 16.932790
#> 36  10.902844 16.332695
#> 37  11.843764 17.416912
#> 38  16.036860 13.536330
#> 39  10.782535 15.936266
#> 40   7.332702 12.488798
#> 41   5.155952 17.141346
#> 42   5.859038 11.946675
#> 43   7.896515  9.672227
#> 44  14.147666 17.108548
#> 45   9.052875 12.722261
#> 46   7.636260 16.507085
#> 47   9.380554 12.802877
#> 48  10.676258 12.985043
#> 49   7.559516 17.327967
#> 50  10.844291 13.425957
#> 51   8.777332 15.323007
#> 52   7.925160 12.579856
#> 53  10.202817 17.086053
#> 54   6.982833 15.726655
#> 55  11.994597 18.521194
#> 56   8.283546  9.734722
#> 57  10.201641 13.006570
#> 58   7.111470 19.880625
#> 59  10.845926 19.731474
#> 60  10.834874 21.522077
#> 61   9.861411 18.443558
#> 62   9.377515 17.799563
#> 63   8.499223 19.653071
#> 64   7.903452 11.494335
#> 65  11.562188 15.533072
#> 66  12.141993 16.608428
#> 67  13.613405 23.836377
#> 68   8.186192 16.309573
#> 69  10.210280 14.734098
#> 70  12.334557 18.646742
#> 71   9.975590  8.566771
#> 72  10.787299 12.893912
#> 73   9.339438 11.040808
#> 74   8.380733 16.451205
#> 75  12.324905 16.865268
#> 76   7.728790 19.465974
#> 77  14.031566  9.588177
#> 78  12.882147 14.771710
#> 79   8.999719 15.936306
#> 80   9.074393 15.116825
#> 81   7.622625 14.261920
#> 82  10.414867  8.927686
#> 83  13.420480 13.669218
#> 84   8.884409 12.400234
#> 85  11.201724 15.911575
#> 86   9.003996 12.838750
#> 87  11.132423 15.592801
#> 88  10.554731  9.399879
#> 89   8.580039 12.279376
#> 90  11.363142 15.298476
#> 91  13.388828 19.182147
#> 92   9.423201 17.946667
#> 93   6.582051 12.174989
#> 94  10.015492 20.256459
#> 95   9.627381 14.442929
#> 96   9.850368  7.630832
#> 97  10.487508 21.134666
#> 98  11.321766 13.351058
#> 99  11.431116 14.792342
#> 100 10.092062 16.461432
#> 
#> 
#> $misty.uniqueid
#> [1] "438181f117e0c45426f6af1b3902646f"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1    marker2
#> 1   10.0276031 25.9514734
#> 2    0.9313269 13.6143918
#> 3   13.3756911 26.3548004
#> 4   21.8100820 11.4812338
#> 5   14.9942002 17.4641595
#> 6    6.7064621 12.4152965
#> 7   19.4473341 15.8811230
#> 8    4.5226075 12.5160742
#> 9   16.0694111 15.8267619
#> 10  13.2411369 12.2875752
#> 11  14.2761339 16.2451075
#> 12  15.9345975 15.1752917
#> 13   6.4832555  4.7993045
#> 14   6.7916570 21.7301798
#> 15   3.4485818 25.8434744
#> 16   9.7739738 12.2201332
#> 17  13.6257315 13.2764877
#> 18  16.6855846 23.8034532
#> 19   7.3264521 14.3151438
#> 20   4.5180780 16.3583686
#> 21  19.9248228 20.0596122
#> 22   4.1467890  9.2867236
#> 23   4.8180777 16.0879388
#> 24  17.6574257  5.8930940
#> 25   4.5283361 21.6915163
#> 26   4.7174869 17.9366450
#> 27  19.2290640 18.0407807
#> 28  15.1357700 11.2541127
#> 29  19.0901696  6.8221240
#> 30  13.5839676 15.0950738
#> 31   8.6739017 11.5211163
#> 32  11.5204887 14.2169791
#> 33  10.0801586 19.4967219
#> 34   9.6660920 18.3642814
#> 35  10.6058621 10.5759485
#> 36   6.4920616 15.4820154
#> 37   7.3890927 13.3466703
#> 38   9.7532323 17.9033652
#> 39  11.8095492 13.1403744
#> 40   8.2527868 15.2486584
#> 41   4.3663369 18.2620293
#> 42  21.3535664 17.7170956
#> 43  10.7364690  5.2290989
#> 44  13.5543306 10.6149740
#> 45   8.9274911 17.4046390
#> 46   9.7962867 17.5389634
#> 47   9.4949434  9.9876844
#> 48  11.7904224 17.4304717
#> 49  16.6272210  5.2568349
#> 50  11.1606682 19.8203157
#> 51  18.0238674  7.3369667
#> 52   4.9288979 16.2438183
#> 53   4.8714469 15.5120732
#> 54   9.3486218 26.0078894
#> 55  14.0045349 10.1275993
#> 56   4.8686712 12.6200599
#> 57   8.6071535 20.2238050
#> 58   7.4230465  9.8106443
#> 59  11.6161797 17.9410526
#> 60  16.6437175 12.8159166
#> 61   9.4967108 -0.9127759
#> 62   8.3828423 15.3003121
#> 63   8.2357228 17.1301844
#> 64  12.7920605 18.8280140
#> 65   3.5797732 13.9388942
#> 66   9.9066045 17.8974765
#> 67  14.4560332 16.0506725
#> 68  10.0365551 10.4727924
#> 69   6.6375850  8.2726632
#> 70  11.3860856  8.5422279
#> 71  11.8478868  4.8079454
#> 72  14.1877642 17.0908035
#> 73  12.8225435 20.4628902
#> 74  10.8339265 19.1473554
#> 75  17.6622125  7.8915641
#> 76  11.4185123 19.9210220
#> 77  15.0511705 20.5219410
#> 78   3.9692409 11.4475828
#> 79   6.9002127 21.2683230
#> 80   9.5855542 14.1437132
#> 81  10.8626924 21.0565613
#> 82  11.8660831 18.3827909
#> 83   9.4131454 16.6150838
#> 84  14.5641973 18.0671483
#> 85  15.1603126  9.4202704
#> 86   3.8925001 17.5637709
#> 87  18.5901171 16.6894980
#> 88   3.4363120 22.9880982
#> 89  15.0290662  0.6578729
#> 90  12.6333293 12.3179513
#> 91   8.5011572 20.7789679
#> 92   8.4901053 14.4263016
#> 93   3.7541255  9.8365181
#> 94   8.7812622 16.4091303
#> 95   4.3953498  6.6137376
#> 96  12.0604845 11.6364003
#> 97   4.8025846  7.8962363
#> 98  15.3970564 20.9367074
#> 99   4.1142854 18.8613177
#> 100  5.8762499  8.9371214
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    6.367578 13.660528
#> 2   11.412268 18.795591
#> 3   10.435780 19.133470
#> 4   11.005681 18.005244
#> 5   12.210940 16.537712
#> 6   10.472112 12.640753
#> 7   10.162135 17.476992
#> 8   10.168064 15.088304
#> 9   10.373959 12.054767
#> 10   8.112558 17.418659
#> 11  10.624522 16.243355
#> 12   6.833295 18.747411
#> 13   9.780231 17.684040
#> 14   7.298886 14.388469
#> 15  13.049443 12.599050
#> 16   8.780371 13.582626
#> 17  12.122010 14.869543
#> 18  11.920200 16.522173
#> 19   6.750819 16.912246
#> 20   9.667504 12.540667
#> 21  13.870126 18.718635
#> 22   8.175921 15.291707
#> 23  13.103743 18.656798
#> 24  12.545750 15.188442
#> 25   5.609032 12.780907
#> 26   9.068654 17.894356
#> 27   8.658243 13.556836
#> 28  12.272979 16.250483
#> 29  11.720532 15.888718
#> 30  10.073815 18.450650
#> 31   8.100606 13.584757
#> 32  10.688784 13.038888
#> 33   8.953193 13.220668
#> 34  10.677577 11.257240
#> 35   9.315523 16.932790
#> 36  10.902844 16.332695
#> 37  11.843764 17.416912
#> 38  16.036860 13.536330
#> 39  10.782535 15.936266
#> 40   7.332702 12.488798
#> 41   5.155952 17.141346
#> 42   5.859038 11.946675
#> 43   7.896515  9.672227
#> 44  14.147666 17.108548
#> 45   9.052875 12.722261
#> 46   7.636260 16.507085
#> 47   9.380554 12.802877
#> 48  10.676258 12.985043
#> 49   7.559516 17.327967
#> 50  10.844291 13.425957
#> 51   8.777332 15.323007
#> 52   7.925160 12.579856
#> 53  10.202817 17.086053
#> 54   6.982833 15.726655
#> 55  11.994597 18.521194
#> 56   8.283546  9.734722
#> 57  10.201641 13.006570
#> 58   7.111470 19.880625
#> 59  10.845926 19.731474
#> 60  10.834874 21.522077
#> 61   9.861411 18.443558
#> 62   9.377515 17.799563
#> 63   8.499223 19.653071
#> 64   7.903452 11.494335
#> 65  11.562188 15.533072
#> 66  12.141993 16.608428
#> 67  13.613405 23.836377
#> 68   8.186192 16.309573
#> 69  10.210280 14.734098
#> 70  12.334557 18.646742
#> 71   9.975590  8.566771
#> 72  10.787299 12.893912
#> 73   9.339438 11.040808
#> 74   8.380733 16.451205
#> 75  12.324905 16.865268
#> 76   7.728790 19.465974
#> 77  14.031566  9.588177
#> 78  12.882147 14.771710
#> 79   8.999719 15.936306
#> 80   9.074393 15.116825
#> 81   7.622625 14.261920
#> 82  10.414867  8.927686
#> 83  13.420480 13.669218
#> 84   8.884409 12.400234
#> 85  11.201724 15.911575
#> 86   9.003996 12.838750
#> 87  11.132423 15.592801
#> 88  10.554731  9.399879
#> 89   8.580039 12.279376
#> 90  11.363142 15.298476
#> 91  13.388828 19.182147
#> 92   9.423201 17.946667
#> 93   6.582051 12.174989
#> 94  10.015492 20.256459
#> 95   9.627381 14.442929
#> 96   9.850368  7.630832
#> 97  10.487508 21.134666
#> 98  11.321766 13.351058
#> 99  11.431116 14.792342
#> 100 10.092062 16.461432
#> 
#> 
#> $misty.uniqueid
#> [1] "438181f117e0c45426f6af1b3902646f"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1    marker2
#> 1   10.0276031 25.9514734
#> 2    0.9313269 13.6143918
#> 3   13.3756911 26.3548004
#> 4   21.8100820 11.4812338
#> 5   14.9942002 17.4641595
#> 6    6.7064621 12.4152965
#> 7   19.4473341 15.8811230
#> 8    4.5226075 12.5160742
#> 9   16.0694111 15.8267619
#> 10  13.2411369 12.2875752
#> 11  14.2761339 16.2451075
#> 12  15.9345975 15.1752917
#> 13   6.4832555  4.7993045
#> 14   6.7916570 21.7301798
#> 15   3.4485818 25.8434744
#> 16   9.7739738 12.2201332
#> 17  13.6257315 13.2764877
#> 18  16.6855846 23.8034532
#> 19   7.3264521 14.3151438
#> 20   4.5180780 16.3583686
#> 21  19.9248228 20.0596122
#> 22   4.1467890  9.2867236
#> 23   4.8180777 16.0879388
#> 24  17.6574257  5.8930940
#> 25   4.5283361 21.6915163
#> 26   4.7174869 17.9366450
#> 27  19.2290640 18.0407807
#> 28  15.1357700 11.2541127
#> 29  19.0901696  6.8221240
#> 30  13.5839676 15.0950738
#> 31   8.6739017 11.5211163
#> 32  11.5204887 14.2169791
#> 33  10.0801586 19.4967219
#> 34   9.6660920 18.3642814
#> 35  10.6058621 10.5759485
#> 36   6.4920616 15.4820154
#> 37   7.3890927 13.3466703
#> 38   9.7532323 17.9033652
#> 39  11.8095492 13.1403744
#> 40   8.2527868 15.2486584
#> 41   4.3663369 18.2620293
#> 42  21.3535664 17.7170956
#> 43  10.7364690  5.2290989
#> 44  13.5543306 10.6149740
#> 45   8.9274911 17.4046390
#> 46   9.7962867 17.5389634
#> 47   9.4949434  9.9876844
#> 48  11.7904224 17.4304717
#> 49  16.6272210  5.2568349
#> 50  11.1606682 19.8203157
#> 51  18.0238674  7.3369667
#> 52   4.9288979 16.2438183
#> 53   4.8714469 15.5120732
#> 54   9.3486218 26.0078894
#> 55  14.0045349 10.1275993
#> 56   4.8686712 12.6200599
#> 57   8.6071535 20.2238050
#> 58   7.4230465  9.8106443
#> 59  11.6161797 17.9410526
#> 60  16.6437175 12.8159166
#> 61   9.4967108 -0.9127759
#> 62   8.3828423 15.3003121
#> 63   8.2357228 17.1301844
#> 64  12.7920605 18.8280140
#> 65   3.5797732 13.9388942
#> 66   9.9066045 17.8974765
#> 67  14.4560332 16.0506725
#> 68  10.0365551 10.4727924
#> 69   6.6375850  8.2726632
#> 70  11.3860856  8.5422279
#> 71  11.8478868  4.8079454
#> 72  14.1877642 17.0908035
#> 73  12.8225435 20.4628902
#> 74  10.8339265 19.1473554
#> 75  17.6622125  7.8915641
#> 76  11.4185123 19.9210220
#> 77  15.0511705 20.5219410
#> 78   3.9692409 11.4475828
#> 79   6.9002127 21.2683230
#> 80   9.5855542 14.1437132
#> 81  10.8626924 21.0565613
#> 82  11.8660831 18.3827909
#> 83   9.4131454 16.6150838
#> 84  14.5641973 18.0671483
#> 85  15.1603126  9.4202704
#> 86   3.8925001 17.5637709
#> 87  18.5901171 16.6894980
#> 88   3.4363120 22.9880982
#> 89  15.0290662  0.6578729
#> 90  12.6333293 12.3179513
#> 91   8.5011572 20.7789679
#> 92   8.4901053 14.4263016
#> 93   3.7541255  9.8365181
#> 94   8.7812622 16.4091303
#> 95   4.3953498  6.6137376
#> 96  12.0604845 11.6364003
#> 97   4.8025846  7.8962363
#> 98  15.3970564 20.9367074
#> 99   4.1142854 18.8613177
#> 100  5.8762499  8.9371214
#> 
#>