Add custom views to the current view composition

Add one or more custom views to the current view composition.

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.897426 15.628996
#> 2    9.489982 17.515542
#> 3   10.899167 16.924460
#> 4    9.557038 16.636028
#> 5   13.248059 17.470729
#> 6    8.921393 15.623710
#> 7   12.350109 14.968299
#> 8    9.936825 14.080293
#> 9   11.992414 19.467372
#> 10   9.816964 15.671812
#> 11  12.795330 17.279010
#> 12   9.189504 11.573680
#> 13   8.963876 15.199938
#> 14  11.508882 13.638819
#> 15  11.490532 10.282532
#> 16  10.345678 12.948936
#> 17   8.589176 16.854012
#> 18  11.042852 16.525212
#> 19   9.415818 13.877198
#> 20   9.963354  8.744769
#> 21  10.369407 14.562527
#> 22  11.897225 17.007029
#> 23   7.481076 20.578704
#> 24   7.032599 15.818562
#> 25   9.462337 11.045708
#> 26  11.712407 18.220783
#> 27   7.867010 13.290301
#> 28   8.206582 16.401483
#> 29  14.549281 13.310295
#> 30   9.901133 14.221551
#> 31   5.833755 15.612693
#> 32  11.765367 15.621569
#> 33  10.804254 11.719599
#> 34   8.791821 15.165242
#> 35   7.179091  9.399579
#> 36  13.540396 14.072171
#> 37  10.534119 15.037942
#> 38  10.755327 12.174058
#> 39   7.912922 17.755826
#> 40  10.541188 21.145557
#> 41   9.973699 17.347766
#> 42  10.613073 16.666149
#> 43   8.658547 13.366568
#> 44   7.746370 10.613196
#> 45   8.038038 11.781672
#> 46  11.482650 11.813602
#> 47   8.105785 16.784943
#> 48  13.598637 13.568082
#> 49   9.168567 14.787366
#> 50  12.181740 16.416722
#> 51  13.046169 16.380919
#> 52  10.053576 15.988315
#> 53  11.311818 17.406068
#> 54   7.376751 13.821193
#> 55   6.005435 17.810493
#> 56  10.883429 16.943222
#> 57  11.058037  9.685707
#> 58  11.294646 18.494181
#> 59  12.738549 14.961841
#> 60   8.027399 13.321012
#> 61  12.126009 12.501483
#> 62   7.025694 17.337098
#> 63  12.067267  8.299879
#> 64  11.093616 13.721972
#> 65  14.496708 17.318900
#> 66  11.492704 14.383323
#> 67  13.287272 20.417394
#> 68  14.238708 15.148602
#> 69   7.088242 14.662639
#> 70  13.121425 16.350485
#> 71  10.374313 14.830534
#> 72  12.393407 16.841573
#> 73   9.932068 10.421100
#> 74  10.991312 14.592069
#> 75   7.158842 19.527063
#> 76   9.157741 16.254587
#> 77   8.277633 18.059394
#> 78   8.191401 10.330993
#> 79   9.931926 14.382181
#> 80  10.753095 16.922133
#> 81   7.779289 15.385658
#> 82   9.838824 15.722801
#> 83  11.557265 19.872728
#> 84   7.414177 19.616386
#> 85   8.511947 14.209667
#> 86  10.437682 11.591386
#> 87   7.181100 12.746858
#> 88   9.554363 14.302469
#> 89  11.239688 17.108369
#> 90   6.917459 15.264325
#> 91   9.652731 14.327124
#> 92  12.770147 15.222959
#> 93   9.942140 17.001552
#> 94   7.468248 16.803188
#> 95   9.469086 14.848743
#> 96   9.716302 15.430933
#> 97   6.471807 16.452919
#> 98  12.902177 12.392716
#> 99   8.628905 13.435292
#> 100 11.098104 14.530345
#> 
#> 
#> $misty.uniqueid
#> [1] "6609220f82a4c7896ca36900153e51f7"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   12.7495136 14.733188
#> 2    6.1470251 24.585409
#> 3   15.3422971 10.189402
#> 4   14.7228680 17.766546
#> 5    5.8496502 18.071413
#> 6    4.0921205 15.077123
#> 7   13.5691965 17.778094
#> 8    4.0089296 16.983677
#> 9    9.7433295 11.981977
#> 10   7.4186508 30.382770
#> 11   9.7358137 11.768966
#> 12   3.7343514  6.981499
#> 13  11.3188841 13.349026
#> 14   7.5040166 24.590238
#> 15  16.3120162  5.071523
#> 16  10.7557703 21.474422
#> 17  11.8880940 15.087992
#> 18   5.9816752  7.872901
#> 19  12.1870267 13.827005
#> 20   9.0368389 15.667251
#> 21   9.2887387 20.971397
#> 22   9.0494362 14.029106
#> 23  21.5479768 23.615906
#> 24  11.3335726 17.316331
#> 25  13.9564815 18.876667
#> 26   7.6359319 11.930413
#> 27   3.4618611  6.189676
#> 28   9.3530787  7.930837
#> 29   6.1690133 18.634802
#> 30   4.9481698 13.134798
#> 31  11.4124195 19.148644
#> 32   9.2920701 20.158033
#> 33   6.2910258  6.583601
#> 34  13.2077372 16.726482
#> 35   8.0624535 11.975829
#> 36  10.3330969 18.079195
#> 37   6.4689731 14.511987
#> 38   8.4445276 16.232046
#> 39  12.5739945 23.345441
#> 40  12.0370036 11.911750
#> 41  -0.1987883 11.460330
#> 42  10.3959378 22.601777
#> 43  11.2757152  8.136036
#> 44  10.4042232  6.153900
#> 45  14.5319770 17.802197
#> 46   9.2075890 14.250120
#> 47   7.5024619 12.486494
#> 48   3.6446165 15.001908
#> 49  12.1592501  6.940818
#> 50  16.5815569 10.353827
#> 51   9.7010867 19.822078
#> 52   8.9014579 12.343370
#> 53  11.5588876 21.125109
#> 54   8.5758703 21.991644
#> 55  10.0322075 11.865010
#> 56   1.6979135 15.090542
#> 57  11.3866397 17.773748
#> 58  11.1012532  8.968941
#> 59   7.3360651 19.067607
#> 60   8.7682733 19.284356
#> 61  12.1625651 17.048111
#> 62  12.1554292 10.407884
#> 63   9.8547476 16.111071
#> 64  20.2753782 16.126304
#> 65  10.6388090  9.950860
#> 66  10.8933491 18.576438
#> 67   4.1075068 13.290884
#> 68  10.3806029 19.621104
#> 69  16.6179944 20.028408
#> 70  11.7794820  7.246783
#> 71   3.8124400 12.549566
#> 72  14.3857546 21.386215
#> 73   6.3036891  5.346881
#> 74  14.6122335 18.430510
#> 75  18.4309755 14.548124
#> 76  12.7375344 13.918706
#> 77  10.3163085 13.454197
#> 78  15.7431064 15.599880
#> 79  14.2780845 17.536838
#> 80  11.4502588 19.360779
#> 81  16.1882965 18.740838
#> 82  -1.1812163 15.379665
#> 83   1.9605499 14.845974
#> 84  10.3954691 15.393318
#> 85  14.6041751 13.027410
#> 86  -4.0741301  6.632104
#> 87  -0.7385186 17.318094
#> 88   6.4077297  8.822546
#> 89  14.1962798 25.603817
#> 90   5.1936639  8.787417
#> 91  18.4769257  2.409971
#> 92   0.7742162 19.870309
#> 93   8.6757208 10.296635
#> 94   5.7772676 12.396127
#> 95   4.1366867  1.111553
#> 96   6.5643259 12.278892
#> 97   7.7034083 13.055278
#> 98  13.3511964  9.715689
#> 99  13.6877856 14.050930
#> 100 19.0540134 18.570596
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1   10.897426 15.628996
#> 2    9.489982 17.515542
#> 3   10.899167 16.924460
#> 4    9.557038 16.636028
#> 5   13.248059 17.470729
#> 6    8.921393 15.623710
#> 7   12.350109 14.968299
#> 8    9.936825 14.080293
#> 9   11.992414 19.467372
#> 10   9.816964 15.671812
#> 11  12.795330 17.279010
#> 12   9.189504 11.573680
#> 13   8.963876 15.199938
#> 14  11.508882 13.638819
#> 15  11.490532 10.282532
#> 16  10.345678 12.948936
#> 17   8.589176 16.854012
#> 18  11.042852 16.525212
#> 19   9.415818 13.877198
#> 20   9.963354  8.744769
#> 21  10.369407 14.562527
#> 22  11.897225 17.007029
#> 23   7.481076 20.578704
#> 24   7.032599 15.818562
#> 25   9.462337 11.045708
#> 26  11.712407 18.220783
#> 27   7.867010 13.290301
#> 28   8.206582 16.401483
#> 29  14.549281 13.310295
#> 30   9.901133 14.221551
#> 31   5.833755 15.612693
#> 32  11.765367 15.621569
#> 33  10.804254 11.719599
#> 34   8.791821 15.165242
#> 35   7.179091  9.399579
#> 36  13.540396 14.072171
#> 37  10.534119 15.037942
#> 38  10.755327 12.174058
#> 39   7.912922 17.755826
#> 40  10.541188 21.145557
#> 41   9.973699 17.347766
#> 42  10.613073 16.666149
#> 43   8.658547 13.366568
#> 44   7.746370 10.613196
#> 45   8.038038 11.781672
#> 46  11.482650 11.813602
#> 47   8.105785 16.784943
#> 48  13.598637 13.568082
#> 49   9.168567 14.787366
#> 50  12.181740 16.416722
#> 51  13.046169 16.380919
#> 52  10.053576 15.988315
#> 53  11.311818 17.406068
#> 54   7.376751 13.821193
#> 55   6.005435 17.810493
#> 56  10.883429 16.943222
#> 57  11.058037  9.685707
#> 58  11.294646 18.494181
#> 59  12.738549 14.961841
#> 60   8.027399 13.321012
#> 61  12.126009 12.501483
#> 62   7.025694 17.337098
#> 63  12.067267  8.299879
#> 64  11.093616 13.721972
#> 65  14.496708 17.318900
#> 66  11.492704 14.383323
#> 67  13.287272 20.417394
#> 68  14.238708 15.148602
#> 69   7.088242 14.662639
#> 70  13.121425 16.350485
#> 71  10.374313 14.830534
#> 72  12.393407 16.841573
#> 73   9.932068 10.421100
#> 74  10.991312 14.592069
#> 75   7.158842 19.527063
#> 76   9.157741 16.254587
#> 77   8.277633 18.059394
#> 78   8.191401 10.330993
#> 79   9.931926 14.382181
#> 80  10.753095 16.922133
#> 81   7.779289 15.385658
#> 82   9.838824 15.722801
#> 83  11.557265 19.872728
#> 84   7.414177 19.616386
#> 85   8.511947 14.209667
#> 86  10.437682 11.591386
#> 87   7.181100 12.746858
#> 88   9.554363 14.302469
#> 89  11.239688 17.108369
#> 90   6.917459 15.264325
#> 91   9.652731 14.327124
#> 92  12.770147 15.222959
#> 93   9.942140 17.001552
#> 94   7.468248 16.803188
#> 95   9.469086 14.848743
#> 96   9.716302 15.430933
#> 97   6.471807 16.452919
#> 98  12.902177 12.392716
#> 99   8.628905 13.435292
#> 100 11.098104 14.530345
#> 
#> 
#> $misty.uniqueid
#> [1] "6609220f82a4c7896ca36900153e51f7"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   12.7495136 14.733188
#> 2    6.1470251 24.585409
#> 3   15.3422971 10.189402
#> 4   14.7228680 17.766546
#> 5    5.8496502 18.071413
#> 6    4.0921205 15.077123
#> 7   13.5691965 17.778094
#> 8    4.0089296 16.983677
#> 9    9.7433295 11.981977
#> 10   7.4186508 30.382770
#> 11   9.7358137 11.768966
#> 12   3.7343514  6.981499
#> 13  11.3188841 13.349026
#> 14   7.5040166 24.590238
#> 15  16.3120162  5.071523
#> 16  10.7557703 21.474422
#> 17  11.8880940 15.087992
#> 18   5.9816752  7.872901
#> 19  12.1870267 13.827005
#> 20   9.0368389 15.667251
#> 21   9.2887387 20.971397
#> 22   9.0494362 14.029106
#> 23  21.5479768 23.615906
#> 24  11.3335726 17.316331
#> 25  13.9564815 18.876667
#> 26   7.6359319 11.930413
#> 27   3.4618611  6.189676
#> 28   9.3530787  7.930837
#> 29   6.1690133 18.634802
#> 30   4.9481698 13.134798
#> 31  11.4124195 19.148644
#> 32   9.2920701 20.158033
#> 33   6.2910258  6.583601
#> 34  13.2077372 16.726482
#> 35   8.0624535 11.975829
#> 36  10.3330969 18.079195
#> 37   6.4689731 14.511987
#> 38   8.4445276 16.232046
#> 39  12.5739945 23.345441
#> 40  12.0370036 11.911750
#> 41  -0.1987883 11.460330
#> 42  10.3959378 22.601777
#> 43  11.2757152  8.136036
#> 44  10.4042232  6.153900
#> 45  14.5319770 17.802197
#> 46   9.2075890 14.250120
#> 47   7.5024619 12.486494
#> 48   3.6446165 15.001908
#> 49  12.1592501  6.940818
#> 50  16.5815569 10.353827
#> 51   9.7010867 19.822078
#> 52   8.9014579 12.343370
#> 53  11.5588876 21.125109
#> 54   8.5758703 21.991644
#> 55  10.0322075 11.865010
#> 56   1.6979135 15.090542
#> 57  11.3866397 17.773748
#> 58  11.1012532  8.968941
#> 59   7.3360651 19.067607
#> 60   8.7682733 19.284356
#> 61  12.1625651 17.048111
#> 62  12.1554292 10.407884
#> 63   9.8547476 16.111071
#> 64  20.2753782 16.126304
#> 65  10.6388090  9.950860
#> 66  10.8933491 18.576438
#> 67   4.1075068 13.290884
#> 68  10.3806029 19.621104
#> 69  16.6179944 20.028408
#> 70  11.7794820  7.246783
#> 71   3.8124400 12.549566
#> 72  14.3857546 21.386215
#> 73   6.3036891  5.346881
#> 74  14.6122335 18.430510
#> 75  18.4309755 14.548124
#> 76  12.7375344 13.918706
#> 77  10.3163085 13.454197
#> 78  15.7431064 15.599880
#> 79  14.2780845 17.536838
#> 80  11.4502588 19.360779
#> 81  16.1882965 18.740838
#> 82  -1.1812163 15.379665
#> 83   1.9605499 14.845974
#> 84  10.3954691 15.393318
#> 85  14.6041751 13.027410
#> 86  -4.0741301  6.632104
#> 87  -0.7385186 17.318094
#> 88   6.4077297  8.822546
#> 89  14.1962798 25.603817
#> 90   5.1936639  8.787417
#> 91  18.4769257  2.409971
#> 92   0.7742162 19.870309
#> 93   8.6757208 10.296635
#> 94   5.7772676 12.396127
#> 95   4.1366867  1.111553
#> 96   6.5643259 12.278892
#> 97   7.7034083 13.055278
#> 98  13.3511964  9.715689
#> 99  13.6877856 14.050930
#> 100 19.0540134 18.570596
#> 
#>