Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.147116 14.773349
#> 2    9.379281  9.598865
#> 3   12.471372 12.584100
#> 4   10.341245 19.549720
#> 5   10.341238 17.569026
#> 6    8.127741 11.400255
#> 7    9.192327 18.759454
#> 8    7.991191 13.873981
#> 9    7.778586 15.960021
#> 10   8.617160 16.927435
#> 11   8.660713 19.156778
#> 12  11.810934 16.916394
#> 13  10.728475 14.944677
#> 14  11.303032 15.738555
#> 15  10.552506 16.953432
#> 16  10.954474 19.140693
#> 17  11.819402 13.999430
#> 18   8.573062 14.266932
#> 19   7.039639 17.932174
#> 20   8.257285 12.727924
#> 21  11.196253 14.491926
#> 22  10.262893 18.100687
#> 23  10.045391 11.693906
#> 24  12.371541 10.066523
#> 25   8.200097 21.228335
#> 26   9.392253 13.364929
#> 27  10.540035 18.172582
#> 28  10.582997 14.606449
#> 29  13.805375 17.673744
#> 30   6.944830 14.031214
#> 31  10.842000 17.959361
#> 32   8.974109 17.611423
#> 33  11.600573 12.285631
#> 34  11.247682 10.313818
#> 35  10.986234 10.557505
#> 36   4.604489 14.430758
#> 37  12.172167 11.708435
#> 38  10.420431 14.581907
#> 39   7.199396 15.635287
#> 40   7.875248 13.008051
#> 41  10.686682 12.388456
#> 42   8.477913 17.149223
#> 43   9.488551 12.825695
#> 44  12.622561 10.512821
#> 45   8.100540 18.588889
#> 46  10.547171 18.138914
#> 47  11.443992 13.241828
#> 48   7.837032 15.194163
#> 49   7.285705 16.535340
#> 50   9.628270 13.898895
#> 51  13.929758 15.040375
#> 52  11.064995 17.485847
#> 53  10.896790 17.435841
#> 54   8.485408 12.406793
#> 55  11.204250 15.638445
#> 56   7.341554 18.712210
#> 57   6.823032 11.758395
#> 58  12.448538 11.880617
#> 59  11.888459 12.325656
#> 60  11.373085 12.018831
#> 61   8.824511 16.028936
#> 62   7.760347 14.133224
#> 63   7.208120 13.682043
#> 64   8.369269 12.463294
#> 65   9.343171 16.927003
#> 66   8.893487 16.289842
#> 67  10.629718  8.095231
#> 68   9.338371 14.082113
#> 69   7.498090 10.388512
#> 70   8.282610 11.033060
#> 71  10.194826 22.555975
#> 72   8.646241 16.938945
#> 73  12.968828 19.994221
#> 74  10.050579 18.379473
#> 75   9.973016 12.593697
#> 76   9.863692 10.840088
#> 77   9.603544 12.990747
#> 78  11.581431 20.790503
#> 79   8.881165 13.218355
#> 80   8.403656 13.568303
#> 81  12.106995 12.604467
#> 82   8.791174 15.720998
#> 83  12.108503 13.613823
#> 84  14.555612 15.189932
#> 85   6.758847 11.161352
#> 86   7.615607  8.356509
#> 87  12.168551 15.595379
#> 88   7.901084 22.752864
#> 89  14.032439  9.763586
#> 90  11.732212 17.608203
#> 91   7.072087 19.445792
#> 92   9.271814 16.848258
#> 93   5.752941 15.013441
#> 94   7.520027 16.103703
#> 95   9.160238 16.943678
#> 96   9.054491 14.319986
#> 97  12.941847 13.301676
#> 98   6.144297 19.512165
#> 99  11.975934 16.186339
#> 100 12.227423 18.228549
#> 
#> 
#> $misty.uniqueid
#> [1] "790c4cc3e0be70477421eb2a0a5057a3"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   15.4325216 12.196657
#> 2   17.2725073 10.542491
#> 3    7.0312589 15.273506
#> 4    9.5205542 15.331941
#> 5    0.5857943 19.178378
#> 6    9.1958625 10.520786
#> 7    7.4191778 14.827394
#> 8    6.7665015 11.394387
#> 9    9.5681059 10.949289
#> 10  12.8940065 11.200911
#> 11   7.1138365 14.348012
#> 12   2.3424527 17.928963
#> 13  10.3914779 12.307287
#> 14  10.2741325  6.330625
#> 15  14.8573177 15.229007
#> 16  11.1066147 18.878023
#> 17  17.3306749 11.867334
#> 18  11.6958422 11.855958
#> 19  10.9388989  1.630758
#> 20  13.8997664 11.214635
#> 21   7.5815462 19.324120
#> 22   8.5659007 22.617628
#> 23  11.5347740 20.283152
#> 24   4.4165057  5.782458
#> 25   7.2185265 13.089618
#> 26  12.7465816 20.861671
#> 27  14.1983716  9.170548
#> 28  16.9065572 10.843393
#> 29  14.4480255 16.864736
#> 30   3.7115773 17.113913
#> 31   9.0951101 13.865950
#> 32  10.4888856 14.416103
#> 33  12.9629553 10.422704
#> 34   5.4403867 20.305971
#> 35  10.9280881 11.626687
#> 36  20.0996672  5.360340
#> 37  11.5513039 18.529238
#> 38  12.6520048 12.383338
#> 39  14.6736306 11.565209
#> 40   7.6896529 22.110151
#> 41   9.8179885  8.691901
#> 42   3.7538343 17.885095
#> 43   9.8418147 13.916670
#> 44   8.8671723  2.851097
#> 45  11.7033453  7.273878
#> 46  11.1601635  8.429435
#> 47  11.2243017 22.433918
#> 48   3.6467930  6.880682
#> 49   7.3339703 13.325778
#> 50  20.0427331  5.800592
#> 51  -1.6797850 25.409744
#> 52  13.3770979 14.496853
#> 53   6.6826816 10.772240
#> 54   4.5630740 22.942833
#> 55  17.9328461 19.529736
#> 56  11.1551001 18.061527
#> 57  14.4967838 19.530814
#> 58  23.7038631 14.594430
#> 59   1.7225141 21.087970
#> 60  11.6996767 10.602354
#> 61   8.6116118 21.113845
#> 62  16.5405817 18.276203
#> 63   0.3709174  9.700346
#> 64  10.3711960 10.546526
#> 65   7.3770224 10.430074
#> 66   9.0548265 17.787692
#> 67  12.3101996 11.702569
#> 68  14.4590148 18.111699
#> 69   8.1058905 14.617330
#> 70   9.8003026 23.831542
#> 71  10.1974744 15.363684
#> 72  10.8338533  6.533359
#> 73   9.9157588 10.290579
#> 74  10.1466736 13.604972
#> 75   4.7594172 12.817860
#> 76   8.3190180 22.393860
#> 77  -0.3923340 14.402588
#> 78   9.1455180  8.228548
#> 79   6.5677559 15.082416
#> 80   8.5926602 19.976364
#> 81  11.8569538 13.436973
#> 82   6.0568812 15.543345
#> 83  11.8152126  8.676890
#> 84  11.5619152 13.105979
#> 85   9.2731839 21.402112
#> 86  16.0865758 15.034537
#> 87  15.8171716 14.581668
#> 88  14.8384004 16.811186
#> 89  14.2539181 15.853230
#> 90   7.3003148 13.589548
#> 91   9.9732440 19.856860
#> 92   7.8443344 20.984054
#> 93   6.2126264  4.428040
#> 94   4.0078642 13.763909
#> 95  10.0587509 20.934616
#> 96  18.9805758 12.784631
#> 97   7.5453897 17.766827
#> 98   8.2656999 17.422929
#> 99  13.0533179 10.387294
#> 100  2.5225715 21.149169
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.147116 14.773349
#> 2    9.379281  9.598865
#> 3   12.471372 12.584100
#> 4   10.341245 19.549720
#> 5   10.341238 17.569026
#> 6    8.127741 11.400255
#> 7    9.192327 18.759454
#> 8    7.991191 13.873981
#> 9    7.778586 15.960021
#> 10   8.617160 16.927435
#> 11   8.660713 19.156778
#> 12  11.810934 16.916394
#> 13  10.728475 14.944677
#> 14  11.303032 15.738555
#> 15  10.552506 16.953432
#> 16  10.954474 19.140693
#> 17  11.819402 13.999430
#> 18   8.573062 14.266932
#> 19   7.039639 17.932174
#> 20   8.257285 12.727924
#> 21  11.196253 14.491926
#> 22  10.262893 18.100687
#> 23  10.045391 11.693906
#> 24  12.371541 10.066523
#> 25   8.200097 21.228335
#> 26   9.392253 13.364929
#> 27  10.540035 18.172582
#> 28  10.582997 14.606449
#> 29  13.805375 17.673744
#> 30   6.944830 14.031214
#> 31  10.842000 17.959361
#> 32   8.974109 17.611423
#> 33  11.600573 12.285631
#> 34  11.247682 10.313818
#> 35  10.986234 10.557505
#> 36   4.604489 14.430758
#> 37  12.172167 11.708435
#> 38  10.420431 14.581907
#> 39   7.199396 15.635287
#> 40   7.875248 13.008051
#> 41  10.686682 12.388456
#> 42   8.477913 17.149223
#> 43   9.488551 12.825695
#> 44  12.622561 10.512821
#> 45   8.100540 18.588889
#> 46  10.547171 18.138914
#> 47  11.443992 13.241828
#> 48   7.837032 15.194163
#> 49   7.285705 16.535340
#> 50   9.628270 13.898895
#> 51  13.929758 15.040375
#> 52  11.064995 17.485847
#> 53  10.896790 17.435841
#> 54   8.485408 12.406793
#> 55  11.204250 15.638445
#> 56   7.341554 18.712210
#> 57   6.823032 11.758395
#> 58  12.448538 11.880617
#> 59  11.888459 12.325656
#> 60  11.373085 12.018831
#> 61   8.824511 16.028936
#> 62   7.760347 14.133224
#> 63   7.208120 13.682043
#> 64   8.369269 12.463294
#> 65   9.343171 16.927003
#> 66   8.893487 16.289842
#> 67  10.629718  8.095231
#> 68   9.338371 14.082113
#> 69   7.498090 10.388512
#> 70   8.282610 11.033060
#> 71  10.194826 22.555975
#> 72   8.646241 16.938945
#> 73  12.968828 19.994221
#> 74  10.050579 18.379473
#> 75   9.973016 12.593697
#> 76   9.863692 10.840088
#> 77   9.603544 12.990747
#> 78  11.581431 20.790503
#> 79   8.881165 13.218355
#> 80   8.403656 13.568303
#> 81  12.106995 12.604467
#> 82   8.791174 15.720998
#> 83  12.108503 13.613823
#> 84  14.555612 15.189932
#> 85   6.758847 11.161352
#> 86   7.615607  8.356509
#> 87  12.168551 15.595379
#> 88   7.901084 22.752864
#> 89  14.032439  9.763586
#> 90  11.732212 17.608203
#> 91   7.072087 19.445792
#> 92   9.271814 16.848258
#> 93   5.752941 15.013441
#> 94   7.520027 16.103703
#> 95   9.160238 16.943678
#> 96   9.054491 14.319986
#> 97  12.941847 13.301676
#> 98   6.144297 19.512165
#> 99  11.975934 16.186339
#> 100 12.227423 18.228549
#> 
#> 
#> $misty.uniqueid
#> [1] "790c4cc3e0be70477421eb2a0a5057a3"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   15.4325216 12.196657
#> 2   17.2725073 10.542491
#> 3    7.0312589 15.273506
#> 4    9.5205542 15.331941
#> 5    0.5857943 19.178378
#> 6    9.1958625 10.520786
#> 7    7.4191778 14.827394
#> 8    6.7665015 11.394387
#> 9    9.5681059 10.949289
#> 10  12.8940065 11.200911
#> 11   7.1138365 14.348012
#> 12   2.3424527 17.928963
#> 13  10.3914779 12.307287
#> 14  10.2741325  6.330625
#> 15  14.8573177 15.229007
#> 16  11.1066147 18.878023
#> 17  17.3306749 11.867334
#> 18  11.6958422 11.855958
#> 19  10.9388989  1.630758
#> 20  13.8997664 11.214635
#> 21   7.5815462 19.324120
#> 22   8.5659007 22.617628
#> 23  11.5347740 20.283152
#> 24   4.4165057  5.782458
#> 25   7.2185265 13.089618
#> 26  12.7465816 20.861671
#> 27  14.1983716  9.170548
#> 28  16.9065572 10.843393
#> 29  14.4480255 16.864736
#> 30   3.7115773 17.113913
#> 31   9.0951101 13.865950
#> 32  10.4888856 14.416103
#> 33  12.9629553 10.422704
#> 34   5.4403867 20.305971
#> 35  10.9280881 11.626687
#> 36  20.0996672  5.360340
#> 37  11.5513039 18.529238
#> 38  12.6520048 12.383338
#> 39  14.6736306 11.565209
#> 40   7.6896529 22.110151
#> 41   9.8179885  8.691901
#> 42   3.7538343 17.885095
#> 43   9.8418147 13.916670
#> 44   8.8671723  2.851097
#> 45  11.7033453  7.273878
#> 46  11.1601635  8.429435
#> 47  11.2243017 22.433918
#> 48   3.6467930  6.880682
#> 49   7.3339703 13.325778
#> 50  20.0427331  5.800592
#> 51  -1.6797850 25.409744
#> 52  13.3770979 14.496853
#> 53   6.6826816 10.772240
#> 54   4.5630740 22.942833
#> 55  17.9328461 19.529736
#> 56  11.1551001 18.061527
#> 57  14.4967838 19.530814
#> 58  23.7038631 14.594430
#> 59   1.7225141 21.087970
#> 60  11.6996767 10.602354
#> 61   8.6116118 21.113845
#> 62  16.5405817 18.276203
#> 63   0.3709174  9.700346
#> 64  10.3711960 10.546526
#> 65   7.3770224 10.430074
#> 66   9.0548265 17.787692
#> 67  12.3101996 11.702569
#> 68  14.4590148 18.111699
#> 69   8.1058905 14.617330
#> 70   9.8003026 23.831542
#> 71  10.1974744 15.363684
#> 72  10.8338533  6.533359
#> 73   9.9157588 10.290579
#> 74  10.1466736 13.604972
#> 75   4.7594172 12.817860
#> 76   8.3190180 22.393860
#> 77  -0.3923340 14.402588
#> 78   9.1455180  8.228548
#> 79   6.5677559 15.082416
#> 80   8.5926602 19.976364
#> 81  11.8569538 13.436973
#> 82   6.0568812 15.543345
#> 83  11.8152126  8.676890
#> 84  11.5619152 13.105979
#> 85   9.2731839 21.402112
#> 86  16.0865758 15.034537
#> 87  15.8171716 14.581668
#> 88  14.8384004 16.811186
#> 89  14.2539181 15.853230
#> 90   7.3003148 13.589548
#> 91   9.9732440 19.856860
#> 92   7.8443344 20.984054
#> 93   6.2126264  4.428040
#> 94   4.0078642 13.763909
#> 95  10.0587509 20.934616
#> 96  18.9805758 12.784631
#> 97   7.5453897 17.766827
#> 98   8.2656999 17.422929
#> 99  13.0533179 10.387294
#> 100  2.5225715 21.149169
#> 
#>