Add custom views to the current view composition

Add one or more custom views to the current view composition.

add_views(current.views, new.views)

Arguments

current.views
the current view composition.

new.views
a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)

#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.199913 13.838359
#> 2   10.510634 12.643702
#> 3    5.125473 11.829789
#> 4    9.988857 12.613376
#> 5   11.243105  9.731174
#> 6   12.296823 12.928386
#> 7    6.356365 13.324374
#> 8    9.505349 13.390010
#> 9    9.511601 15.681381
#> 10   9.434589 17.935365
#> 11   8.892601 14.373352
#> 12  11.257964 10.801769
#> 13  14.130050 15.775612
#> 14   6.738021 13.674602
#> 15  11.024854 16.705800
#> 16   6.273977 21.380551
#> 17   8.955975 16.274575
#> 18   9.894796  9.947155
#> 19  11.085993 15.748205
#> 20   8.171850 18.218515
#> 21  10.936309 21.118108
#> 22  10.725903 16.348361
#> 23   7.390913 19.175442
#> 24  11.475553 16.279700
#> 25  13.777010 15.322752
#> 26   9.805110 15.066884
#> 27   8.128305 16.810833
#> 28   9.968099 14.212048
#> 29   8.346422 13.415208
#> 30   6.975201 15.576448
#> 31  11.870726 11.561401
#> 32  10.352977 17.538554
#> 33  10.487371 15.245159
#> 34  13.247098 11.084649
#> 35  10.224076 12.165264
#> 36   9.732006 16.363025
#> 37   6.179825 12.434392
#> 38   9.441526 14.139314
#> 39   9.373108 17.684885
#> 40  12.134616 15.201913
#> 41  10.140070 14.511971
#> 42   8.721753 12.518069
#> 43   9.900070 20.629517
#> 44   9.497033 17.299321
#> 45  10.889594 17.939870
#> 46  15.510835 18.965343
#> 47  10.093063 11.640868
#> 48  11.155418 16.543799
#> 49  10.236390 10.472700
#> 50   6.176559 19.598224
#> 51  11.724173 16.287442
#> 52   9.513527 15.366310
#> 53   9.587826 11.585963
#> 54  10.038355 13.325955
#> 55  10.059122 18.157616
#> 56  11.099655 17.033051
#> 57   5.451770 15.115499
#> 58  15.365114 13.930856
#> 59   9.277557 17.348532
#> 60  10.426711 17.413235
#> 61  12.148692  9.299818
#> 62   8.669824 17.807353
#> 63  12.227905 14.072845
#> 64   9.508207 15.789200
#> 65   7.644873  9.628224
#> 66   8.048299 12.635223
#> 67  12.130115 11.600935
#> 68  10.263341 16.090958
#> 69  10.977258 14.142336
#> 70   6.601099 16.553007
#> 71   7.058527 14.691274
#> 72  10.568301 12.077791
#> 73  12.674641 18.812017
#> 74  10.473393 17.882594
#> 75  12.636587 17.306164
#> 76  11.047820 18.107792
#> 77  11.213496 13.578339
#> 78   9.780129 11.173995
#> 79  10.344363 14.083138
#> 80   9.819345 21.635308
#> 81  13.848687 11.874995
#> 82  12.596786 11.560428
#> 83  11.497583  9.974018
#> 84  11.112449 19.577816
#> 85   8.903485 16.662557
#> 86  12.221070 20.979331
#> 87   4.775331 14.537638
#> 88   9.688612 22.693225
#> 89  10.867780 18.185997
#> 90   9.236098 18.428085
#> 91  10.848375 18.371517
#> 92  12.126204 13.808996
#> 93  12.097425 12.530217
#> 94   9.923794 13.263346
#> 95  10.972298 20.291368
#> 96  13.345765 15.398976
#> 97   9.291278 16.129498
#> 98  11.892696 18.416123
#> 99  12.633653 18.723789
#> 100  9.406720 16.836273
#> 
#> 
#> $misty.uniqueid
#> [1] "b0782b57cbeec2bcbbdf3e54abecd6fe"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    7.8530996 13.219378
#> 2   16.8023066  9.677679
#> 3    9.6457128 20.385583
#> 4    8.6392316 20.907878
#> 5   -2.2334001 15.991960
#> 6   10.3274332 12.997974
#> 7    4.5074555 18.080771
#> 8    6.8341091 24.870784
#> 9   -0.3182723 24.423312
#> 10  23.2446601  7.056897
#> 11   4.2330081 12.300384
#> 12   8.2968106  9.152693
#> 13  13.9318129 17.795530
#> 14   3.6474344  5.903264
#> 15  12.7107077 16.966720
#> 16  10.3755295 15.210671
#> 17  12.7925721 20.898321
#> 18  12.0770320 13.715394
#> 19   2.7385012  9.718320
#> 20  14.7060306 15.993886
#> 21   8.3053206 18.252668
#> 22   9.6221288 16.719567
#> 23  10.2010220 22.387662
#> 24  10.6215053 15.360128
#> 25   5.0078372 25.632223
#> 26  16.1669503  7.619015
#> 27  11.7021224 17.039442
#> 28   7.6364876 21.969889
#> 29  13.5437653 16.801391
#> 30   2.3552064 18.272751
#> 31  11.1871267 20.260777
#> 32   3.4359288  5.102224
#> 33  13.7351429 21.041928
#> 34   2.1874078 14.153600
#> 35  10.3552668 16.475149
#> 36   6.8023261 21.331703
#> 37   5.7740213  9.323284
#> 38  13.3762235  9.344731
#> 39  15.7668790 15.549967
#> 40   1.5674763 19.264527
#> 41   5.4859253 13.828311
#> 42  16.5881685 25.433443
#> 43  15.5009487 14.445403
#> 44  16.0188392  8.035765
#> 45   2.8436461  9.288546
#> 46  16.9145543 23.523044
#> 47  10.0156297 14.599632
#> 48   9.6105659 12.813594
#> 49  12.2071411 14.403925
#> 50  10.6446145 18.932314
#> 51   5.8489287 12.105274
#> 52   7.4820355 14.272866
#> 53   4.0317941 17.632290
#> 54   6.2413834 23.667891
#> 55  17.2792070 22.243286
#> 56   5.8569823 22.590966
#> 57  11.4488723 13.079964
#> 58   7.5997326 24.135626
#> 59   6.9758532 12.242541
#> 60  17.3005509 10.671232
#> 61  10.7483968 13.280843
#> 62   2.8333945 20.314382
#> 63   9.9484834 19.065291
#> 64   8.9388198 24.017417
#> 65   5.4682991 14.474657
#> 66  -0.5107624 19.912267
#> 67  19.4668023  6.433487
#> 68   5.1593708 10.839902
#> 69   9.4869848 20.502459
#> 70  11.1997979 14.130899
#> 71  10.3044945 15.894060
#> 72  -0.8878801 11.507853
#> 73   9.4106993 10.197754
#> 74  10.5614739 10.122885
#> 75  10.0394310 13.307117
#> 76  19.3887194 20.761735
#> 77  20.7937828 17.025506
#> 78  13.5485726 12.645388
#> 79  13.8349169 14.333745
#> 80   8.4589429 21.133412
#> 81  15.0600092 16.664720
#> 82   5.4047420 13.264558
#> 83  12.8169004 14.507247
#> 84  11.6124137 15.173830
#> 85  11.8333718 16.930635
#> 86  15.6491758 15.104156
#> 87   5.2925096 15.037934
#> 88  11.0891882 19.654220
#> 89  17.0770615 11.576250
#> 90   8.0813348 16.687008
#> 91   9.1295681 12.939311
#> 92   8.8912774 19.671306
#> 93   4.9523564 24.201584
#> 94  12.4036263 11.475902
#> 95  18.0220366 15.042552
#> 96   2.4248774 25.170949
#> 97   2.9198804  8.291570
#> 98  14.3838866 20.794896
#> 99  13.1206621 13.983955
#> 100 20.5613864 13.109857
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))

#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.199913 13.838359
#> 2   10.510634 12.643702
#> 3    5.125473 11.829789
#> 4    9.988857 12.613376
#> 5   11.243105  9.731174
#> 6   12.296823 12.928386
#> 7    6.356365 13.324374
#> 8    9.505349 13.390010
#> 9    9.511601 15.681381
#> 10   9.434589 17.935365
#> 11   8.892601 14.373352
#> 12  11.257964 10.801769
#> 13  14.130050 15.775612
#> 14   6.738021 13.674602
#> 15  11.024854 16.705800
#> 16   6.273977 21.380551
#> 17   8.955975 16.274575
#> 18   9.894796  9.947155
#> 19  11.085993 15.748205
#> 20   8.171850 18.218515
#> 21  10.936309 21.118108
#> 22  10.725903 16.348361
#> 23   7.390913 19.175442
#> 24  11.475553 16.279700
#> 25  13.777010 15.322752
#> 26   9.805110 15.066884
#> 27   8.128305 16.810833
#> 28   9.968099 14.212048
#> 29   8.346422 13.415208
#> 30   6.975201 15.576448
#> 31  11.870726 11.561401
#> 32  10.352977 17.538554
#> 33  10.487371 15.245159
#> 34  13.247098 11.084649
#> 35  10.224076 12.165264
#> 36   9.732006 16.363025
#> 37   6.179825 12.434392
#> 38   9.441526 14.139314
#> 39   9.373108 17.684885
#> 40  12.134616 15.201913
#> 41  10.140070 14.511971
#> 42   8.721753 12.518069
#> 43   9.900070 20.629517
#> 44   9.497033 17.299321
#> 45  10.889594 17.939870
#> 46  15.510835 18.965343
#> 47  10.093063 11.640868
#> 48  11.155418 16.543799
#> 49  10.236390 10.472700
#> 50   6.176559 19.598224
#> 51  11.724173 16.287442
#> 52   9.513527 15.366310
#> 53   9.587826 11.585963
#> 54  10.038355 13.325955
#> 55  10.059122 18.157616
#> 56  11.099655 17.033051
#> 57   5.451770 15.115499
#> 58  15.365114 13.930856
#> 59   9.277557 17.348532
#> 60  10.426711 17.413235
#> 61  12.148692  9.299818
#> 62   8.669824 17.807353
#> 63  12.227905 14.072845
#> 64   9.508207 15.789200
#> 65   7.644873  9.628224
#> 66   8.048299 12.635223
#> 67  12.130115 11.600935
#> 68  10.263341 16.090958
#> 69  10.977258 14.142336
#> 70   6.601099 16.553007
#> 71   7.058527 14.691274
#> 72  10.568301 12.077791
#> 73  12.674641 18.812017
#> 74  10.473393 17.882594
#> 75  12.636587 17.306164
#> 76  11.047820 18.107792
#> 77  11.213496 13.578339
#> 78   9.780129 11.173995
#> 79  10.344363 14.083138
#> 80   9.819345 21.635308
#> 81  13.848687 11.874995
#> 82  12.596786 11.560428
#> 83  11.497583  9.974018
#> 84  11.112449 19.577816
#> 85   8.903485 16.662557
#> 86  12.221070 20.979331
#> 87   4.775331 14.537638
#> 88   9.688612 22.693225
#> 89  10.867780 18.185997
#> 90   9.236098 18.428085
#> 91  10.848375 18.371517
#> 92  12.126204 13.808996
#> 93  12.097425 12.530217
#> 94   9.923794 13.263346
#> 95  10.972298 20.291368
#> 96  13.345765 15.398976
#> 97   9.291278 16.129498
#> 98  11.892696 18.416123
#> 99  12.633653 18.723789
#> 100  9.406720 16.836273
#> 
#> 
#> $misty.uniqueid
#> [1] "b0782b57cbeec2bcbbdf3e54abecd6fe"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1    7.8530996 13.219378
#> 2   16.8023066  9.677679
#> 3    9.6457128 20.385583
#> 4    8.6392316 20.907878
#> 5   -2.2334001 15.991960
#> 6   10.3274332 12.997974
#> 7    4.5074555 18.080771
#> 8    6.8341091 24.870784
#> 9   -0.3182723 24.423312
#> 10  23.2446601  7.056897
#> 11   4.2330081 12.300384
#> 12   8.2968106  9.152693
#> 13  13.9318129 17.795530
#> 14   3.6474344  5.903264
#> 15  12.7107077 16.966720
#> 16  10.3755295 15.210671
#> 17  12.7925721 20.898321
#> 18  12.0770320 13.715394
#> 19   2.7385012  9.718320
#> 20  14.7060306 15.993886
#> 21   8.3053206 18.252668
#> 22   9.6221288 16.719567
#> 23  10.2010220 22.387662
#> 24  10.6215053 15.360128
#> 25   5.0078372 25.632223
#> 26  16.1669503  7.619015
#> 27  11.7021224 17.039442
#> 28   7.6364876 21.969889
#> 29  13.5437653 16.801391
#> 30   2.3552064 18.272751
#> 31  11.1871267 20.260777
#> 32   3.4359288  5.102224
#> 33  13.7351429 21.041928
#> 34   2.1874078 14.153600
#> 35  10.3552668 16.475149
#> 36   6.8023261 21.331703
#> 37   5.7740213  9.323284
#> 38  13.3762235  9.344731
#> 39  15.7668790 15.549967
#> 40   1.5674763 19.264527
#> 41   5.4859253 13.828311
#> 42  16.5881685 25.433443
#> 43  15.5009487 14.445403
#> 44  16.0188392  8.035765
#> 45   2.8436461  9.288546
#> 46  16.9145543 23.523044
#> 47  10.0156297 14.599632
#> 48   9.6105659 12.813594
#> 49  12.2071411 14.403925
#> 50  10.6446145 18.932314
#> 51   5.8489287 12.105274
#> 52   7.4820355 14.272866
#> 53   4.0317941 17.632290
#> 54   6.2413834 23.667891
#> 55  17.2792070 22.243286
#> 56   5.8569823 22.590966
#> 57  11.4488723 13.079964
#> 58   7.5997326 24.135626
#> 59   6.9758532 12.242541
#> 60  17.3005509 10.671232
#> 61  10.7483968 13.280843
#> 62   2.8333945 20.314382
#> 63   9.9484834 19.065291
#> 64   8.9388198 24.017417
#> 65   5.4682991 14.474657
#> 66  -0.5107624 19.912267
#> 67  19.4668023  6.433487
#> 68   5.1593708 10.839902
#> 69   9.4869848 20.502459
#> 70  11.1997979 14.130899
#> 71  10.3044945 15.894060
#> 72  -0.8878801 11.507853
#> 73   9.4106993 10.197754
#> 74  10.5614739 10.122885
#> 75  10.0394310 13.307117
#> 76  19.3887194 20.761735
#> 77  20.7937828 17.025506
#> 78  13.5485726 12.645388
#> 79  13.8349169 14.333745
#> 80   8.4589429 21.133412
#> 81  15.0600092 16.664720
#> 82   5.4047420 13.264558
#> 83  12.8169004 14.507247
#> 84  11.6124137 15.173830
#> 85  11.8333718 16.930635
#> 86  15.6491758 15.104156
#> 87   5.2925096 15.037934
#> 88  11.0891882 19.654220
#> 89  17.0770615 11.576250
#> 90   8.0813348 16.687008
#> 91   9.1295681 12.939311
#> 92   8.8912774 19.671306
#> 93   4.9523564 24.201584
#> 94  12.4036263 11.475902
#> 95  18.0220366 15.042552
#> 96   2.4248774 25.170949
#> 97   2.9198804  8.291570
#> 98  14.3838866 20.794896
#> 99  13.1206621 13.983955
#> 100 20.5613864 13.109857
#> 
#>