Add custom views to the current view composition

Add one or more custom views to the current view composition.

add_views(current.views, new.views)

Arguments

current.views
the current view composition.

new.views
a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from current.views and new.views.

See also

Examples

# Add custom views created with create_view() to the view composition.

# Alternatives
if (FALSE) {

new.view <- create_view("nearest", nnexpr, "")
add_views(misty.views, new.view)
  
misty.views %>% add_views(create_view("nearest", nnexpr, ""))

misty.views %>% 
  add_views(c(create_view("nearest", nnexpr, ""), view2, view3))
  
misty.views %>% add_views(list(new.view, another.view))
}