Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage,

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.406231 17.893163
#> 2    8.438305 12.592391
#> 3    6.951036 11.845643
#> 4   12.366460 16.310448
#> 5   11.831156 18.494045
#> 6    7.749688 13.288896
#> 7   10.358862 16.176849
#> 8    5.348827  9.995543
#> 9    8.684906 20.431140
#> 10  13.838740 18.065080
#> 11   8.689653 15.393096
#> 12   8.531492 16.776026
#> 13   8.193998 20.183738
#> 14  10.033295 12.555771
#> 15  13.503739 12.966003
#> 16   7.325875 14.070319
#> 17   9.632226 19.171403
#> 18  10.899270 16.090845
#> 19   6.973525 18.661530
#> 20  10.827554 19.248763
#> 21  10.287072 12.475214
#> 22   6.354200 13.410965
#> 23   9.808126 15.821262
#> 24   8.938590 14.258552
#> 25   6.816345 15.745692
#> 26   8.973325 13.887669
#> 27  10.312562 15.912912
#> 28   9.892835 14.765853
#> 29   9.784100 18.207639
#> 30  11.711990 19.200811
#> 31  11.186372 14.326022
#> 32   9.006762 11.522628
#> 33  11.119647 17.559537
#> 34  11.366636 16.311351
#> 35   9.875982 15.118568
#> 36  10.447546 15.422524
#> 37   9.882318 13.309907
#> 38   7.365584 11.551938
#> 39  13.255510 11.738024
#> 40   7.445469 12.375216
#> 41   6.123031 12.731385
#> 42   8.047282 15.885151
#> 43   8.335945 20.361493
#> 44  10.008909 13.892441
#> 45   9.401846 12.778791
#> 46  11.532427 14.229447
#> 47  14.838202 17.934660
#> 48  11.026816 11.706907
#> 49  11.683822 11.566211
#> 50   6.747461 20.135810
#> 51   9.937813 14.740987
#> 52  11.591970 11.466428
#> 53  11.222415  8.309773
#> 54   8.291038  9.844486
#> 55   9.036828 13.832067
#> 56  11.645221 18.384156
#> 57  13.158134 11.613525
#> 58   8.002984 13.699928
#> 59   9.976432 13.189589
#> 60   9.616478 14.080558
#> 61  11.430415 17.170690
#> 62   9.998650 16.080797
#> 63   6.677694 13.744525
#> 64   6.508618 13.028598
#> 65  11.525321 16.114534
#> 66   8.082871 14.723286
#> 67  12.519740 15.922084
#> 68   9.640083 15.173144
#> 69   9.530751 10.220240
#> 70   9.462595 19.129188
#> 71   8.004879 14.278656
#> 72  14.064105 13.955327
#> 73  10.965953 18.482156
#> 74  10.724531 16.258308
#> 75  11.307759 17.188076
#> 76   7.343868 13.991729
#> 77  10.548968 12.730862
#> 78  13.322497 16.496589
#> 79   7.025799 12.042636
#> 80  10.623708 11.440584
#> 81   5.632134 11.032893
#> 82  10.976730 17.597235
#> 83  12.593007 15.992528
#> 84  11.192613  9.871099
#> 85   7.686005 13.527529
#> 86   7.004779 19.337338
#> 87  12.563095 14.253698
#> 88   7.694299  9.893288
#> 89   9.524235 10.212616
#> 90  10.979292 10.031433
#> 91   7.078768 16.733615
#> 92   8.642047 16.137746
#> 93   8.027155 10.453641
#> 94   9.302254 15.372839
#> 95  12.211606 15.136641
#> 96   7.531607 16.408254
#> 97  10.689105 17.865946
#> 98  10.530326 14.565584
#> 99  11.238571 18.910249
#> 100  8.426724 16.892504
#> 
#> 
#> $misty.uniqueid
#> [1] "fe3063b7f1052126b102958f2107e467"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   11.0586536 16.541178
#> 2    4.6981946 17.510625
#> 3    7.3004992 20.677006
#> 4   17.6156984  9.797385
#> 5    8.0236135 16.078173
#> 6   -0.3771677 20.713012
#> 7   10.8745040 19.083178
#> 8   16.0106647 16.746344
#> 9    8.8549539 11.247250
#> 10  13.4283134 15.012881
#> 11  15.5847155 12.674218
#> 12  10.1864216 17.492638
#> 13   7.1031097 11.677833
#> 14   9.0692662 10.457786
#> 15   9.1454036 18.516708
#> 16  11.6110357 23.106567
#> 17   7.0318199 20.508739
#> 18  15.6788592 24.355329
#> 19  10.0118107 13.213233
#> 20  13.2156306  4.600867
#> 21  15.7116873 16.671202
#> 22   5.5270966 21.434528
#> 23  12.8159230 18.907552
#> 24  15.8162256 15.013999
#> 25   5.2155070 17.894398
#> 26   7.3903473 17.014396
#> 27  11.1052729 13.084178
#> 28  11.5913160 17.382083
#> 29  12.3621620 11.541415
#> 30  14.8500235 12.451947
#> 31   6.7021180 15.499338
#> 32   9.3620163 13.199737
#> 33   7.2656118 22.777451
#> 34   6.0297113 17.014184
#> 35   6.8953556 11.332554
#> 36   3.2515845 13.549248
#> 37  10.9086109  7.321351
#> 38   7.1728649 11.683247
#> 39  13.9138909  9.901596
#> 40  16.2952787 10.457751
#> 41  12.3021074 19.917380
#> 42  17.1157700 17.840026
#> 43   7.7552356  8.388634
#> 44   5.8463905 20.975091
#> 45   6.1371385 18.302954
#> 46   9.6245140 13.383586
#> 47  14.8956665 18.724111
#> 48  11.2407965 16.360043
#> 49  13.2142500  9.946872
#> 50  12.2626063 11.814471
#> 51   7.2821971 18.859458
#> 52   8.7367045  8.421039
#> 53  14.5625982 17.828160
#> 54   4.1538574 18.498972
#> 55   5.1294793 17.159111
#> 56  11.8182077 11.886845
#> 57  11.6801718 18.257370
#> 58  10.8657782 15.669771
#> 59   4.7758303 10.593123
#> 60  13.4040548 11.406963
#> 61   8.0666102 15.295180
#> 62  18.1250595 11.314787
#> 63   7.5482878  9.042611
#> 64  11.0034892 17.729586
#> 65  10.0370817 13.917206
#> 66   7.5817486 12.625453
#> 67   7.8274762 22.681863
#> 68  11.7420072 26.105234
#> 69   9.9283187 22.368008
#> 70  13.7227589  4.424425
#> 71   8.6895097 17.001890
#> 72   5.8333453 13.202740
#> 73  10.9050930  9.029316
#> 74  12.3335323  4.073588
#> 75  12.2256646 10.516095
#> 76  11.2789403 21.283544
#> 77   5.0702831 17.025937
#> 78   3.5393555 18.123422
#> 79   8.4915075 16.669542
#> 80  10.5449904 13.124116
#> 81   9.0218102 13.441281
#> 82   5.5209443 20.802694
#> 83  10.3030627 25.017049
#> 84   4.6846835 15.635426
#> 85   7.4505550 20.179646
#> 86  13.4595233 20.269693
#> 87   8.5198241 18.281370
#> 88   3.7277686 17.558842
#> 89  17.6075043 11.522630
#> 90  13.9038286 20.342307
#> 91  21.0464362 12.262069
#> 92   5.9222365 22.175034
#> 93   7.4983089 18.624484
#> 94  16.1795024 15.522881
#> 95  13.9716890 11.781957
#> 96   6.3495255  7.989329
#> 97   7.5652037  6.657879
#> 98  12.1863070 11.022903
#> 99   6.4064754 13.545668
#> 100 18.9628941 12.956597
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.406231 17.893163
#> 2    8.438305 12.592391
#> 3    6.951036 11.845643
#> 4   12.366460 16.310448
#> 5   11.831156 18.494045
#> 6    7.749688 13.288896
#> 7   10.358862 16.176849
#> 8    5.348827  9.995543
#> 9    8.684906 20.431140
#> 10  13.838740 18.065080
#> 11   8.689653 15.393096
#> 12   8.531492 16.776026
#> 13   8.193998 20.183738
#> 14  10.033295 12.555771
#> 15  13.503739 12.966003
#> 16   7.325875 14.070319
#> 17   9.632226 19.171403
#> 18  10.899270 16.090845
#> 19   6.973525 18.661530
#> 20  10.827554 19.248763
#> 21  10.287072 12.475214
#> 22   6.354200 13.410965
#> 23   9.808126 15.821262
#> 24   8.938590 14.258552
#> 25   6.816345 15.745692
#> 26   8.973325 13.887669
#> 27  10.312562 15.912912
#> 28   9.892835 14.765853
#> 29   9.784100 18.207639
#> 30  11.711990 19.200811
#> 31  11.186372 14.326022
#> 32   9.006762 11.522628
#> 33  11.119647 17.559537
#> 34  11.366636 16.311351
#> 35   9.875982 15.118568
#> 36  10.447546 15.422524
#> 37   9.882318 13.309907
#> 38   7.365584 11.551938
#> 39  13.255510 11.738024
#> 40   7.445469 12.375216
#> 41   6.123031 12.731385
#> 42   8.047282 15.885151
#> 43   8.335945 20.361493
#> 44  10.008909 13.892441
#> 45   9.401846 12.778791
#> 46  11.532427 14.229447
#> 47  14.838202 17.934660
#> 48  11.026816 11.706907
#> 49  11.683822 11.566211
#> 50   6.747461 20.135810
#> 51   9.937813 14.740987
#> 52  11.591970 11.466428
#> 53  11.222415  8.309773
#> 54   8.291038  9.844486
#> 55   9.036828 13.832067
#> 56  11.645221 18.384156
#> 57  13.158134 11.613525
#> 58   8.002984 13.699928
#> 59   9.976432 13.189589
#> 60   9.616478 14.080558
#> 61  11.430415 17.170690
#> 62   9.998650 16.080797
#> 63   6.677694 13.744525
#> 64   6.508618 13.028598
#> 65  11.525321 16.114534
#> 66   8.082871 14.723286
#> 67  12.519740 15.922084
#> 68   9.640083 15.173144
#> 69   9.530751 10.220240
#> 70   9.462595 19.129188
#> 71   8.004879 14.278656
#> 72  14.064105 13.955327
#> 73  10.965953 18.482156
#> 74  10.724531 16.258308
#> 75  11.307759 17.188076
#> 76   7.343868 13.991729
#> 77  10.548968 12.730862
#> 78  13.322497 16.496589
#> 79   7.025799 12.042636
#> 80  10.623708 11.440584
#> 81   5.632134 11.032893
#> 82  10.976730 17.597235
#> 83  12.593007 15.992528
#> 84  11.192613  9.871099
#> 85   7.686005 13.527529
#> 86   7.004779 19.337338
#> 87  12.563095 14.253698
#> 88   7.694299  9.893288
#> 89   9.524235 10.212616
#> 90  10.979292 10.031433
#> 91   7.078768 16.733615
#> 92   8.642047 16.137746
#> 93   8.027155 10.453641
#> 94   9.302254 15.372839
#> 95  12.211606 15.136641
#> 96   7.531607 16.408254
#> 97  10.689105 17.865946
#> 98  10.530326 14.565584
#> 99  11.238571 18.910249
#> 100  8.426724 16.892504
#> 
#> 
#> $misty.uniqueid
#> [1] "fe3063b7f1052126b102958f2107e467"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>        marker1   marker2
#> 1   11.0586536 16.541178
#> 2    4.6981946 17.510625
#> 3    7.3004992 20.677006
#> 4   17.6156984  9.797385
#> 5    8.0236135 16.078173
#> 6   -0.3771677 20.713012
#> 7   10.8745040 19.083178
#> 8   16.0106647 16.746344
#> 9    8.8549539 11.247250
#> 10  13.4283134 15.012881
#> 11  15.5847155 12.674218
#> 12  10.1864216 17.492638
#> 13   7.1031097 11.677833
#> 14   9.0692662 10.457786
#> 15   9.1454036 18.516708
#> 16  11.6110357 23.106567
#> 17   7.0318199 20.508739
#> 18  15.6788592 24.355329
#> 19  10.0118107 13.213233
#> 20  13.2156306  4.600867
#> 21  15.7116873 16.671202
#> 22   5.5270966 21.434528
#> 23  12.8159230 18.907552
#> 24  15.8162256 15.013999
#> 25   5.2155070 17.894398
#> 26   7.3903473 17.014396
#> 27  11.1052729 13.084178
#> 28  11.5913160 17.382083
#> 29  12.3621620 11.541415
#> 30  14.8500235 12.451947
#> 31   6.7021180 15.499338
#> 32   9.3620163 13.199737
#> 33   7.2656118 22.777451
#> 34   6.0297113 17.014184
#> 35   6.8953556 11.332554
#> 36   3.2515845 13.549248
#> 37  10.9086109  7.321351
#> 38   7.1728649 11.683247
#> 39  13.9138909  9.901596
#> 40  16.2952787 10.457751
#> 41  12.3021074 19.917380
#> 42  17.1157700 17.840026
#> 43   7.7552356  8.388634
#> 44   5.8463905 20.975091
#> 45   6.1371385 18.302954
#> 46   9.6245140 13.383586
#> 47  14.8956665 18.724111
#> 48  11.2407965 16.360043
#> 49  13.2142500  9.946872
#> 50  12.2626063 11.814471
#> 51   7.2821971 18.859458
#> 52   8.7367045  8.421039
#> 53  14.5625982 17.828160
#> 54   4.1538574 18.498972
#> 55   5.1294793 17.159111
#> 56  11.8182077 11.886845
#> 57  11.6801718 18.257370
#> 58  10.8657782 15.669771
#> 59   4.7758303 10.593123
#> 60  13.4040548 11.406963
#> 61   8.0666102 15.295180
#> 62  18.1250595 11.314787
#> 63   7.5482878  9.042611
#> 64  11.0034892 17.729586
#> 65  10.0370817 13.917206
#> 66   7.5817486 12.625453
#> 67   7.8274762 22.681863
#> 68  11.7420072 26.105234
#> 69   9.9283187 22.368008
#> 70  13.7227589  4.424425
#> 71   8.6895097 17.001890
#> 72   5.8333453 13.202740
#> 73  10.9050930  9.029316
#> 74  12.3335323  4.073588
#> 75  12.2256646 10.516095
#> 76  11.2789403 21.283544
#> 77   5.0702831 17.025937
#> 78   3.5393555 18.123422
#> 79   8.4915075 16.669542
#> 80  10.5449904 13.124116
#> 81   9.0218102 13.441281
#> 82   5.5209443 20.802694
#> 83  10.3030627 25.017049
#> 84   4.6846835 15.635426
#> 85   7.4505550 20.179646
#> 86  13.4595233 20.269693
#> 87   8.5198241 18.281370
#> 88   3.7277686 17.558842
#> 89  17.6075043 11.522630
#> 90  13.9038286 20.342307
#> 91  21.0464362 12.262069
#> 92   5.9222365 22.175034
#> 93   7.4983089 18.624484
#> 94  16.1795024 15.522881
#> 95  13.9716890 11.781957
#> 96   6.3495255  7.989329
#> 97   7.5652037  6.657879
#> 98  12.1863070 11.022903
#> 99   6.4064754 13.545668
#> 100 18.9628941 12.956597
#> 
#>