Add custom views to the current view composition

Add one or more custom views to the current view composition.

Usage

add_views(current.views, new.views)

Arguments

current.views

the current view composition.

new.views

a view or a list of views created with create_view() or otherwise.

Value

A mistyR view composition containing an union of views from

current.views and new.views.

See also

Examples

# create random views
view1 <- data.frame(marker1 = rnorm(100, 10, 2), marker2 = rnorm(100, 15, 3))
view2 <- data.frame(marker1 = rnorm(100, 10, 5), marker2 = rnorm(100, 15, 5))

misty.views <- create_initial_view(view1)

new.view <- create_view("dummyname", view2, "dname")
add_views(misty.views, new.view)
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.377431 11.575408
#> 2    7.794241 14.652013
#> 3    9.546742 10.868123
#> 4    8.886201 15.432423
#> 5    9.703178 16.089305
#> 6    9.990599 18.453890
#> 7    7.775015 11.158741
#> 8   12.150078 13.472632
#> 9   11.740351 12.996268
#> 10  12.005619 17.413666
#> 11  12.149108 12.382710
#> 12   9.633888 14.082961
#> 13  13.026786 15.362537
#> 14  10.766854 15.330265
#> 15  10.935294 20.143866
#> 16   9.197073 14.853910
#> 17  10.069325 17.013377
#> 18   9.458682 14.266146
#> 19  11.474345 20.289679
#> 20  11.212417 17.292345
#> 21  11.560610  9.640527
#> 22  10.510212 15.959987
#> 23   8.475769 16.499671
#> 24   7.153394 12.462511
#> 25  12.737456 11.820217
#> 26  11.070473 17.769697
#> 27  11.153027 15.300581
#> 28   8.995077 10.841359
#> 29  11.583158 17.164516
#> 30   8.370704 18.314271
#> 31  10.445481 17.114638
#> 32  14.195911 14.271379
#> 33  10.878244 18.564387
#> 34  10.920205 13.784765
#> 35   8.233868 15.029371
#> 36  11.441335 18.664030
#> 37  13.239513 15.220468
#> 38   9.044124 15.291023
#> 39  11.229146 17.806718
#> 40   7.941866 18.786631
#> 41  11.401803 16.226421
#> 42  10.873482 20.081910
#> 43   9.837498 11.106255
#> 44  12.526039 12.461710
#> 45   8.100004 13.612464
#> 46   8.229611 15.276898
#> 47  10.541663 15.483472
#> 48  10.528633 11.069462
#> 49   8.708126 11.667607
#> 50  10.584584 12.282389
#> 51  11.874188 14.219660
#> 52   5.719785 15.984736
#> 53   8.037629 14.645443
#> 54   6.823450  9.022715
#> 55  10.611356 12.302336
#> 56   7.268333 16.794219
#> 57   7.460874 17.244827
#> 58   9.934416 18.126710
#> 59   9.177738 14.396921
#> 60  11.993567 10.431339
#> 61  10.153774 18.184009
#> 62  14.265554 14.904995
#> 63   8.424027 11.258837
#> 64  10.692384  8.035713
#> 65   7.367035  9.378813
#> 66   5.886320  9.931551
#> 67  11.978127 17.497162
#> 68   9.963001  9.396366
#> 69   7.693132  9.803208
#> 70  10.888084 15.081354
#> 71  11.902841 14.753715
#> 72  12.909799 10.536414
#> 73   9.819952 16.425054
#> 74   7.575065 19.666278
#> 75   8.461661 15.012537
#> 76   8.960771 17.255828
#> 77   9.873571 16.876376
#> 78  11.740948 14.313035
#> 79   8.179438 17.940507
#> 80   9.849521  8.729360
#> 81  12.725149 10.217743
#> 82   7.909449 12.913824
#> 83  10.320317 13.257962
#> 84  10.825999 11.593971
#> 85   8.294297 10.664774
#> 86   8.904920 13.984680
#> 87   6.591544 16.102480
#> 88  10.744175 13.630184
#> 89  11.437526 10.802483
#> 90   8.537684 11.437093
#> 91   9.328113 20.848193
#> 92   9.010348 13.010099
#> 93  12.587192 18.541201
#> 94  11.171533 17.912316
#> 95   8.327446 13.017048
#> 96   9.317464 18.896677
#> 97  10.332561 13.468874
#> 98   8.097576 13.921577
#> 99  11.829073 16.098622
#> 100 10.240606 16.188407
#> 
#> 
#> $misty.uniqueid
#> [1] "ff8790973c082cd5c176669811ad7cf9"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>          marker1   marker2
#> 1    3.899948846 11.204549
#> 2    8.108206366 12.824588
#> 3   16.267751605 14.530842
#> 4   12.830925636 24.077974
#> 5   22.150189297 13.671899
#> 6   17.122740925 16.564214
#> 7   -1.135273647 15.429825
#> 8   11.267052988  7.345819
#> 9    8.407993462 17.114810
#> 10  11.933801955 15.205851
#> 11  10.046157036 18.516203
#> 12   9.683645528 15.980142
#> 13   6.033461208 17.730000
#> 14   9.883864173 17.822348
#> 15   8.983905233 14.336830
#> 16  14.121775983 13.728820
#> 17  12.658571949  5.382853
#> 18   5.754032825 19.161900
#> 19   8.700046612 16.656035
#> 20  10.396867249  9.983365
#> 21   4.281798763  7.869408
#> 22   6.092945772  9.426415
#> 23  11.218747534 19.842888
#> 24   3.181538563 14.465889
#> 25   7.682413707 16.200672
#> 26  22.825741610 20.582602
#> 27   4.947409831 10.666081
#> 28  11.525135645 14.404332
#> 29   1.250027759 10.325661
#> 30  11.107539586 20.536214
#> 31  -2.350304054 15.516766
#> 32  20.568539705 17.333337
#> 33   1.958145254 15.331443
#> 34  11.813698803 18.070282
#> 35   6.150727641 10.224089
#> 36  16.347917933 17.762127
#> 37   6.645264917  6.310316
#> 38  10.035850305 10.236884
#> 39   2.054760715 14.597452
#> 40  19.665109349 11.312523
#> 41  12.350145113 16.652420
#> 42  11.621810687 14.780172
#> 43  10.281087415 11.930451
#> 44   7.606352400 16.698199
#> 45  10.917449101 14.795561
#> 46   9.912793421 14.554843
#> 47   3.366048056 11.372404
#> 48  13.952710713 19.722347
#> 49  11.075014480 27.114925
#> 50   4.801270493 13.143013
#> 51   8.961001620 14.264755
#> 52  15.526601673 16.951737
#> 53  16.229186543 13.059256
#> 54  11.058683347  4.295285
#> 55  17.879666536 23.743616
#> 56   7.783035981 16.501373
#> 57   8.459839628 16.649993
#> 58   9.743338833 18.862881
#> 59  -0.058657313  8.781659
#> 60   0.001366583 10.150031
#> 61   8.262481360 15.457308
#> 62   3.013660702 15.718722
#> 63  13.818210116  9.239281
#> 64  21.783148846  8.918705
#> 65  16.006502210 11.030574
#> 66  11.484014242 22.735284
#> 67  13.896434924 16.439864
#> 68  -1.997289890 10.930957
#> 69   9.924536370 31.320804
#> 70  11.502026935 14.557477
#> 71   1.748731138 21.593312
#> 72   3.692685416 14.814392
#> 73  11.492944577 15.844503
#> 74  14.732761805 24.049896
#> 75  12.382811436  9.417268
#> 76  12.510294363 12.837609
#> 77  13.498749986 19.696957
#> 78  16.946808311 20.116535
#> 79   5.339037311 11.917995
#> 80   7.237811365 13.526597
#> 81   3.800628128 18.066560
#> 82   8.481558744 16.659054
#> 83  18.094343796 20.844745
#> 84  12.466553862 16.718795
#> 85  10.468252474 25.026423
#> 86  15.218057805 19.790155
#> 87   4.250962061 19.894834
#> 88  12.165451008 12.835083
#> 89  12.599593790 17.128601
#> 90   5.422638700 24.432647
#> 91  14.210803248 19.626388
#> 92   9.676001729 12.339478
#> 93   9.810941808 12.840664
#> 94  10.097077025 14.937497
#> 95   5.156349989  6.658048
#> 96  15.139955187 17.039449
#> 97  17.099705733 24.761180
#> 98   2.517018364 16.168941
#> 99  13.031220792 15.241051
#> 100 11.533647767 10.149358
#> 
#> 

misty.views %>% add_views(create_view("dummyname", view2, "dname"))
#> $intraview
#> $intraview$abbrev
#> [1] "intra"
#> 
#> $intraview$data
#>       marker1   marker2
#> 1    7.377431 11.575408
#> 2    7.794241 14.652013
#> 3    9.546742 10.868123
#> 4    8.886201 15.432423
#> 5    9.703178 16.089305
#> 6    9.990599 18.453890
#> 7    7.775015 11.158741
#> 8   12.150078 13.472632
#> 9   11.740351 12.996268
#> 10  12.005619 17.413666
#> 11  12.149108 12.382710
#> 12   9.633888 14.082961
#> 13  13.026786 15.362537
#> 14  10.766854 15.330265
#> 15  10.935294 20.143866
#> 16   9.197073 14.853910
#> 17  10.069325 17.013377
#> 18   9.458682 14.266146
#> 19  11.474345 20.289679
#> 20  11.212417 17.292345
#> 21  11.560610  9.640527
#> 22  10.510212 15.959987
#> 23   8.475769 16.499671
#> 24   7.153394 12.462511
#> 25  12.737456 11.820217
#> 26  11.070473 17.769697
#> 27  11.153027 15.300581
#> 28   8.995077 10.841359
#> 29  11.583158 17.164516
#> 30   8.370704 18.314271
#> 31  10.445481 17.114638
#> 32  14.195911 14.271379
#> 33  10.878244 18.564387
#> 34  10.920205 13.784765
#> 35   8.233868 15.029371
#> 36  11.441335 18.664030
#> 37  13.239513 15.220468
#> 38   9.044124 15.291023
#> 39  11.229146 17.806718
#> 40   7.941866 18.786631
#> 41  11.401803 16.226421
#> 42  10.873482 20.081910
#> 43   9.837498 11.106255
#> 44  12.526039 12.461710
#> 45   8.100004 13.612464
#> 46   8.229611 15.276898
#> 47  10.541663 15.483472
#> 48  10.528633 11.069462
#> 49   8.708126 11.667607
#> 50  10.584584 12.282389
#> 51  11.874188 14.219660
#> 52   5.719785 15.984736
#> 53   8.037629 14.645443
#> 54   6.823450  9.022715
#> 55  10.611356 12.302336
#> 56   7.268333 16.794219
#> 57   7.460874 17.244827
#> 58   9.934416 18.126710
#> 59   9.177738 14.396921
#> 60  11.993567 10.431339
#> 61  10.153774 18.184009
#> 62  14.265554 14.904995
#> 63   8.424027 11.258837
#> 64  10.692384  8.035713
#> 65   7.367035  9.378813
#> 66   5.886320  9.931551
#> 67  11.978127 17.497162
#> 68   9.963001  9.396366
#> 69   7.693132  9.803208
#> 70  10.888084 15.081354
#> 71  11.902841 14.753715
#> 72  12.909799 10.536414
#> 73   9.819952 16.425054
#> 74   7.575065 19.666278
#> 75   8.461661 15.012537
#> 76   8.960771 17.255828
#> 77   9.873571 16.876376
#> 78  11.740948 14.313035
#> 79   8.179438 17.940507
#> 80   9.849521  8.729360
#> 81  12.725149 10.217743
#> 82   7.909449 12.913824
#> 83  10.320317 13.257962
#> 84  10.825999 11.593971
#> 85   8.294297 10.664774
#> 86   8.904920 13.984680
#> 87   6.591544 16.102480
#> 88  10.744175 13.630184
#> 89  11.437526 10.802483
#> 90   8.537684 11.437093
#> 91   9.328113 20.848193
#> 92   9.010348 13.010099
#> 93  12.587192 18.541201
#> 94  11.171533 17.912316
#> 95   8.327446 13.017048
#> 96   9.317464 18.896677
#> 97  10.332561 13.468874
#> 98   8.097576 13.921577
#> 99  11.829073 16.098622
#> 100 10.240606 16.188407
#> 
#> 
#> $misty.uniqueid
#> [1] "ff8790973c082cd5c176669811ad7cf9"
#> 
#> $dummyname
#> $dummyname$abbrev
#> [1] "dname"
#> 
#> $dummyname$data
#>          marker1   marker2
#> 1    3.899948846 11.204549
#> 2    8.108206366 12.824588
#> 3   16.267751605 14.530842
#> 4   12.830925636 24.077974
#> 5   22.150189297 13.671899
#> 6   17.122740925 16.564214
#> 7   -1.135273647 15.429825
#> 8   11.267052988  7.345819
#> 9    8.407993462 17.114810
#> 10  11.933801955 15.205851
#> 11  10.046157036 18.516203
#> 12   9.683645528 15.980142
#> 13   6.033461208 17.730000
#> 14   9.883864173 17.822348
#> 15   8.983905233 14.336830
#> 16  14.121775983 13.728820
#> 17  12.658571949  5.382853
#> 18   5.754032825 19.161900
#> 19   8.700046612 16.656035
#> 20  10.396867249  9.983365
#> 21   4.281798763  7.869408
#> 22   6.092945772  9.426415
#> 23  11.218747534 19.842888
#> 24   3.181538563 14.465889
#> 25   7.682413707 16.200672
#> 26  22.825741610 20.582602
#> 27   4.947409831 10.666081
#> 28  11.525135645 14.404332
#> 29   1.250027759 10.325661
#> 30  11.107539586 20.536214
#> 31  -2.350304054 15.516766
#> 32  20.568539705 17.333337
#> 33   1.958145254 15.331443
#> 34  11.813698803 18.070282
#> 35   6.150727641 10.224089
#> 36  16.347917933 17.762127
#> 37   6.645264917  6.310316
#> 38  10.035850305 10.236884
#> 39   2.054760715 14.597452
#> 40  19.665109349 11.312523
#> 41  12.350145113 16.652420
#> 42  11.621810687 14.780172
#> 43  10.281087415 11.930451
#> 44   7.606352400 16.698199
#> 45  10.917449101 14.795561
#> 46   9.912793421 14.554843
#> 47   3.366048056 11.372404
#> 48  13.952710713 19.722347
#> 49  11.075014480 27.114925
#> 50   4.801270493 13.143013
#> 51   8.961001620 14.264755
#> 52  15.526601673 16.951737
#> 53  16.229186543 13.059256
#> 54  11.058683347  4.295285
#> 55  17.879666536 23.743616
#> 56   7.783035981 16.501373
#> 57   8.459839628 16.649993
#> 58   9.743338833 18.862881
#> 59  -0.058657313  8.781659
#> 60   0.001366583 10.150031
#> 61   8.262481360 15.457308
#> 62   3.013660702 15.718722
#> 63  13.818210116  9.239281
#> 64  21.783148846  8.918705
#> 65  16.006502210 11.030574
#> 66  11.484014242 22.735284
#> 67  13.896434924 16.439864
#> 68  -1.997289890 10.930957
#> 69   9.924536370 31.320804
#> 70  11.502026935 14.557477
#> 71   1.748731138 21.593312
#> 72   3.692685416 14.814392
#> 73  11.492944577 15.844503
#> 74  14.732761805 24.049896
#> 75  12.382811436  9.417268
#> 76  12.510294363 12.837609
#> 77  13.498749986 19.696957
#> 78  16.946808311 20.116535
#> 79   5.339037311 11.917995
#> 80   7.237811365 13.526597
#> 81   3.800628128 18.066560
#> 82   8.481558744 16.659054
#> 83  18.094343796 20.844745
#> 84  12.466553862 16.718795
#> 85  10.468252474 25.026423
#> 86  15.218057805 19.790155
#> 87   4.250962061 19.894834
#> 88  12.165451008 12.835083
#> 89  12.599593790 17.128601
#> 90   5.422638700 24.432647
#> 91  14.210803248 19.626388
#> 92   9.676001729 12.339478
#> 93   9.810941808 12.840664
#> 94  10.097077025 14.937497
#> 95   5.156349989  6.658048
#> 96  15.139955187 17.039449
#> 97  17.099705733 24.761180
#> 98   2.517018364 16.168941
#> 99  13.031220792 15.241051
#> 100 11.533647767 10.149358
#> 
#>