Collect and aggregate results

Collect and aggregate performance, contribution and importance estimations of a set of raw results produced by run_misty().

collect_results(folders)

Arguments

folders
Paths to folders containing the raw results from run_misty().

Value

List of collected performance, contributions and importances per sample, performance and contribution statistics and aggregated importances.

improvements

Long format tibble with measurements of performance for each target and each sample. Available performance measures are RMSE and variance explained (R2) for a model containing only an intrinsic view (intra.RMSE, intra.R2), model with all views (multi.RMSE, multi.R2), gain of RMSE and gain of variance explained of multi-view model over the intrisic model where gain.RMSE is the relative decrease of RMSE in percent, while gain.R2 is the absolute increase of variance explained in percent. Each value represents the mean performance across folds (k-fold cross-validation). The p values of a one sided t-test of improvement of performance (p.RMSE, p.R2) are also available as a measure.

improvements.stats

Long format tibble with summary statistics (mean, standard deviation and coefficient of variation) for all performance measures for each target over all samples.

contributions

Long format tibble with the values of the coefficients for each view in the meta-model, for each target and each sample. The p values for the coefficient for each view, under the null hypothesis of zero contribution to the meta model are also available.

contributions.stats

Long format tibble with summary statistics for all views per target over all samples. Including mean coffecient value, fraction of contribution, mean and standard deviation of p values.

importances

List of view-specific predictor-target importance tables per sample. The importances in each table are standardized per target and weighted by the quantile of the coefficient for the target in that view. Columns other than Predictor represent target markers.

importances.aggregated

A list of aggregated view-specific predictor-target importance tables . Aggregation is reducing by mean over all samples.

See also

Examples

# Train and collect results for 3 samples in synthetic

library(dplyr)
library(purrr)

data("synthetic")

misty.results <- synthetic[seq_len(3)] %>%
  imap_chr(~ create_initial_view(.x %>% select(-c(row, col, type))) %>%
    add_paraview(.x %>% select(row, col), l = 10) %>%
    run_misty(paste0("results/", .y))) %>%
  collect_results()
#> Generating paraview
#> Training models
#> Generating paraview
#> Training models
#> Generating paraview
#> Training models
#> 
#> Collecting improvements
#> 
#> Collecting contributions
#> 
#> Collecting importances
#> 
#> Aggregating
str(misty.results)
#> List of 6
#>  $ improvements          : tibble [264 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ target : chr [1:264] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ sample : chr [1:264] "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" ...
#>   ..$ measure: chr [1:264] "intra.RMSE" "intra.R2" "multi.RMSE" "multi.R2" ...
#>   ..$ value  : num [1:264] 0.1056 0.9252 0.0985 0.9349 0.0274 ...
#>  $ improvements.stats    : tibble [66 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ target : chr [1:66] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ measure: chr [1:66] "gain.R2" "gain.RMSE" "intra.R2" "intra.RMSE" ...
#>   ..$ mean   : num [1:66] 0.86 6.188 0.928 0.101 0.937 ...
#>   ..$ sd     : num [1:66] 0.11674 0.8589 0.00339 0.00577 0.0033 ...
#>   ..$ cv     : num [1:66] 0.13572 0.1388 0.00365 0.05744 0.00353 ...
#>  $ contributions         : tibble [198 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ target: chr [1:198] "ECM" "ECM" "ECM" "ECM" ...
#>   ..$ sample: chr [1:198] "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" "/tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1" ...
#>   ..$ view  : chr [1:198] "intercept" "intra" "para.10" "p.intercept" ...
#>   ..$ value : num [1:198] -7.41e-02 9.92e-01 2.59e-01 2.80e-106 0.00 ...
#>  $ contributions.stats   : tibble [22 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ target  : chr [1:22] "ECM" "ECM" "ligA" "ligA" ...
#>   ..$ view    : chr [1:22] "intra" "para.10" "intra" "para.10" ...
#>   ..$ mean    : num [1:22] 0.989 0.245 0.996 0.066 0.985 ...
#>   ..$ fraction: num [1:22] 0.8012 0.1988 0.9379 0.0621 0.7747 ...
#>   ..$ p.mean  : num [1:22] 0.00 1.27e-101 0.00 1.26e-44 0.00 ...
#>   ..$ p.sd    : num [1:22] 0.00 2.20e-101 0.00 2.12e-44 0.00 ...
#>  $ importances           :List of 3
#>   ..$ /tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic1 :List of 2
#>   .. ..$ intra  : tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.505 -0.524 -0.268 -0.669 ...
#>   .. .. ..$ ligA     : num [1:11] -0.621 NA -0.61 -0.603 2.058 ...
#>   .. .. ..$ ligB     : num [1:11] 0.305 -0.522 NA -0.495 -0.53 ...
#>   .. .. ..$ ligC     : num [1:11] -0.375 -0.42 -0.428 NA -0.471 ...
#>   .. .. ..$ ligD     : num [1:11] -0.598 2.427 -0.62 -0.61 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.7088 1.2277 -0.7943 -0.7729 0.0769 ...
#>   .. .. ..$ prodB    : num [1:11] 0.266 -0.655 2.609 -0.396 -0.681 ...
#>   .. .. ..$ prodC    : num [1:11] -0.613 -0.703 -0.477 2.403 -0.769 ...
#>   .. .. ..$ prodD    : num [1:11] -0.801 0.738 -0.819 -0.81 0.677 ...
#>   .. .. ..$ protE    : num [1:11] 2.551 -0.68 -0.874 0.584 -0.698 ...
#>   .. .. ..$ protF    : num [1:11] -0.832 1.695 -0.833 -0.752 0.16 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.719 0.461 0.364 -1.447 ...
#>   .. .. ..$ ligA     : num [1:11] -0.3742 NA -1.29 -0.0105 1.8888 ...
#>   .. .. ..$ ligB     : num [1:11] 1.665 -0.695 NA -0.169 -0.758 ...
#>   .. .. ..$ ligC     : num [1:11] 0.212 -0.217 -0.471 NA -0.566 ...
#>   .. .. ..$ ligD     : num [1:11] 0.228 1.724 -1.276 -0.434 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.4293 0.0797 -0.7394 -0.8152 0.5933 ...
#>   .. .. ..$ prodB    : num [1:11] 1.815 -0.837 1.196 -0.111 -0.972 ...
#>   .. .. ..$ prodC    : num [1:11] 0.13 -0.372 -1.04 0.952 -0.691 ...
#>   .. .. ..$ prodD    : num [1:11] -0.5223 0.6688 -0.7129 -0.627 -0.0721 ...
#>   .. .. ..$ protE    : num [1:11] 0.05136 -0.04569 -0.11412 -0.00404 0.01572 ...
#>   .. .. ..$ protF    : num [1:11] -0.507 0.658 -1.003 -0.684 0.643 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ /tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic10:List of 2
#>   .. ..$ intra  : tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.538 -0.583 -0.194 -0.317 ...
#>   .. .. ..$ ligA     : num [1:11] -0.625 NA -0.663 -0.668 2.3 ...
#>   .. .. ..$ ligB     : num [1:11] 0.444 -0.472 NA -0.457 -0.546 ...
#>   .. .. ..$ ligC     : num [1:11] -0.389 -0.51 -0.438 NA -0.487 ...
#>   .. .. ..$ ligD     : num [1:11] -0.64 2.489 -0.649 -0.62 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.737 0.468 -0.811 -0.763 0.694 ...
#>   .. .. ..$ prodB    : num [1:11] 0.413 -0.638 2.629 -0.429 -0.639 ...
#>   .. .. ..$ prodC    : num [1:11] -0.568 -0.714 -0.561 2.476 -0.692 ...
#>   .. .. ..$ prodD    : num [1:11] -0.7037 0.0901 -0.7903 -0.7737 0.8501 ...
#>   .. .. ..$ protE    : num [1:11] 2.4 -0.977 -1.009 0.645 -0.552 ...
#>   .. .. ..$ protF    : num [1:11] -0.82 0.926 -0.886 -0.806 0.819 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -1.151 0.727 -0.174 -0.706 ...
#>   .. .. ..$ ligA     : num [1:11] -0.59 NA -0.548 -0.438 1.864 ...
#>   .. .. ..$ ligB     : num [1:11] 1.0865 -0.8025 NA -0.0635 -0.698 ...
#>   .. .. ..$ ligC     : num [1:11] 0.222 -0.411 -0.428 NA -0.625 ...
#>   .. .. ..$ ligD     : num [1:11] -0.247 2.095 -0.487 -0.672 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.62 0.659 -0.651 -0.646 0.347 ...
#>   .. .. ..$ prodB    : num [1:11] 0.8609 -0.0851 1.6933 -0.2533 -0.6974 ...
#>   .. .. ..$ prodC    : num [1:11] -0.684 -0.171 -1.195 0.821 -0.662 ...
#>   .. .. ..$ prodD    : num [1:11] -0.393 0.9524 -0.6933 -0.5718 -0.0744 ...
#>   .. .. ..$ protE    : num [1:11] 0.0812 -0.1258 -1.054 -0.2519 -0.5068 ...
#>   .. .. ..$ protF    : num [1:11] -0.691 1.288 -0.795 -0.881 0.68 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ /tmp/Rtmp8eaknE/file534c7c8db405/reference/results/synthetic2 :List of 2
#>   .. ..$ intra  : tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -0.614 -0.528 -0.237 -0.426 ...
#>   .. .. ..$ ligA     : num [1:11] -0.603 NA -0.607 -0.621 2.186 ...
#>   .. .. ..$ ligB     : num [1:11] 0.287 -0.457 NA -0.449 -0.487 ...
#>   .. .. ..$ ligC     : num [1:11] -0.4 -0.534 -0.471 NA -0.493 ...
#>   .. .. ..$ ligD     : num [1:11] -0.608 2.338 -0.639 -0.633 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.714 0.367 -0.75 -0.726 0.193 ...
#>   .. .. ..$ prodB    : num [1:11] 0.0969 -0.5583 2.7212 -0.415 -0.5612 ...
#>   .. .. ..$ prodC    : num [1:11] -0.562 -0.662 -0.381 2.609 -0.654 ...
#>   .. .. ..$ prodD    : num [1:11] -0.709 0.028 -0.765 -0.731 0.717 ...
#>   .. .. ..$ protE    : num [1:11] 2.449 -0.814 -0.998 0.694 -0.755 ...
#>   .. .. ..$ protF    : num [1:11] -0.735 0.763 -0.759 -0.666 0.279 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   .. ..$ para.10: tibble [11 × 12] (S3: tbl_df/tbl/data.frame)
#>   .. .. ..$ ECM      : num [1:11] NA -1.06 0.354 0.492 -1.361 ...
#>   .. .. ..$ ligA     : num [1:11] -0.276 NA -0.389 -1.05 1.962 ...
#>   .. .. ..$ ligB     : num [1:11] 0.464 -0.223 NA -0.271 -0.747 ...
#>   .. .. ..$ ligC     : num [1:11] 0.419 -0.646 -0.269 NA -0.73 ...
#>   .. .. ..$ ligD     : num [1:11] 0.285 1.997 -0.555 -0.83 NA ...
#>   .. .. ..$ prodA    : num [1:11] -0.323 0.268 -0.62 -0.673 0.949 ...
#>   .. .. ..$ prodB    : num [1:11] -0.166 0.11 1.503 -0.925 -0.471 ...
#>   .. .. ..$ prodC    : num [1:11] 0.118 -1.027 -0.923 0.935 -0.791 ...
#>   .. .. ..$ prodD    : num [1:11] -0.3725 1.0838 -0.7076 -0.644 0.0112 ...
#>   .. .. ..$ protE    : num [1:11] -0.0519 -0.2362 -0.906 0.014 0.099 ...
#>   .. .. ..$ protF    : num [1:11] -0.497 1.175 -0.739 -1.013 0.728 ...
#>   .. .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>  $ importances.aggregated:List of 2
#>   ..$ intra  :'data.frame':	11 obs. of  12 variables:
#>   .. ..$ ECM      : num [1:11] NA -0.552 -0.545 -0.233 -0.471 ...
#>   .. ..$ ligA     : num [1:11] -0.616 NA -0.626 -0.63 2.182 ...
#>   .. ..$ ligB     : num [1:11] 0.345 -0.484 NA -0.467 -0.521 ...
#>   .. ..$ ligC     : num [1:11] -0.388 -0.488 -0.446 NA -0.484 ...
#>   .. ..$ ligD     : num [1:11] -0.615 2.418 -0.636 -0.621 NA ...
#>   .. ..$ prodA    : num [1:11] -0.72 0.688 -0.785 -0.754 0.321 ...
#>   .. ..$ prodB    : num [1:11] 0.259 -0.617 2.653 -0.413 -0.627 ...
#>   .. ..$ prodC    : num [1:11] -0.581 -0.693 -0.473 2.496 -0.705 ...
#>   .. ..$ prodD    : num [1:11] -0.738 0.285 -0.791 -0.771 0.748 ...
#>   .. ..$ protE    : num [1:11] 2.467 -0.824 -0.96 0.641 -0.668 ...
#>   .. ..$ protF    : num [1:11] -0.796 1.128 -0.826 -0.741 0.419 ...
#>   .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...
#>   ..$ para.10:'data.frame':	11 obs. of  12 variables:
#>   .. ..$ ECM      : num [1:11] NA -0.977 0.514 0.227 -1.171 ...
#>   .. ..$ ligA     : num [1:11] -0.413 NA -0.742 -0.499 1.905 ...
#>   .. ..$ ligB     : num [1:11] 1.072 -0.573 NA -0.168 -0.734 ...
#>   .. ..$ ligC     : num [1:11] 0.285 -0.424 -0.389 NA -0.64 ...
#>   .. ..$ ligD     : num [1:11] 0.0884 1.9388 -0.7727 -0.6452 NA ...
#>   .. ..$ prodA    : num [1:11] -0.457 0.336 -0.67 -0.711 0.63 ...
#>   .. ..$ prodB    : num [1:11] 0.837 -0.271 1.464 -0.43 -0.714 ...
#>   .. ..$ prodC    : num [1:11] -0.145 -0.523 -1.053 0.903 -0.715 ...
#>   .. ..$ prodD    : num [1:11] -0.4293 0.9016 -0.7046 -0.6143 -0.0451 ...
#>   .. ..$ protE    : num [1:11] 0.0269 -0.1359 -0.6914 -0.0806 -0.1307 ...
#>   .. ..$ protF    : num [1:11] -0.565 1.04 -0.846 -0.859 0.684 ...
#>   .. ..$ Predictor: chr [1:11] "ECM" "ligA" "ligB" "ligC" ...