Extract signatures from the results

Signature is a representation of each sample in the space of mistyR results.

Usage

extract_signature(
  misty.results,
  type = c("performance", "contribution", "importance"),
  trim = -Inf,
  trim.measure = c("gain.R2", "multi.R2", "intra.R2", "gain.RMSE", "multi.RMSE",
    "intra.RMSE")
)

Arguments

misty.results

a results list generated by collect_results().

type

type of signature to extract from the results.

trim

display targets with performance value above (if R2 or gain) or below (otherwise) this value only.

trim.measure

the measure used for trimming.

Value

A table with one row per sample from misty.results representing its signature.

Details

The performance signature of each sample is a concatenation of the estimated values of variance explained using only the intraview, the variance explained by the multiview model and the gain in variance explained for each marker. The performance signature vector for each sample available in misty.results is of length \(\textrm{markers} \cdot 3\).

The contribution signature of each sample is a concatenation of the estimated fraction of contribution of each view for each marker. The contribution signature vector for each sample available in misty.results is of length \(\textrm{markers} \cdot \textrm{views}\).

The importance signature of each sample is a concatenation of the estimated and weighted importances for each predictor-target marker pair from all views. The importance signature vector for each sample available in misty.results is of length \(\textrm{markers}^2 \cdot \textrm{views}\).

See also

Examples

library(dplyr)

misty.results <-
  list.files("results", full.names = TRUE) %>% collect_results()
#> 
#> Collecting improvements
#> 
#> Collecting contributions
#> 
#> Collecting importances
#> 
#> Aggregating

extract_signature(misty.results, "performance")
#> # A tibble: 3 × 34
#>   sample       ECM_intra.R2 ECM_multi.R2 ECM_gain.R2 ligA_intra.R2 ligA_multi.R2
#>   <chr>               <dbl>        <dbl>       <dbl>         <dbl>         <dbl>
#> 1 /tmp/RtmpOk…         92.5         93.4       0.958          98.6          98.7
#> 2 /tmp/RtmpOk…         93.2         94.1       0.882          98.7          98.7
#> 3 /tmp/RtmpOk…         92.7         93.5       0.743          98.6          98.6
#> # ℹ 28 more variables: ligA_gain.R2 <dbl>, ligB_intra.R2 <dbl>,
#> #   ligB_multi.R2 <dbl>, ligB_gain.R2 <dbl>, ligC_intra.R2 <dbl>,
#> #   ligC_multi.R2 <dbl>, ligC_gain.R2 <dbl>, ligD_intra.R2 <dbl>,
#> #   ligD_multi.R2 <dbl>, ligD_gain.R2 <dbl>, protE_intra.R2 <dbl>,
#> #   protE_multi.R2 <dbl>, protE_gain.R2 <dbl>, protF_intra.R2 <dbl>,
#> #   protF_multi.R2 <dbl>, protF_gain.R2 <dbl>, prodA_intra.R2 <dbl>,
#> #   prodA_multi.R2 <dbl>, prodA_gain.R2 <dbl>, prodB_intra.R2 <dbl>, …