Select a subset of markers in a view

Select a subset of markers in a view

select_markers(current.views, view = "intraview", ...)

Arguments

current.views
the current view composition.

view
the name of the view to select markers for.

...
one or more select expressions dplyr::select() for the specified view.

Value

A mistyR view composition with selected markers in view.

See also

Examples

# Create a view composition with an intraview and select

library(dplyr)

# get the expression data
data("synthetic")
expr <- synthetic[[1]] %>% select(-c(row, col, type))

# compose
misty.views <- create_initial_view(expr)

# select markers from the intraview not starting with lig and preview
misty.views %>%
  select_markers("intraview", !starts_with("lig")) %>%
  str()

#> List of 2
#>  $ intraview     :List of 2
#>   ..$ abbrev: chr "intra"
#>   ..$ data  : tibble [4,205 × 7] (S3: tbl_df/tbl/data.frame)
#>   .. ..$ ECM  : num [1:4205] 0.0385 0.0327 0.1444 0.387 0.1635 ...
#>   .. ..$ protE: num [1:4205] 1.18 0 1.67 0 0 ...
#>   .. ..$ protF: num [1:4205] 2.547 0.386 1.614 0.913 0.162 ...
#>   .. ..$ prodA: num [1:4205] 0.382 0 0.472 0 0 ...
#>   .. ..$ prodB: num [1:4205] 0 0 0 0 0.16 ...
#>   .. ..$ prodC: num [1:4205] 0 0.536 0 0.418 0 ...
#>   .. ..$ prodD: num [1:4205] 0.588 0 0.379 0 0 ...
#>  $ misty.uniqueid: chr "b98ad35f4e671871cba35f2155228612"